Add tables.go file
