--- crypto/internal/falcon1024/fft_tables.go
@@ -1,517 +1,0 @@
-package falcon1024
-
-// falconGMTabBits contains Falcon reference fpr_gm_tab entries as float64 bit patterns.
-var falconGMTabBits = [2 * n]uint64{
-	0x0000000000000000, 0x0000000000000000, 0x8000000000000000, 0x3ff0000000000000,
-	0x3fe6a09e667f3bcd, 0x3fe6a09e667f3bcd, 0xbfe6a09e667f3bcd, 0x3fe6a09e667f3bcd,
-	0x3fed906bcf328d46, 0x3fd87de2a6aea963, 0xbfd87de2a6aea963, 0x3fed906bcf328d46,
-	0x3fd87de2a6aea963, 0x3fed906bcf328d46, 0xbfed906bcf328d46, 0x3fd87de2a6aea963,
-	0x3fef6297cff75cb0, 0x3fc8f8b83c69a60b, 0xbfc8f8b83c69a60b, 0x3fef6297cff75cb0,
-	0x3fe1c73b39ae68c8, 0x3fea9b66290ea1a3, 0xbfea9b66290ea1a3, 0x3fe1c73b39ae68c8,
-	0x3fea9b66290ea1a3, 0x3fe1c73b39ae68c8, 0xbfe1c73b39ae68c8, 0x3fea9b66290ea1a3,
-	0x3fc8f8b83c69a60b, 0x3fef6297cff75cb0, 0xbfef6297cff75cb0, 0x3fc8f8b83c69a60b,
-	0x3fefd88da3d12526, 0x3fb917a6bc29b42c, 0xbfb917a6bc29b42c, 0x3fefd88da3d12526,
-	0x3fe44cf325091dd6, 0x3fe8bc806b151741, 0xbfe8bc806b151741, 0x3fe44cf325091dd6,
-	0x3fec38b2f180bdb1, 0x3fde2b5d3806f63b, 0xbfde2b5d3806f63b, 0x3fec38b2f180bdb1,
-	0x3fd294062ed59f06, 0x3fee9f4156c62dda, 0xbfee9f4156c62dda, 0x3fd294062ed59f06,
-	0x3fee9f4156c62dda, 0x3fd294062ed59f06, 0xbfd294062ed59f06, 0x3fee9f4156c62dda,
-	0x3fde2b5d3806f63b, 0x3fec38b2f180bdb1, 0xbfec38b2f180bdb1, 0x3fde2b5d3806f63b,
-	0x3fe8bc806b151741, 0x3fe44cf325091dd6, 0xbfe44cf325091dd6, 0x3fe8bc806b151741,
-	0x3fb917a6bc29b42c, 0x3fefd88da3d12526, 0xbfefd88da3d12526, 0x3fb917a6bc29b42c,
-	0x3feff621e3796d7e, 0x3fa91f65f10dd814, 0xbfa91f65f10dd814, 0x3feff621e3796d7e,
-	0x3fe57d69348ceca0, 0x3fe7b5df226aafaf, 0xbfe7b5df226aafaf, 0x3fe57d69348ceca0,
-	0x3feced7af43cc773, 0x3fdb5d1009e15cc0, 0xbfdb5d1009e15cc0, 0x3feced7af43cc773,
-	0x3fd58f9a75ab1fdd, 0x3fee212104f686e5, 0xbfee212104f686e5, 0x3fd58f9a75ab1fdd,
-	0x3fef0a7efb9230d7, 0x3fcf19f97b215f1b, 0xbfcf19f97b215f1b, 0x3fef0a7efb9230d7,
-	0x3fe073879922ffee, 0x3feb728345196e3e, 0xbfeb728345196e3e, 0x3fe073879922ffee,
-	0x3fe9b3e047f38741, 0x3fe30ff7fce17035, 0xbfe30ff7fce17035, 0x3fe9b3e047f38741,
-	0x3fc2c8106e8e613a, 0x3fefa7557f08a517, 0xbfefa7557f08a517, 0x3fc2c8106e8e613a,
-	0x3fefa7557f08a517, 0x3fc2c8106e8e613a, 0xbfc2c8106e8e613a, 0x3fefa7557f08a517,
-	0x3fe30ff7fce17035, 0x3fe9b3e047f38741, 0xbfe9b3e047f38741, 0x3fe30ff7fce17035,
-	0x3feb728345196e3e, 0x3fe073879922ffee, 0xbfe073879922ffee, 0x3feb728345196e3e,
-	0x3fcf19f97b215f1b, 0x3fef0a7efb9230d7, 0xbfef0a7efb9230d7, 0x3fcf19f97b215f1b,
-	0x3fee212104f686e5, 0x3fd58f9a75ab1fdd, 0xbfd58f9a75ab1fdd, 0x3fee212104f686e5,
-	0x3fdb5d1009e15cc0, 0x3feced7af43cc773, 0xbfeced7af43cc773, 0x3fdb5d1009e15cc0,
-	0x3fe7b5df226aafaf, 0x3fe57d69348ceca0, 0xbfe57d69348ceca0, 0x3fe7b5df226aafaf,
-	0x3fa91f65f10dd814, 0x3feff621e3796d7e, 0xbfeff621e3796d7e, 0x3fa91f65f10dd814,
-	0x3feffd886084cd0d, 0x3f992155f7a3667e, 0xbf992155f7a3667e, 0x3feffd886084cd0d,
-	0x3fe610b7551d2cdf, 0x3fe72d0837efff96, 0xbfe72d0837efff96, 0x3fe610b7551d2cdf,
-	0x3fed4134d14dc93a, 0x3fd9ef7943a8ed8a, 0xbfd9ef7943a8ed8a, 0x3fed4134d14dc93a,
-	0x3fd7088530fa459f, 0x3feddb13b6ccc23c, 0xbfeddb13b6ccc23c, 0x3fd7088530fa459f,
-	0x3fef38f3ac64e589, 0x3fcc0b826a7e4f63, 0xbfcc0b826a7e4f63, 0x3fef38f3ac64e589,
-	0x3fe11eb3541b4b23, 0x3feb090a58150200, 0xbfeb090a58150200, 0x3fe11eb3541b4b23,
-	0x3fea29a7a0462782, 0x3fe26d054cdd12df, 0xbfe26d054cdd12df, 0x3fea29a7a0462782,
-	0x3fc5e214448b3fc6, 0x3fef8764fa714ba9, 0xbfef8764fa714ba9, 0x3fc5e214448b3fc6,
-	0x3fefc26470e19fd3, 0x3fbf564e56a9730e, 0xbfbf564e56a9730e, 0x3fefc26470e19fd3,
-	0x3fe3affa292050b9, 0x3fe93a22499263fb, 0xbfe93a22499263fb, 0x3fe3affa292050b9,
-	0x3febd7c0ac6f952a, 0x3fdf8ba4dbf89aba, 0xbfdf8ba4dbf89aba, 0x3febd7c0ac6f952a,
-	0x3fd111d262b1f677, 0x3feed740e7684963, 0xbfeed740e7684963, 0x3fd111d262b1f677,
-	0x3fee6288ec48e112, 0x3fd4135c94176601, 0xbfd4135c94176601, 0x3fee6288ec48e112,
-	0x3fdcc66e9931c45e, 0x3fec954b213411f5, 0xbfec954b213411f5, 0x3fdcc66e9931c45e,
-	0x3fe83b0e0bff976e, 0x3fe4e6cabbe3e5e9, 0xbfe4e6cabbe3e5e9, 0x3fe83b0e0bff976e,
-	0x3fb2d52092ce19f6, 0x3fefe9cdad01883a, 0xbfefe9cdad01883a, 0x3fb2d52092ce19f6,
-	0x3fefe9cdad01883a, 0x3fb2d52092ce19f6, 0xbfb2d52092ce19f6, 0x3fefe9cdad01883a,
-	0x3fe4e6cabbe3e5e9, 0x3fe83b0e0bff976e, 0xbfe83b0e0bff976e, 0x3fe4e6cabbe3e5e9,
-	0x3fec954b213411f5, 0x3fdcc66e9931c45e, 0xbfdcc66e9931c45e, 0x3fec954b213411f5,
-	0x3fd4135c94176601, 0x3fee6288ec48e112, 0xbfee6288ec48e112, 0x3fd4135c94176601,
-	0x3feed740e7684963, 0x3fd111d262b1f677, 0xbfd111d262b1f677, 0x3feed740e7684963,
-	0x3fdf8ba4dbf89aba, 0x3febd7c0ac6f952a, 0xbfebd7c0ac6f952a, 0x3fdf8ba4dbf89aba,
-	0x3fe93a22499263fb, 0x3fe3affa292050b9, 0xbfe3affa292050b9, 0x3fe93a22499263fb,
-	0x3fbf564e56a9730e, 0x3fefc26470e19fd3, 0xbfefc26470e19fd3, 0x3fbf564e56a9730e,
-	0x3fef8764fa714ba9, 0x3fc5e214448b3fc6, 0xbfc5e214448b3fc6, 0x3fef8764fa714ba9,
-	0x3fe26d054cdd12df, 0x3fea29a7a0462782, 0xbfea29a7a0462782, 0x3fe26d054cdd12df,
-	0x3feb090a58150200, 0x3fe11eb3541b4b23, 0xbfe11eb3541b4b23, 0x3feb090a58150200,
-	0x3fcc0b826a7e4f63, 0x3fef38f3ac64e589, 0xbfef38f3ac64e589, 0x3fcc0b826a7e4f63,
-	0x3feddb13b6ccc23c, 0x3fd7088530fa459f, 0xbfd7088530fa459f, 0x3feddb13b6ccc23c,
-	0x3fd9ef7943a8ed8a, 0x3fed4134d14dc93a, 0xbfed4134d14dc93a, 0x3fd9ef7943a8ed8a,
-	0x3fe72d0837efff96, 0x3fe610b7551d2cdf, 0xbfe610b7551d2cdf, 0x3fe72d0837efff96,
-	0x3f992155f7a3667e, 0x3feffd886084cd0d, 0xbfeffd886084cd0d, 0x3f992155f7a3667e,
-	0x3fefff62169b92db, 0x3f8921d1fcdec784, 0xbf8921d1fcdec784, 0x3fefff62169b92db,
-	0x3fe6591925f0783d, 0x3fe6e74454eaa8af, 0xbfe6e74454eaa8af, 0x3fe6591925f0783d,
-	0x3fed696173c9e68b, 0x3fd9372a63bc93d7, 0xbfd9372a63bc93d7, 0x3fed696173c9e68b,
-	0x3fd7c3a9311dcce7, 0x3fedb6526238a09b, 0xbfedb6526238a09b, 0x3fd7c3a9311dcce7,
-	0x3fef4e603b0b2f2d, 0x3fca82a025b00451, 0xbfca82a025b00451, 0x3fef4e603b0b2f2d,
-	0x3fe1734d63dedb49, 0x3fead2bc9e21d511, 0xbfead2bc9e21d511, 0x3fe1734d63dedb49,
-	0x3fea63091b02fae2, 0x3fe21a799933eb59, 0xbfe21a799933eb59, 0x3fea63091b02fae2,
-	0x3fc76dd9de50bf31, 0x3fef7599a3a12077, 0xbfef7599a3a12077, 0x3fc76dd9de50bf31,
-	0x3fefce15fd6da67b, 0x3fbc3785c79ec2d5, 0xbfbc3785c79ec2d5, 0x3fefce15fd6da67b,
-	0x3fe3fed9534556d4, 0x3fe8fbcca3ef940d, 0xbfe8fbcca3ef940d, 0x3fe3fed9534556d4,
-	0x3fec08c426725549, 0x3fdedc1952ef78d6, 0xbfdedc1952ef78d6, 0x3fec08c426725549,
-	0x3fd1d3443f4cdb3e, 0x3feebbd8c8df0b74, 0xbfeebbd8c8df0b74, 0x3fd1d3443f4cdb3e,
-	0x3fee817bab4cd10d, 0x3fd35410c2e18152, 0xbfd35410c2e18152, 0x3fee817bab4cd10d,
-	0x3fdd79775b86e389, 0x3fec678b3488739b, 0xbfec678b3488739b, 0x3fdd79775b86e389,
-	0x3fe87c400fba2ebf, 0x3fe49a449b9b0939, 0xbfe49a449b9b0939, 0x3fe87c400fba2ebf,
-	0x3fb5f6d00a9aa419, 0x3fefe1cafcbd5b09, 0xbfefe1cafcbd5b09, 0x3fb5f6d00a9aa419,
-	0x3feff095658e71ad, 0x3faf656e79f820e0, 0xbfaf656e79f820e0, 0x3feff095658e71ad,
-	0x3fe5328292a35596, 0x3fe7f8ece3571771, 0xbfe7f8ece3571771, 0x3fe5328292a35596,
-	0x3fecc1f0f3fcfc5c, 0x3fdc1249d8011ee7, 0xbfdc1249d8011ee7, 0x3fecc1f0f3fcfc5c,
-	0x3fd4d1e24278e76a, 0x3fee426a4b2bc17e, 0xbfee426a4b2bc17e, 0x3fd4d1e24278e76a,
-	0x3feef178a3e473c2, 0x3fd04fb80e37fdae, 0xbfd04fb80e37fdae, 0x3feef178a3e473c2,
-	0x3fe01cfc874c3eb7, 0x3feba5aa673590d2, 0xbfeba5aa673590d2, 0x3fe01cfc874c3eb7,
-	0x3fe9777ef4c7d742, 0x3fe36058b10659f3, 0xbfe36058b10659f3, 0x3fe9777ef4c7d742,
-	0x3fc139f0cedaf577, 0x3fefb5797195d741, 0xbfefb5797195d741, 0x3fc139f0cedaf577,
-	0x3fef97f924c9099b, 0x3fc45576b1293e5a, 0xbfc45576b1293e5a, 0x3fef97f924c9099b,
-	0x3fe2bedb25faf3ea, 0x3fe9ef43ef29af94, 0xbfe9ef43ef29af94, 0x3fe2bedb25faf3ea,
-	0x3feb3e4d3ef55712, 0x3fe0c9704d5d898f, 0xbfe0c9704d5d898f, 0x3feb3e4d3ef55712,
-	0x3fcd934fe5454311, 0x3fef2252f7763ada, 0xbfef2252f7763ada, 0x3fcd934fe5454311,
-	0x3fedfeae622dbe2b, 0x3fd64c7ddd3f27c6, 0xbfd64c7ddd3f27c6, 0x3fedfeae622dbe2b,
-	0x3fdaa6c82b6d3fca, 0x3fed17e7743e35dc, 0xbfed17e7743e35dc, 0x3fdaa6c82b6d3fca,
-	0x3fe771e75f037261, 0x3fe5c77bbe65018c, 0xbfe5c77bbe65018c, 0x3fe771e75f037261,
-	0x3fa2d865759455cd, 0x3feffa72effef75d, 0xbfeffa72effef75d, 0x3fa2d865759455cd,
-	0x3feffa72effef75d, 0x3fa2d865759455cd, 0xbfa2d865759455cd, 0x3feffa72effef75d,
-	0x3fe5c77bbe65018c, 0x3fe771e75f037261, 0xbfe771e75f037261, 0x3fe5c77bbe65018c,
-	0x3fed17e7743e35dc, 0x3fdaa6c82b6d3fca, 0xbfdaa6c82b6d3fca, 0x3fed17e7743e35dc,
-	0x3fd64c7ddd3f27c6, 0x3fedfeae622dbe2b, 0xbfedfeae622dbe2b, 0x3fd64c7ddd3f27c6,
-	0x3fef2252f7763ada, 0x3fcd934fe5454311, 0xbfcd934fe5454311, 0x3fef2252f7763ada,
-	0x3fe0c9704d5d898f, 0x3feb3e4d3ef55712, 0xbfeb3e4d3ef55712, 0x3fe0c9704d5d898f,
-	0x3fe9ef43ef29af94, 0x3fe2bedb25faf3ea, 0xbfe2bedb25faf3ea, 0x3fe9ef43ef29af94,
-	0x3fc45576b1293e5a, 0x3fef97f924c9099b, 0xbfef97f924c9099b, 0x3fc45576b1293e5a,
-	0x3fefb5797195d741, 0x3fc139f0cedaf577, 0xbfc139f0cedaf577, 0x3fefb5797195d741,
-	0x3fe36058b10659f3, 0x3fe9777ef4c7d742, 0xbfe9777ef4c7d742, 0x3fe36058b10659f3,
-	0x3feba5aa673590d2, 0x3fe01cfc874c3eb7, 0xbfe01cfc874c3eb7, 0x3feba5aa673590d2,
-	0x3fd04fb80e37fdae, 0x3feef178a3e473c2, 0xbfeef178a3e473c2, 0x3fd04fb80e37fdae,
-	0x3fee426a4b2bc17e, 0x3fd4d1e24278e76a, 0xbfd4d1e24278e76a, 0x3fee426a4b2bc17e,
-	0x3fdc1249d8011ee7, 0x3fecc1f0f3fcfc5c, 0xbfecc1f0f3fcfc5c, 0x3fdc1249d8011ee7,
-	0x3fe7f8ece3571771, 0x3fe5328292a35596, 0xbfe5328292a35596, 0x3fe7f8ece3571771,
-	0x3faf656e79f820e0, 0x3feff095658e71ad, 0xbfeff095658e71ad, 0x3faf656e79f820e0,
-	0x3fefe1cafcbd5b09, 0x3fb5f6d00a9aa419, 0xbfb5f6d00a9aa419, 0x3fefe1cafcbd5b09,
-	0x3fe49a449b9b0939, 0x3fe87c400fba2ebf, 0xbfe87c400fba2ebf, 0x3fe49a449b9b0939,
-	0x3fec678b3488739b, 0x3fdd79775b86e389, 0xbfdd79775b86e389, 0x3fec678b3488739b,
-	0x3fd35410c2e18152, 0x3fee817bab4cd10d, 0xbfee817bab4cd10d, 0x3fd35410c2e18152,
-	0x3feebbd8c8df0b74, 0x3fd1d3443f4cdb3e, 0xbfd1d3443f4cdb3e, 0x3feebbd8c8df0b74,
-	0x3fdedc1952ef78d6, 0x3fec08c426725549, 0xbfec08c426725549, 0x3fdedc1952ef78d6,
-	0x3fe8fbcca3ef940d, 0x3fe3fed9534556d4, 0xbfe3fed9534556d4, 0x3fe8fbcca3ef940d,
-	0x3fbc3785c79ec2d5, 0x3fefce15fd6da67b, 0xbfefce15fd6da67b, 0x3fbc3785c79ec2d5,
-	0x3fef7599a3a12077, 0x3fc76dd9de50bf31, 0xbfc76dd9de50bf31, 0x3fef7599a3a12077,
-	0x3fe21a799933eb59, 0x3fea63091b02fae2, 0xbfea63091b02fae2, 0x3fe21a799933eb59,
-	0x3fead2bc9e21d511, 0x3fe1734d63dedb49, 0xbfe1734d63dedb49, 0x3fead2bc9e21d511,
-	0x3fca82a025b00451, 0x3fef4e603b0b2f2d, 0xbfef4e603b0b2f2d, 0x3fca82a025b00451,
-	0x3fedb6526238a09b, 0x3fd7c3a9311dcce7, 0xbfd7c3a9311dcce7, 0x3fedb6526238a09b,
-	0x3fd9372a63bc93d7, 0x3fed696173c9e68b, 0xbfed696173c9e68b, 0x3fd9372a63bc93d7,
-	0x3fe6e74454eaa8af, 0x3fe6591925f0783d, 0xbfe6591925f0783d, 0x3fe6e74454eaa8af,
-	0x3f8921d1fcdec784, 0x3fefff62169b92db, 0xbfefff62169b92db, 0x3f8921d1fcdec784,
-	0x3fefffd8858e8a92, 0x3f7921f0fe670071, 0xbf7921f0fe670071, 0x3fefffd8858e8a92,
-	0x3fe67cf78491af10, 0x3fe6c40d73c18275, 0xbfe6c40d73c18275, 0x3fe67cf78491af10,
-	0x3fed7d0b02b8ecf9, 0x3fd8daa52ec8a4b0, 0xbfd8daa52ec8a4b0, 0x3fed7d0b02b8ecf9,
-	0x3fd820e3b04eaac4, 0x3feda383a9668988, 0xbfeda383a9668988, 0x3fd820e3b04eaac4,
-	0x3fef58a2b1789e84, 0x3fc9bdcbf2dc4366, 0xbfc9bdcbf2dc4366, 0x3fef58a2b1789e84,
-	0x3fe19d5a09f2b9b8, 0x3feab7325916c0d4, 0xbfeab7325916c0d4, 0x3fe19d5a09f2b9b8,
-	0x3fea7f58529fe69d, 0x3fe1f0f08bbc861b, 0xbfe1f0f08bbc861b, 0x3fea7f58529fe69d,
-	0x3fc83366e89c64c6, 0x3fef6c3f7df5bbb7, 0xbfef6c3f7df5bbb7, 0x3fc83366e89c64c6,
-	0x3fefd37914220b84, 0x3fbaa7b724495c03, 0xbfbaa7b724495c03, 0x3fefd37914220b84,
-	0x3fe425ff178e6bb1, 0x3fe8dc45331698cc, 0xbfe8dc45331698cc, 0x3fe425ff178e6bb1,
-	0x3fec20de3fa971b0, 0x3fde83e0eaf85114, 0xbfde83e0eaf85114, 0x3fec20de3fa971b0,
-	0x3fd233bbabc3bb71, 0x3feeadb2e8e7a88e, 0xbfeeadb2e8e7a88e, 0x3fd233bbabc3bb71,
-	0x3fee9084361df7f2, 0x3fd2f422daec0387, 0xbfd2f422daec0387, 0x3fee9084361df7f2,
-	0x3fddd28f1481cc58, 0x3fec5042012b6907, 0xbfec5042012b6907, 0x3fddd28f1481cc58,
-	0x3fe89c7e9a4dd4aa, 0x3fe473b51b987347, 0xbfe473b51b987347, 0x3fe89c7e9a4dd4aa,
-	0x3fb787586a5d5b21, 0x3fefdd539ff1f456, 0xbfefdd539ff1f456, 0x3fb787586a5d5b21,
-	0x3feff3830f8d575c, 0x3fac428d12c0d7e3, 0xbfac428d12c0d7e3, 0x3feff3830f8d575c,
-	0x3fe5581038975137, 0x3fe7d7836cc33db2, 0xbfe7d7836cc33db2, 0x3fe5581038975137,
-	0x3fecd7d9898b32f6, 0x3fdbb7cf2304bd01, 0xbfdbb7cf2304bd01, 0x3fecd7d9898b32f6,
-	0x3fd530d880af3c24, 0x3fee31eae870ce25, 0xbfee31eae870ce25, 0x3fd530d880af3c24,
-	0x3feefe220c0b95ec, 0x3fcfdcdc1adfedf9, 0xbfcfdcdc1adfedf9, 0x3feefe220c0b95ec,
-	0x3fe0485626ae221a, 0x3feb8c38d27504e9, 0xbfeb8c38d27504e9, 0x3fe0485626ae221a,
-	0x3fe995cf2ed80d22, 0x3fe338400d0c8e57, 0xbfe338400d0c8e57, 0x3fe995cf2ed80d22,
-	0x3fc20116d4ec7bcf, 0x3fefae8e8e46cfbb, 0xbfefae8e8e46cfbb, 0x3fc20116d4ec7bcf,
-	0x3fef9fce55adb2c8, 0x3fc38edbb0cd8d14, 0xbfc38edbb0cd8d14, 0x3fef9fce55adb2c8,
-	0x3fe2e780e3e8ea17, 0x3fe9d1b1f5ea80d5, 0xbfe9d1b1f5ea80d5, 0x3fe2e780e3e8ea17,
-	0x3feb5889fe921405, 0x3fe09e907417c5e1, 0xbfe09e907417c5e1, 0x3feb5889fe921405,
-	0x3fce56ca1e101a1b, 0x3fef168f53f7205d, 0xbfef168f53f7205d, 0x3fce56ca1e101a1b,
-	0x3fee100cca2980ac, 0x3fd5ee27379ea693, 0xbfd5ee27379ea693, 0x3fee100cca2980ac,
-	0x3fdb020d6c7f4009, 0x3fed02d4feb2bd92, 0xbfed02d4feb2bd92, 0x3fdb020d6c7f4009,
-	0x3fe79400574f55e5, 0x3fe5a28d2a5d7250, 0xbfe5a28d2a5d7250, 0x3fe79400574f55e5,
-	0x3fa5fc00d290cd43, 0x3feff871dadb81df, 0xbfeff871dadb81df, 0x3fa5fc00d290cd43,
-	0x3feffc251df1d3f8, 0x3f9f693731d1cf01, 0xbf9f693731d1cf01, 0x3feffc251df1d3f8,
-	0x3fe5ec3495837074, 0x3fe74f948da8d28d, 0xbfe74f948da8d28d, 0x3fe5ec3495837074,
-	0x3fed2cb220e0ef9f, 0x3fda4b4127dea1e5, 0xbfda4b4127dea1e5, 0x3fed2cb220e0ef9f,
-	0x3fd6aa9d7dc77e17, 0x3feded05f7de47da, 0xbfeded05f7de47da, 0x3fd6aa9d7dc77e17,
-	0x3fef2dc9c9089a9d, 0x3fcccf8cb312b286, 0xbfcccf8cb312b286, 0x3fef2dc9c9089a9d,
-	0x3fe0f426bb2a8e7e, 0x3feb23cd470013b4, 0xbfeb23cd470013b4, 0x3fe0f426bb2a8e7e,
-	0x3fea0c95eabaf937, 0x3fe2960727629ca8, 0xbfe2960727629ca8, 0x3fea0c95eabaf937,
-	0x3fc51bdf8597c5f2, 0x3fef8fd5ffae41db, 0xbfef8fd5ffae41db, 0x3fc51bdf8597c5f2,
-	0x3fefbc1617e44186, 0x3fc072a047ba831d, 0xbfc072a047ba831d, 0x3fefbc1617e44186,
-	0x3fe3884185dfeb22, 0x3fe958efe48e6dd7, 0xbfe958efe48e6dd7, 0x3fe3884185dfeb22,
-	0x3febbed7c49380ea, 0x3fdfe2f64be71210, 0xbfdfe2f64be71210, 0x3febbed7c49380ea,
-	0x3fd0b0d9cfdbdb90, 0x3feee482e25a9dbc, 0xbfeee482e25a9dbc, 0x3fd0b0d9cfdbdb90,
-	0x3fee529f04729ffc, 0x3fd472b8a5571054, 0xbfd472b8a5571054, 0x3fee529f04729ffc,
-	0x3fdc6c7f4997000b, 0x3fecabc169a0b900, 0xbfecabc169a0b900, 0x3fdc6c7f4997000b,
-	0x3fe81a1b33b57acc, 0x3fe50cc09f59a09b, 0xbfe50cc09f59a09b, 0x3fe81a1b33b57acc,
-	0x3fb1440134d709b3, 0x3fefed58ecb673c4, 0xbfefed58ecb673c4, 0x3fb1440134d709b3,
-	0x3fefe5f3af2e3940, 0x3fb4661179272096, 0xbfb4661179272096, 0x3fefe5f3af2e3940,
-	0x3fe4c0a145ec0004, 0x3fe85bc51ae958cc, 0xbfe85bc51ae958cc, 0x3fe4c0a145ec0004,
-	0x3fec7e8e52233cf3, 0x3fdd2016e8e9db5b, 0xbfdd2016e8e9db5b, 0x3fec7e8e52233cf3,
-	0x3fd3b3cefa0414b7, 0x3fee7227db6a9744, 0xbfee7227db6a9744, 0x3fd3b3cefa0414b7,
-	0x3feec9b2d3c3bf84, 0x3fd172a0d7765177, 0xbfd172a0d7765177, 0x3feec9b2d3c3bf84,
-	0x3fdf3405963fd067, 0x3febf064e15377dd, 0xbfebf064e15377dd, 0x3fdf3405963fd067,
-	0x3fe91b166fd49da2, 0x3fe3d78238c58344, 0xbfe3d78238c58344, 0x3fe91b166fd49da2,
-	0x3fbdc70ecbae9fc9, 0x3fefc8646cfeb721, 0xbfefc8646cfeb721, 0x3fbdc70ecbae9fc9,
-	0x3fef7ea629e63d6e, 0x3fc6a81304f64ab2, 0xbfc6a81304f64ab2, 0x3fef7ea629e63d6e,
-	0x3fe243d5fb98ac1f, 0x3fea4678c8119ac8, 0xbfea4678c8119ac8, 0x3fe243d5fb98ac1f,
-	0x3feaee04b43c1474, 0x3fe14915af336ceb, 0xbfe14915af336ceb, 0x3feaee04b43c1474,
-	0x3fcb4732ef3d6722, 0x3fef43d085ff92dd, 0xbfef43d085ff92dd, 0x3fcb4732ef3d6722,
-	0x3fedc8d7cb410260, 0x3fd766340f2418f6, 0xbfd766340f2418f6, 0x3fedc8d7cb410260,
-	0x3fd993716141bdff, 0x3fed556f52e93eb1, 0xbfed556f52e93eb1, 0x3fd993716141bdff,
-	0x3fe70a42b3176d7a, 0x3fe63503a31c1be9, 0xbfe63503a31c1be9, 0x3fe70a42b3176d7a,
-	0x3f92d936bbe30efd, 0x3feffe9cb44b51a1, 0xbfeffe9cb44b51a1, 0x3f92d936bbe30efd,
-	0x3feffe9cb44b51a1, 0x3f92d936bbe30efd, 0xbf92d936bbe30efd, 0x3feffe9cb44b51a1,
-	0x3fe63503a31c1be9, 0x3fe70a42b3176d7a, 0xbfe70a42b3176d7a, 0x3fe63503a31c1be9,
-	0x3fed556f52e93eb1, 0x3fd993716141bdff, 0xbfd993716141bdff, 0x3fed556f52e93eb1,
-	0x3fd766340f2418f6, 0x3fedc8d7cb410260, 0xbfedc8d7cb410260, 0x3fd766340f2418f6,
-	0x3fef43d085ff92dd, 0x3fcb4732ef3d6722, 0xbfcb4732ef3d6722, 0x3fef43d085ff92dd,
-	0x3fe14915af336ceb, 0x3feaee04b43c1474, 0xbfeaee04b43c1474, 0x3fe14915af336ceb,
-	0x3fea4678c8119ac8, 0x3fe243d5fb98ac1f, 0xbfe243d5fb98ac1f, 0x3fea4678c8119ac8,
-	0x3fc6a81304f64ab2, 0x3fef7ea629e63d6e, 0xbfef7ea629e63d6e, 0x3fc6a81304f64ab2,
-	0x3fefc8646cfeb721, 0x3fbdc70ecbae9fc9, 0xbfbdc70ecbae9fc9, 0x3fefc8646cfeb721,
-	0x3fe3d78238c58344, 0x3fe91b166fd49da2, 0xbfe91b166fd49da2, 0x3fe3d78238c58344,
-	0x3febf064e15377dd, 0x3fdf3405963fd067, 0xbfdf3405963fd067, 0x3febf064e15377dd,
-	0x3fd172a0d7765177, 0x3feec9b2d3c3bf84, 0xbfeec9b2d3c3bf84, 0x3fd172a0d7765177,
-	0x3fee7227db6a9744, 0x3fd3b3cefa0414b7, 0xbfd3b3cefa0414b7, 0x3fee7227db6a9744,
-	0x3fdd2016e8e9db5b, 0x3fec7e8e52233cf3, 0xbfec7e8e52233cf3, 0x3fdd2016e8e9db5b,
-	0x3fe85bc51ae958cc, 0x3fe4c0a145ec0004, 0xbfe4c0a145ec0004, 0x3fe85bc51ae958cc,
-	0x3fb4661179272096, 0x3fefe5f3af2e3940, 0xbfefe5f3af2e3940, 0x3fb4661179272096,
-	0x3fefed58ecb673c4, 0x3fb1440134d709b3, 0xbfb1440134d709b3, 0x3fefed58ecb673c4,
-	0x3fe50cc09f59a09b, 0x3fe81a1b33b57acc, 0xbfe81a1b33b57acc, 0x3fe50cc09f59a09b,
-	0x3fecabc169a0b900, 0x3fdc6c7f4997000b, 0xbfdc6c7f4997000b, 0x3fecabc169a0b900,
-	0x3fd472b8a5571054, 0x3fee529f04729ffc, 0xbfee529f04729ffc, 0x3fd472b8a5571054,
-	0x3feee482e25a9dbc, 0x3fd0b0d9cfdbdb90, 0xbfd0b0d9cfdbdb90, 0x3feee482e25a9dbc,
-	0x3fdfe2f64be71210, 0x3febbed7c49380ea, 0xbfebbed7c49380ea, 0x3fdfe2f64be71210,
-	0x3fe958efe48e6dd7, 0x3fe3884185dfeb22, 0xbfe3884185dfeb22, 0x3fe958efe48e6dd7,
-	0x3fc072a047ba831d, 0x3fefbc1617e44186, 0xbfefbc1617e44186, 0x3fc072a047ba831d,
-	0x3fef8fd5ffae41db, 0x3fc51bdf8597c5f2, 0xbfc51bdf8597c5f2, 0x3fef8fd5ffae41db,
-	0x3fe2960727629ca8, 0x3fea0c95eabaf937, 0xbfea0c95eabaf937, 0x3fe2960727629ca8,
-	0x3feb23cd470013b4, 0x3fe0f426bb2a8e7e, 0xbfe0f426bb2a8e7e, 0x3feb23cd470013b4,
-	0x3fcccf8cb312b286, 0x3fef2dc9c9089a9d, 0xbfef2dc9c9089a9d, 0x3fcccf8cb312b286,
-	0x3feded05f7de47da, 0x3fd6aa9d7dc77e17, 0xbfd6aa9d7dc77e17, 0x3feded05f7de47da,
-	0x3fda4b4127dea1e5, 0x3fed2cb220e0ef9f, 0xbfed2cb220e0ef9f, 0x3fda4b4127dea1e5,
-	0x3fe74f948da8d28d, 0x3fe5ec3495837074, 0xbfe5ec3495837074, 0x3fe74f948da8d28d,
-	0x3f9f693731d1cf01, 0x3feffc251df1d3f8, 0xbfeffc251df1d3f8, 0x3f9f693731d1cf01,
-	0x3feff871dadb81df, 0x3fa5fc00d290cd43, 0xbfa5fc00d290cd43, 0x3feff871dadb81df,
-	0x3fe5a28d2a5d7250, 0x3fe79400574f55e5, 0xbfe79400574f55e5, 0x3fe5a28d2a5d7250,
-	0x3fed02d4feb2bd92, 0x3fdb020d6c7f4009, 0xbfdb020d6c7f4009, 0x3fed02d4feb2bd92,
-	0x3fd5ee27379ea693, 0x3fee100cca2980ac, 0xbfee100cca2980ac, 0x3fd5ee27379ea693,
-	0x3fef168f53f7205d, 0x3fce56ca1e101a1b, 0xbfce56ca1e101a1b, 0x3fef168f53f7205d,
-	0x3fe09e907417c5e1, 0x3feb5889fe921405, 0xbfeb5889fe921405, 0x3fe09e907417c5e1,
-	0x3fe9d1b1f5ea80d5, 0x3fe2e780e3e8ea17, 0xbfe2e780e3e8ea17, 0x3fe9d1b1f5ea80d5,
-	0x3fc38edbb0cd8d14, 0x3fef9fce55adb2c8, 0xbfef9fce55adb2c8, 0x3fc38edbb0cd8d14,
-	0x3fefae8e8e46cfbb, 0x3fc20116d4ec7bcf, 0xbfc20116d4ec7bcf, 0x3fefae8e8e46cfbb,
-	0x3fe338400d0c8e57, 0x3fe995cf2ed80d22, 0xbfe995cf2ed80d22, 0x3fe338400d0c8e57,
-	0x3feb8c38d27504e9, 0x3fe0485626ae221a, 0xbfe0485626ae221a, 0x3feb8c38d27504e9,
-	0x3fcfdcdc1adfedf9, 0x3feefe220c0b95ec, 0xbfeefe220c0b95ec, 0x3fcfdcdc1adfedf9,
-	0x3fee31eae870ce25, 0x3fd530d880af3c24, 0xbfd530d880af3c24, 0x3fee31eae870ce25,
-	0x3fdbb7cf2304bd01, 0x3fecd7d9898b32f6, 0xbfecd7d9898b32f6, 0x3fdbb7cf2304bd01,
-	0x3fe7d7836cc33db2, 0x3fe5581038975137, 0xbfe5581038975137, 0x3fe7d7836cc33db2,
-	0x3fac428d12c0d7e3, 0x3feff3830f8d575c, 0xbfeff3830f8d575c, 0x3fac428d12c0d7e3,
-	0x3fefdd539ff1f456, 0x3fb787586a5d5b21, 0xbfb787586a5d5b21, 0x3fefdd539ff1f456,
-	0x3fe473b51b987347, 0x3fe89c7e9a4dd4aa, 0xbfe89c7e9a4dd4aa, 0x3fe473b51b987347,
-	0x3fec5042012b6907, 0x3fddd28f1481cc58, 0xbfddd28f1481cc58, 0x3fec5042012b6907,
-	0x3fd2f422daec0387, 0x3fee9084361df7f2, 0xbfee9084361df7f2, 0x3fd2f422daec0387,
-	0x3feeadb2e8e7a88e, 0x3fd233bbabc3bb71, 0xbfd233bbabc3bb71, 0x3feeadb2e8e7a88e,
-	0x3fde83e0eaf85114, 0x3fec20de3fa971b0, 0xbfec20de3fa971b0, 0x3fde83e0eaf85114,
-	0x3fe8dc45331698cc, 0x3fe425ff178e6bb1, 0xbfe425ff178e6bb1, 0x3fe8dc45331698cc,
-	0x3fbaa7b724495c03, 0x3fefd37914220b84, 0xbfefd37914220b84, 0x3fbaa7b724495c03,
-	0x3fef6c3f7df5bbb7, 0x3fc83366e89c64c6, 0xbfc83366e89c64c6, 0x3fef6c3f7df5bbb7,
-	0x3fe1f0f08bbc861b, 0x3fea7f58529fe69d, 0xbfea7f58529fe69d, 0x3fe1f0f08bbc861b,
-	0x3feab7325916c0d4, 0x3fe19d5a09f2b9b8, 0xbfe19d5a09f2b9b8, 0x3feab7325916c0d4,
-	0x3fc9bdcbf2dc4366, 0x3fef58a2b1789e84, 0xbfef58a2b1789e84, 0x3fc9bdcbf2dc4366,
-	0x3feda383a9668988, 0x3fd820e3b04eaac4, 0xbfd820e3b04eaac4, 0x3feda383a9668988,
-	0x3fd8daa52ec8a4b0, 0x3fed7d0b02b8ecf9, 0xbfed7d0b02b8ecf9, 0x3fd8daa52ec8a4b0,
-	0x3fe6c40d73c18275, 0x3fe67cf78491af10, 0xbfe67cf78491af10, 0x3fe6c40d73c18275,
-	0x3f7921f0fe670071, 0x3fefffd8858e8a92, 0xbfefffd8858e8a92, 0x3f7921f0fe670071,
-	0x3feffff621621d02, 0x3f6921f8becca4ba, 0xbf6921f8becca4ba, 0x3feffff621621d02,
-	0x3fe68ed1eaa19c71, 0x3fe6b25ced2fe29c, 0xbfe6b25ced2fe29c, 0x3fe68ed1eaa19c71,
-	0x3fed86c48445a44f, 0x3fd8ac4b86d5ed44, 0xbfd8ac4b86d5ed44, 0x3fed86c48445a44f,
-	0x3fd84f6aaaf3903f, 0x3fed9a00dd8b3d46, 0xbfed9a00dd8b3d46, 0x3fd84f6aaaf3903f,
-	0x3fef5da6ed43685d, 0x3fc95b49e9b62afa, 0xbfc95b49e9b62afa, 0x3fef5da6ed43685d,
-	0x3fe1b250171373bf, 0x3feaa9547a2cb98e, 0xbfeaa9547a2cb98e, 0x3fe1b250171373bf,
-	0x3fea8d676e545ad2, 0x3fe1dc1b64dc4872, 0xbfe1dc1b64dc4872, 0x3fea8d676e545ad2,
-	0x3fc8961727c41804, 0x3fef677556883cee, 0xbfef677556883cee, 0x3fc8961727c41804,
-	0x3fefd60d2da75c9e, 0x3fb9dfb6eb24a85c, 0xbfb9dfb6eb24a85c, 0x3fefd60d2da75c9e,
-	0x3fe4397f5b2a4380, 0x3fe8cc6a75184655, 0xbfe8cc6a75184655, 0x3fe4397f5b2a4380,
-	0x3fec2cd14931e3f1, 0x3fde57a86d3cd825, 0xbfde57a86d3cd825, 0x3fec2cd14931e3f1,
-	0x3fd263e6995554ba, 0x3feea68393e65800, 0xbfeea68393e65800, 0x3fd263e6995554ba,
-	0x3fee97ec36016b30, 0x3fd2c41a4e954520, 0xbfd2c41a4e954520, 0x3fee97ec36016b30,
-	0x3fddfeff66a941de, 0x3fec44833141c004, 0xbfec44833141c004, 0x3fddfeff66a941de,
-	0x3fe8ac871ede1d88, 0x3fe4605a692b32a2, 0xbfe4605a692b32a2, 0x3fe8ac871ede1d88,
-	0x3fb84f8712c130a1, 0x3fefdafa7514538c, 0xbfefdafa7514538c, 0x3fb84f8712c130a1,
-	0x3feff4dc54b1bed3, 0x3faab101bd5f8317, 0xbfaab101bd5f8317, 0x3feff4dc54b1bed3,
-	0x3fe56ac35197649f, 0x3fe7c6b89ce2d333, 0xbfe7c6b89ce2d333, 0x3fe56ac35197649f,
-	0x3fece2b32799a060, 0x3fdb8a7814fd5693, 0xbfdb8a7814fd5693, 0x3fece2b32799a060,
-	0x3fd5604012f467b4, 0x3fee298f4439197a, 0xbfee298f4439197a, 0x3fd5604012f467b4,
-	0x3fef045a14cf738c, 0x3fcf7b7480bd3802, 0xbfcf7b7480bd3802, 0x3fef045a14cf738c,
-	0x3fe05df3ec31b8b7, 0x3feb7f6686e792e9, 0xbfeb7f6686e792e9, 0x3fe05df3ec31b8b7,
-	0x3fe9a4dfa42b06b2, 0x3fe32421ec49a61f, 0xbfe32421ec49a61f, 0x3fe9a4dfa42b06b2,
-	0x3fc264994dfd3409, 0x3fefaafbcb0cfddc, 0xbfefaafbcb0cfddc, 0x3fc264994dfd3409,
-	0x3fefa39bac7a1791, 0x3fc32b7bf94516a7, 0xbfc32b7bf94516a7, 0x3fefa39bac7a1791,
-	0x3fe2fbc24b441015, 0x3fe9c2d110f075c2, 0xbfe9c2d110f075c2, 0x3fe2fbc24b441015,
-	0x3feb658f14fdbc47, 0x3fe089112032b08c, 0xbfe089112032b08c, 0x3feb658f14fdbc47,
-	0x3fceb86b462de348, 0x3fef1090bc898f5f, 0xbfef1090bc898f5f, 0x3fceb86b462de348,
-	0x3fee18a02fdc66d9, 0x3fd5bee78b9db3b6, 0xbfd5bee78b9db3b6, 0x3fee18a02fdc66d9,
-	0x3fdb2f971db31972, 0x3fecf830e8ce467b, 0xbfecf830e8ce467b, 0x3fdb2f971db31972,
-	0x3fe7a4f707bf97d2, 0x3fe59001d5f723df, 0xbfe59001d5f723df, 0x3fe7a4f707bf97d2,
-	0x3fa78dbaa5874686, 0x3feff753bb1b9164, 0xbfeff753bb1b9164, 0x3fa78dbaa5874686,
-	0x3feffce09ce2a679, 0x3f9c454f4ce53b1d, 0xbf9c454f4ce53b1d, 0x3feffce09ce2a679,
-	0x3fe5fe7cbde56a10, 0x3fe73e558e079942, 0xbfe73e558e079942, 0x3fe5fe7cbde56a10,
-	0x3fed36fc7bcbfbdc, 0x3fda1d6543b50ac0, 0xbfda1d6543b50ac0, 0x3fed36fc7bcbfbdc,
-	0x3fd6d998638a0cb6, 0x3fede4160f6d8d81, 0xbfede4160f6d8d81, 0x3fd6d998638a0cb6,
-	0x3fef33685a3aaef0, 0x3fcc6d90535d74dd, 0xbfcc6d90535d74dd, 0x3fef33685a3aaef0,
-	0x3fe1097248d0a957, 0x3feb16742a4ca2f5, 0xbfeb16742a4ca2f5, 0x3fe1097248d0a957,
-	0x3fea1b26d2c0a75e, 0x3fe2818bef4d3cba, 0xbfe2818bef4d3cba, 0x3fea1b26d2c0a75e,
-	0x3fc57f008654cbde, 0x3fef8ba737cb4b78, 0xbfef8ba737cb4b78, 0x3fc57f008654cbde,
-	0x3fefbf470f0a8d88, 0x3fc00ee8ad6fb85b, 0xbfc00ee8ad6fb85b, 0x3fefbf470f0a8d88,
-	0x3fe39c23e3d63029, 0x3fe94990e3ac4a6c, 0xbfe94990e3ac4a6c, 0x3fe39c23e3d63029,
-	0x3febcb54cb0d2327, 0x3fdfb7575c24d2de, 0xbfdfb7575c24d2de, 0x3febcb54cb0d2327,
-	0x3fd0e15b4e1749ce, 0x3feeddeb6a078651, 0xbfeeddeb6a078651, 0x3fd0e15b4e1749ce,
-	0x3fee5a9d550467d3, 0x3fd44310dc8936f0, 0xbfd44310dc8936f0, 0x3fee5a9d550467d3,
-	0x3fdc997fc3865389, 0x3feca08f19b9c449, 0xbfeca08f19b9c449, 0x3fdc997fc3865389,
-	0x3fe82a9c13f545ff, 0x3fe4f9cc25cca486, 0xbfe4f9cc25cca486, 0x3fe82a9c13f545ff,
-	0x3fb20c9674ed444d, 0x3fefeb9d2530410f, 0xbfefeb9d2530410f, 0x3fb20c9674ed444d,
-	0x3fefe7ea85482d60, 0x3fb39d9f12c5a299, 0xbfb39d9f12c5a299, 0x3fefe7ea85482d60,
-	0x3fe4d3bc6d589f7f, 0x3fe84b7111af83fa, 0xbfe84b7111af83fa, 0x3fe4d3bc6d589f7f,
-	0x3fec89f587029c13, 0x3fdcf34baee1cd21, 0xbfdcf34baee1cd21, 0x3fec89f587029c13,
-	0x3fd3e39be96ec271, 0x3fee6a61c55d53a7, 0xbfee6a61c55d53a7, 0x3fd3e39be96ec271,
-	0x3feed0835e999009, 0x3fd1423eefc69378, 0xbfd1423eefc69378, 0x3feed0835e999009,
-	0x3fdf5fdee656cda3, 0x3febe41b611154c1, 0xbfebe41b611154c1, 0x3fdf5fdee656cda3,
-	0x3fe92aa41fc5a815, 0x3fe3c3c44981c518, 0xbfe3c3c44981c518, 0x3fe92aa41fc5a815,
-	0x3fbe8eb7fde4aa3f, 0x3fefc56e3b7d9af6, 0xbfefc56e3b7d9af6, 0x3fbe8eb7fde4aa3f,
-	0x3fef830f4a40c60c, 0x3fc6451a831d830d, 0xbfc6451a831d830d, 0x3fef830f4a40c60c,
-	0x3fe258734cbb7110, 0x3fea38184a593bc6, 0xbfea38184a593bc6, 0x3fe258734cbb7110,
-	0x3feafb8fd89f57b6, 0x3fe133e9cfee254f, 0xbfe133e9cfee254f, 0x3feafb8fd89f57b6,
-	0x3fcba96334f15dad, 0x3fef3e6bbc1bbc65, 0xbfef3e6bbc1bbc65, 0x3fcba96334f15dad,
-	0x3fedd1fef38a915a, 0x3fd73763c9261092, 0xbfd73763c9261092, 0x3fedd1fef38a915a,
-	0x3fd9c17d440df9f2, 0x3fed4b5b1b187524, 0xbfed4b5b1b187524, 0x3fd9c17d440df9f2,
-	0x3fe71bac960e41bf, 0x3fe622e44fec22ff, 0xbfe622e44fec22ff, 0x3fe71bac960e41bf,
-	0x3f95fd4d21fab226, 0x3feffe1c6870cb77, 0xbfeffe1c6870cb77, 0x3f95fd4d21fab226,
-	0x3fefff0943c53bd1, 0x3f8f6a296ab997cb, 0xbf8f6a296ab997cb, 0x3fefff0943c53bd1,
-	0x3fe64715437f535b, 0x3fe6f8ca99c95b75, 0xbfe6f8ca99c95b75, 0x3fe64715437f535b,
-	0x3fed5f7172888a7f, 0x3fd96555b7ab948f, 0xbfd96555b7ab948f, 0x3fed5f7172888a7f,
-	0x3fd794f5e613dfae, 0x3fedbf9e4395759a, 0xbfedbf9e4395759a, 0x3fd794f5e613dfae,
-	0x3fef492206bcabb4, 0x3fcae4f1d5f3b9ab, 0xbfcae4f1d5f3b9ab, 0x3fef492206bcabb4,
-	0x3fe15e36e4dbe2bc, 0x3feae068f345ecef, 0xbfeae068f345ecef, 0x3fe15e36e4dbe2bc,
-	0x3fea54c91090f523, 0x3fe22f2d662c13e2, 0xbfe22f2d662c13e2, 0x3fea54c91090f523,
-	0x3fc70afd8d08c4ff, 0x3fef7a299c1a322a, 0xbfef7a299c1a322a, 0x3fc70afd8d08c4ff,
-	0x3fefcb4703914354, 0x3fbcff533b307dc1, 0xbfbcff533b307dc1, 0x3fefcb4703914354,
-	0x3fe3eb33eabe0680, 0x3fe90b7943575efe, 0xbfe90b7943575efe, 0x3fe3eb33eabe0680,
-	0x3febfc9d25a1b147, 0x3fdf081906bff7fe, 0xbfdf081906bff7fe, 0x3febfc9d25a1b147,
-	0x3fd1a2f7fbe8f243, 0x3feec2cf4b1af6b2, 0xbfeec2cf4b1af6b2, 0x3fd1a2f7fbe8f243,
-	0x3fee79db29a5165a, 0x3fd383f5e353b6ab, 0xbfd383f5e353b6ab, 0x3fee79db29a5165a,
-	0x3fdd4cd02ba8609d, 0x3fec7315899eaad7, 0xbfec7315899eaad7, 0x3fdd4cd02ba8609d,
-	0x3fe86c0a1d9aa195, 0x3fe4ad79516722f1, 0xbfe4ad79516722f1, 0x3fe86c0a1d9aa195,
-	0x3fb52e774a4d4d0a, 0x3fefe3e92be9d886, 0xbfefe3e92be9d886, 0x3fb52e774a4d4d0a,
-	0x3fefef0102826191, 0x3fb07b614e463064, 0xbfb07b614e463064, 0x3fefef0102826191,
-	0x3fe51fa81cd99aa6, 0x3fe8098b756e52fa, 0xbfe8098b756e52fa, 0x3fe51fa81cd99aa6,
-	0x3fecb6e20a00da99, 0x3fdc3f6d47263129, 0xbfdc3f6d47263129, 0x3fecb6e20a00da99,
-	0x3fd4a253d11b82f3, 0x3fee4a8dff81ce5e, 0xbfee4a8dff81ce5e, 0x3fd4a253d11b82f3,
-	0x3feeeb074c50a544, 0x3fd0804e05eb661e, 0xbfd0804e05eb661e, 0x3feeeb074c50a544,
-	0x3fe00740c82b82e1, 0x3febb249a0b6c40d, 0xbfebb249a0b6c40d, 0x3fe00740c82b82e1,
-	0x3fe9683f42bd7fe1, 0x3fe374531b817f8d, 0xbfe374531b817f8d, 0x3fe9683f42bd7fe1,
-	0x3fc0d64dbcb26786, 0x3fefb8d18d66adb7, 0xbfefb8d18d66adb7, 0x3fc0d64dbcb26786,
-	0x3fef93f14f85ac08, 0x3fc4b8b17f79fa88, 0xbfc4b8b17f79fa88, 0x3fef93f14f85ac08,
-	0x3fe2aa76e87aeb58, 0x3fe9fdf4f13149de, 0xbfe9fdf4f13149de, 0x3fe2aa76e87aeb58,
-	0x3feb3115a5f37bf3, 0x3fe0ded0b84bc4b6, 0xbfe0ded0b84bc4b6, 0x3feb3115a5f37bf3,
-	0x3fcd31774d2cbdee, 0x3fef2817fc4609ce, 0xbfef2817fc4609ce, 0x3fcd31774d2cbdee,
-	0x3fedf5e36a9ba59c, 0x3fd67b949cad63cb, 0xbfd67b949cad63cb, 0x3fedf5e36a9ba59c,
-	0x3fda790cd3dbf31b, 0x3fed2255c6e5a4e1, 0xbfed2255c6e5a4e1, 0x3fda790cd3dbf31b,
-	0x3fe760c52c304764, 0x3fe5d9dee73e345c, 0xbfe5d9dee73e345c, 0x3fe760c52c304764,
-	0x3fa14685db42c17f, 0x3feffb55e425fdae, 0xbfeffb55e425fdae, 0x3fa14685db42c17f,
-	0x3feff97c4208c014, 0x3fa46a396ff86179, 0xbfa46a396ff86179, 0x3feff97c4208c014,
-	0x3fe5b50b264f7448, 0x3fe782fb1b90b35b, 0xbfe782fb1b90b35b, 0x3fe5b50b264f7448,
-	0x3fed0d672f59d2b9, 0x3fdad473125cdc09, 0xbfdad473125cdc09, 0x3fed0d672f59d2b9,
-	0x3fd61d595c88c202, 0x3fee0766d9280f54, 0xbfee0766d9280f54, 0x3fd61d595c88c202,
-	0x3fef1c7abe284708, 0x3fcdf5163f01099a, 0xbfcdf5163f01099a, 0x3fef1c7abe284708,
-	0x3fe0b405878f85ec, 0x3feb4b7409de7925, 0xbfeb4b7409de7925, 0x3fe0b405878f85ec,
-	0x3fe9e082edb42472, 0x3fe2d333d34e9bb8, 0xbfe2d333d34e9bb8, 0x3fe9e082edb42472,
-	0x3fc3f22f57db4893, 0x3fef9bed7cfbde29, 0xbfef9bed7cfbde29, 0x3fc3f22f57db4893,
-	0x3fefb20dc681d54d, 0x3fc19d8940be24e7, 0xbfc19d8940be24e7, 0x3fefb20dc681d54d,
-	0x3fe34c5252c14de1, 0x3fe986aef1457594, 0xbfe986aef1457594, 0x3fe34c5252c14de1,
-	0x3feb98fa1fd9155e, 0x3fe032ae55edbd96, 0xbfe032ae55edbd96, 0x3feb98fa1fd9155e,
-	0x3fd01f1806b9fdd2, 0x3feef7d6e51ca3c0, 0xbfeef7d6e51ca3c0, 0x3fd01f1806b9fdd2,
-	0x3fee3a33ec75ce85, 0x3fd50163dc197048, 0xbfd50163dc197048, 0x3fee3a33ec75ce85,
-	0x3fdbe51517ffc0d9, 0x3fecccee20c2dea0, 0xbfecccee20c2dea0, 0x3fdbe51517ffc0d9,
-	0x3fe7e83f87b03686, 0x3fe5454ff5159dfc, 0xbfe5454ff5159dfc, 0x3fe7e83f87b03686,
-	0x3fadd406f9808ec9, 0x3feff21614e131ed, 0xbfeff21614e131ed, 0x3fadd406f9808ec9,
-	0x3fefdf9922f73307, 0x3fb6bf1b3e79b129, 0xbfb6bf1b3e79b129, 0x3fefdf9922f73307,
-	0x3fe48703306091ff, 0x3fe88c66e7481ba1, 0xbfe88c66e7481ba1, 0x3fe48703306091ff,
-	0x3fec5bef59fef85a, 0x3fdda60c5cfa10d9, 0xbfdda60c5cfa10d9, 0x3fec5bef59fef85a,
-	0x3fd3241fb638baaf, 0x3fee89095bad6025, 0xbfee89095bad6025, 0x3fd3241fb638baaf,
-	0x3feeb4cf515b8811, 0x3fd2038583d727be, 0xbfd2038583d727be, 0x3feeb4cf515b8811,
-	0x3fdeb00695f25620, 0x3fec14d9dc465e57, 0xbfec14d9dc465e57, 0x3fdeb00695f25620,
-	0x3fe8ec109b486c49, 0x3fe41272663d108c, 0xbfe41272663d108c, 0x3fe8ec109b486c49,
-	0x3fbb6fa6ec38f64c, 0x3fefd0d158d86087, 0xbfefd0d158d86087, 0x3fbb6fa6ec38f64c,
-	0x3fef70f6434b7eb7, 0x3fc7d0a7bbd2cb1c, 0xbfc7d0a7bbd2cb1c, 0x3fef70f6434b7eb7,
-	0x3fe205baa17560d6, 0x3fea7138de9d60f5, 0xbfea7138de9d60f5, 0x3fe205baa17560d6,
-	0x3feac4ffbd3efac8, 0x3fe188591f3a46e5, 0xbfe188591f3a46e5, 0x3feac4ffbd3efac8,
-	0x3fca203e1b1831da, 0x3fef538b1faf2d07, 0xbfef538b1faf2d07, 0x3fca203e1b1831da,
-	0x3fedacf42ce68ab9, 0x3fd7f24dd37341e4, 0xbfd7f24dd37341e4, 0x3fedacf42ce68ab9,
-	0x3fd908ef81ef7bd1, 0x3fed733f508c0dff, 0xbfed733f508c0dff, 0x3fd908ef81ef7bd1,
-	0x3fe6d5afef4aafcd, 0x3fe66b0f3f52b386, 0xbfe66b0f3f52b386, 0x3fe6d5afef4aafcd,
-	0x3f82d96b0e509703, 0x3fefffa72c978c4f, 0xbfefffa72c978c4f, 0x3f82d96b0e509703,
-	0x3fefffa72c978c4f, 0x3f82d96b0e509703, 0xbf82d96b0e509703, 0x3fefffa72c978c4f,
-	0x3fe66b0f3f52b386, 0x3fe6d5afef4aafcd, 0xbfe6d5afef4aafcd, 0x3fe66b0f3f52b386,
-	0x3fed733f508c0dff, 0x3fd908ef81ef7bd1, 0xbfd908ef81ef7bd1, 0x3fed733f508c0dff,
-	0x3fd7f24dd37341e4, 0x3fedacf42ce68ab9, 0xbfedacf42ce68ab9, 0x3fd7f24dd37341e4,
-	0x3fef538b1faf2d07, 0x3fca203e1b1831da, 0xbfca203e1b1831da, 0x3fef538b1faf2d07,
-	0x3fe188591f3a46e5, 0x3feac4ffbd3efac8, 0xbfeac4ffbd3efac8, 0x3fe188591f3a46e5,
-	0x3fea7138de9d60f5, 0x3fe205baa17560d6, 0xbfe205baa17560d6, 0x3fea7138de9d60f5,
-	0x3fc7d0a7bbd2cb1c, 0x3fef70f6434b7eb7, 0xbfef70f6434b7eb7, 0x3fc7d0a7bbd2cb1c,
-	0x3fefd0d158d86087, 0x3fbb6fa6ec38f64c, 0xbfbb6fa6ec38f64c, 0x3fefd0d158d86087,
-	0x3fe41272663d108c, 0x3fe8ec109b486c49, 0xbfe8ec109b486c49, 0x3fe41272663d108c,
-	0x3fec14d9dc465e57, 0x3fdeb00695f25620, 0xbfdeb00695f25620, 0x3fec14d9dc465e57,
-	0x3fd2038583d727be, 0x3feeb4cf515b8811, 0xbfeeb4cf515b8811, 0x3fd2038583d727be,
-	0x3fee89095bad6025, 0x3fd3241fb638baaf, 0xbfd3241fb638baaf, 0x3fee89095bad6025,
-	0x3fdda60c5cfa10d9, 0x3fec5bef59fef85a, 0xbfec5bef59fef85a, 0x3fdda60c5cfa10d9,
-	0x3fe88c66e7481ba1, 0x3fe48703306091ff, 0xbfe48703306091ff, 0x3fe88c66e7481ba1,
-	0x3fb6bf1b3e79b129, 0x3fefdf9922f73307, 0xbfefdf9922f73307, 0x3fb6bf1b3e79b129,
-	0x3feff21614e131ed, 0x3fadd406f9808ec9, 0xbfadd406f9808ec9, 0x3feff21614e131ed,
-	0x3fe5454ff5159dfc, 0x3fe7e83f87b03686, 0xbfe7e83f87b03686, 0x3fe5454ff5159dfc,
-	0x3fecccee20c2dea0, 0x3fdbe51517ffc0d9, 0xbfdbe51517ffc0d9, 0x3fecccee20c2dea0,
-	0x3fd50163dc197048, 0x3fee3a33ec75ce85, 0xbfee3a33ec75ce85, 0x3fd50163dc197048,
-	0x3feef7d6e51ca3c0, 0x3fd01f1806b9fdd2, 0xbfd01f1806b9fdd2, 0x3feef7d6e51ca3c0,
-	0x3fe032ae55edbd96, 0x3feb98fa1fd9155e, 0xbfeb98fa1fd9155e, 0x3fe032ae55edbd96,
-	0x3fe986aef1457594, 0x3fe34c5252c14de1, 0xbfe34c5252c14de1, 0x3fe986aef1457594,
-	0x3fc19d8940be24e7, 0x3fefb20dc681d54d, 0xbfefb20dc681d54d, 0x3fc19d8940be24e7,
-	0x3fef9bed7cfbde29, 0x3fc3f22f57db4893, 0xbfc3f22f57db4893, 0x3fef9bed7cfbde29,
-	0x3fe2d333d34e9bb8, 0x3fe9e082edb42472, 0xbfe9e082edb42472, 0x3fe2d333d34e9bb8,
-	0x3feb4b7409de7925, 0x3fe0b405878f85ec, 0xbfe0b405878f85ec, 0x3feb4b7409de7925,
-	0x3fcdf5163f01099a, 0x3fef1c7abe284708, 0xbfef1c7abe284708, 0x3fcdf5163f01099a,
-	0x3fee0766d9280f54, 0x3fd61d595c88c202, 0xbfd61d595c88c202, 0x3fee0766d9280f54,
-	0x3fdad473125cdc09, 0x3fed0d672f59d2b9, 0xbfed0d672f59d2b9, 0x3fdad473125cdc09,
-	0x3fe782fb1b90b35b, 0x3fe5b50b264f7448, 0xbfe5b50b264f7448, 0x3fe782fb1b90b35b,
-	0x3fa46a396ff86179, 0x3feff97c4208c014, 0xbfeff97c4208c014, 0x3fa46a396ff86179,
-	0x3feffb55e425fdae, 0x3fa14685db42c17f, 0xbfa14685db42c17f, 0x3feffb55e425fdae,
-	0x3fe5d9dee73e345c, 0x3fe760c52c304764, 0xbfe760c52c304764, 0x3fe5d9dee73e345c,
-	0x3fed2255c6e5a4e1, 0x3fda790cd3dbf31b, 0xbfda790cd3dbf31b, 0x3fed2255c6e5a4e1,
-	0x3fd67b949cad63cb, 0x3fedf5e36a9ba59c, 0xbfedf5e36a9ba59c, 0x3fd67b949cad63cb,
-	0x3fef2817fc4609ce, 0x3fcd31774d2cbdee, 0xbfcd31774d2cbdee, 0x3fef2817fc4609ce,
-	0x3fe0ded0b84bc4b6, 0x3feb3115a5f37bf3, 0xbfeb3115a5f37bf3, 0x3fe0ded0b84bc4b6,
-	0x3fe9fdf4f13149de, 0x3fe2aa76e87aeb58, 0xbfe2aa76e87aeb58, 0x3fe9fdf4f13149de,
-	0x3fc4b8b17f79fa88, 0x3fef93f14f85ac08, 0xbfef93f14f85ac08, 0x3fc4b8b17f79fa88,
-	0x3fefb8d18d66adb7, 0x3fc0d64dbcb26786, 0xbfc0d64dbcb26786, 0x3fefb8d18d66adb7,
-	0x3fe374531b817f8d, 0x3fe9683f42bd7fe1, 0xbfe9683f42bd7fe1, 0x3fe374531b817f8d,
-	0x3febb249a0b6c40d, 0x3fe00740c82b82e1, 0xbfe00740c82b82e1, 0x3febb249a0b6c40d,
-	0x3fd0804e05eb661e, 0x3feeeb074c50a544, 0xbfeeeb074c50a544, 0x3fd0804e05eb661e,
-	0x3fee4a8dff81ce5e, 0x3fd4a253d11b82f3, 0xbfd4a253d11b82f3, 0x3fee4a8dff81ce5e,
-	0x3fdc3f6d47263129, 0x3fecb6e20a00da99, 0xbfecb6e20a00da99, 0x3fdc3f6d47263129,
-	0x3fe8098b756e52fa, 0x3fe51fa81cd99aa6, 0xbfe51fa81cd99aa6, 0x3fe8098b756e52fa,
-	0x3fb07b614e463064, 0x3fefef0102826191, 0xbfefef0102826191, 0x3fb07b614e463064,
-	0x3fefe3e92be9d886, 0x3fb52e774a4d4d0a, 0xbfb52e774a4d4d0a, 0x3fefe3e92be9d886,
-	0x3fe4ad79516722f1, 0x3fe86c0a1d9aa195, 0xbfe86c0a1d9aa195, 0x3fe4ad79516722f1,
-	0x3fec7315899eaad7, 0x3fdd4cd02ba8609d, 0xbfdd4cd02ba8609d, 0x3fec7315899eaad7,
-	0x3fd383f5e353b6ab, 0x3fee79db29a5165a, 0xbfee79db29a5165a, 0x3fd383f5e353b6ab,
-	0x3feec2cf4b1af6b2, 0x3fd1a2f7fbe8f243, 0xbfd1a2f7fbe8f243, 0x3feec2cf4b1af6b2,
-	0x3fdf081906bff7fe, 0x3febfc9d25a1b147, 0xbfebfc9d25a1b147, 0x3fdf081906bff7fe,
-	0x3fe90b7943575efe, 0x3fe3eb33eabe0680, 0xbfe3eb33eabe0680, 0x3fe90b7943575efe,
-	0x3fbcff533b307dc1, 0x3fefcb4703914354, 0xbfefcb4703914354, 0x3fbcff533b307dc1,
-	0x3fef7a299c1a322a, 0x3fc70afd8d08c4ff, 0xbfc70afd8d08c4ff, 0x3fef7a299c1a322a,
-	0x3fe22f2d662c13e2, 0x3fea54c91090f523, 0xbfea54c91090f523, 0x3fe22f2d662c13e2,
-	0x3feae068f345ecef, 0x3fe15e36e4dbe2bc, 0xbfe15e36e4dbe2bc, 0x3feae068f345ecef,
-	0x3fcae4f1d5f3b9ab, 0x3fef492206bcabb4, 0xbfef492206bcabb4, 0x3fcae4f1d5f3b9ab,
-	0x3fedbf9e4395759a, 0x3fd794f5e613dfae, 0xbfd794f5e613dfae, 0x3fedbf9e4395759a,
-	0x3fd96555b7ab948f, 0x3fed5f7172888a7f, 0xbfed5f7172888a7f, 0x3fd96555b7ab948f,
-	0x3fe6f8ca99c95b75, 0x3fe64715437f535b, 0xbfe64715437f535b, 0x3fe6f8ca99c95b75,
-	0x3f8f6a296ab997cb, 0x3fefff0943c53bd1, 0xbfefff0943c53bd1, 0x3f8f6a296ab997cb,
-	0x3feffe1c6870cb77, 0x3f95fd4d21fab226, 0xbf95fd4d21fab226, 0x3feffe1c6870cb77,
-	0x3fe622e44fec22ff, 0x3fe71bac960e41bf, 0xbfe71bac960e41bf, 0x3fe622e44fec22ff,
-	0x3fed4b5b1b187524, 0x3fd9c17d440df9f2, 0xbfd9c17d440df9f2, 0x3fed4b5b1b187524,
-	0x3fd73763c9261092, 0x3fedd1fef38a915a, 0xbfedd1fef38a915a, 0x3fd73763c9261092,
-	0x3fef3e6bbc1bbc65, 0x3fcba96334f15dad, 0xbfcba96334f15dad, 0x3fef3e6bbc1bbc65,
-	0x3fe133e9cfee254f, 0x3feafb8fd89f57b6, 0xbfeafb8fd89f57b6, 0x3fe133e9cfee254f,
-	0x3fea38184a593bc6, 0x3fe258734cbb7110, 0xbfe258734cbb7110, 0x3fea38184a593bc6,
-	0x3fc6451a831d830d, 0x3fef830f4a40c60c, 0xbfef830f4a40c60c, 0x3fc6451a831d830d,
-	0x3fefc56e3b7d9af6, 0x3fbe8eb7fde4aa3f, 0xbfbe8eb7fde4aa3f, 0x3fefc56e3b7d9af6,
-	0x3fe3c3c44981c518, 0x3fe92aa41fc5a815, 0xbfe92aa41fc5a815, 0x3fe3c3c44981c518,
-	0x3febe41b611154c1, 0x3fdf5fdee656cda3, 0xbfdf5fdee656cda3, 0x3febe41b611154c1,
-	0x3fd1423eefc69378, 0x3feed0835e999009, 0xbfeed0835e999009, 0x3fd1423eefc69378,
-	0x3fee6a61c55d53a7, 0x3fd3e39be96ec271, 0xbfd3e39be96ec271, 0x3fee6a61c55d53a7,
-	0x3fdcf34baee1cd21, 0x3fec89f587029c13, 0xbfec89f587029c13, 0x3fdcf34baee1cd21,
-	0x3fe84b7111af83fa, 0x3fe4d3bc6d589f7f, 0xbfe4d3bc6d589f7f, 0x3fe84b7111af83fa,
-	0x3fb39d9f12c5a299, 0x3fefe7ea85482d60, 0xbfefe7ea85482d60, 0x3fb39d9f12c5a299,
-	0x3fefeb9d2530410f, 0x3fb20c9674ed444d, 0xbfb20c9674ed444d, 0x3fefeb9d2530410f,
-	0x3fe4f9cc25cca486, 0x3fe82a9c13f545ff, 0xbfe82a9c13f545ff, 0x3fe4f9cc25cca486,
-	0x3feca08f19b9c449, 0x3fdc997fc3865389, 0xbfdc997fc3865389, 0x3feca08f19b9c449,
-	0x3fd44310dc8936f0, 0x3fee5a9d550467d3, 0xbfee5a9d550467d3, 0x3fd44310dc8936f0,
-	0x3feeddeb6a078651, 0x3fd0e15b4e1749ce, 0xbfd0e15b4e1749ce, 0x3feeddeb6a078651,
-	0x3fdfb7575c24d2de, 0x3febcb54cb0d2327, 0xbfebcb54cb0d2327, 0x3fdfb7575c24d2de,
-	0x3fe94990e3ac4a6c, 0x3fe39c23e3d63029, 0xbfe39c23e3d63029, 0x3fe94990e3ac4a6c,
-	0x3fc00ee8ad6fb85b, 0x3fefbf470f0a8d88, 0xbfefbf470f0a8d88, 0x3fc00ee8ad6fb85b,
-	0x3fef8ba737cb4b78, 0x3fc57f008654cbde, 0xbfc57f008654cbde, 0x3fef8ba737cb4b78,
-	0x3fe2818bef4d3cba, 0x3fea1b26d2c0a75e, 0xbfea1b26d2c0a75e, 0x3fe2818bef4d3cba,
-	0x3feb16742a4ca2f5, 0x3fe1097248d0a957, 0xbfe1097248d0a957, 0x3feb16742a4ca2f5,
-	0x3fcc6d90535d74dd, 0x3fef33685a3aaef0, 0xbfef33685a3aaef0, 0x3fcc6d90535d74dd,
-	0x3fede4160f6d8d81, 0x3fd6d998638a0cb6, 0xbfd6d998638a0cb6, 0x3fede4160f6d8d81,
-	0x3fda1d6543b50ac0, 0x3fed36fc7bcbfbdc, 0xbfed36fc7bcbfbdc, 0x3fda1d6543b50ac0,
-	0x3fe73e558e079942, 0x3fe5fe7cbde56a10, 0xbfe5fe7cbde56a10, 0x3fe73e558e079942,
-	0x3f9c454f4ce53b1d, 0x3feffce09ce2a679, 0xbfeffce09ce2a679, 0x3f9c454f4ce53b1d,
-	0x3feff753bb1b9164, 0x3fa78dbaa5874686, 0xbfa78dbaa5874686, 0x3feff753bb1b9164,
-	0x3fe59001d5f723df, 0x3fe7a4f707bf97d2, 0xbfe7a4f707bf97d2, 0x3fe59001d5f723df,
-	0x3fecf830e8ce467b, 0x3fdb2f971db31972, 0xbfdb2f971db31972, 0x3fecf830e8ce467b,
-	0x3fd5bee78b9db3b6, 0x3fee18a02fdc66d9, 0xbfee18a02fdc66d9, 0x3fd5bee78b9db3b6,
-	0x3fef1090bc898f5f, 0x3fceb86b462de348, 0xbfceb86b462de348, 0x3fef1090bc898f5f,
-	0x3fe089112032b08c, 0x3feb658f14fdbc47, 0xbfeb658f14fdbc47, 0x3fe089112032b08c,
-	0x3fe9c2d110f075c2, 0x3fe2fbc24b441015, 0xbfe2fbc24b441015, 0x3fe9c2d110f075c2,
-	0x3fc32b7bf94516a7, 0x3fefa39bac7a1791, 0xbfefa39bac7a1791, 0x3fc32b7bf94516a7,
-	0x3fefaafbcb0cfddc, 0x3fc264994dfd3409, 0xbfc264994dfd3409, 0x3fefaafbcb0cfddc,
-	0x3fe32421ec49a61f, 0x3fe9a4dfa42b06b2, 0xbfe9a4dfa42b06b2, 0x3fe32421ec49a61f,
-	0x3feb7f6686e792e9, 0x3fe05df3ec31b8b7, 0xbfe05df3ec31b8b7, 0x3feb7f6686e792e9,
-	0x3fcf7b7480bd3802, 0x3fef045a14cf738c, 0xbfef045a14cf738c, 0x3fcf7b7480bd3802,
-	0x3fee298f4439197a, 0x3fd5604012f467b4, 0xbfd5604012f467b4, 0x3fee298f4439197a,
-	0x3fdb8a7814fd5693, 0x3fece2b32799a060, 0xbfece2b32799a060, 0x3fdb8a7814fd5693,
-	0x3fe7c6b89ce2d333, 0x3fe56ac35197649f, 0xbfe56ac35197649f, 0x3fe7c6b89ce2d333,
-	0x3faab101bd5f8317, 0x3feff4dc54b1bed3, 0xbfeff4dc54b1bed3, 0x3faab101bd5f8317,
-	0x3fefdafa7514538c, 0x3fb84f8712c130a1, 0xbfb84f8712c130a1, 0x3fefdafa7514538c,
-	0x3fe4605a692b32a2, 0x3fe8ac871ede1d88, 0xbfe8ac871ede1d88, 0x3fe4605a692b32a2,
-	0x3fec44833141c004, 0x3fddfeff66a941de, 0xbfddfeff66a941de, 0x3fec44833141c004,
-	0x3fd2c41a4e954520, 0x3fee97ec36016b30, 0xbfee97ec36016b30, 0x3fd2c41a4e954520,
-	0x3feea68393e65800, 0x3fd263e6995554ba, 0xbfd263e6995554ba, 0x3feea68393e65800,
-	0x3fde57a86d3cd825, 0x3fec2cd14931e3f1, 0xbfec2cd14931e3f1, 0x3fde57a86d3cd825,
-	0x3fe8cc6a75184655, 0x3fe4397f5b2a4380, 0xbfe4397f5b2a4380, 0x3fe8cc6a75184655,
-	0x3fb9dfb6eb24a85c, 0x3fefd60d2da75c9e, 0xbfefd60d2da75c9e, 0x3fb9dfb6eb24a85c,
-	0x3fef677556883cee, 0x3fc8961727c41804, 0xbfc8961727c41804, 0x3fef677556883cee,
-	0x3fe1dc1b64dc4872, 0x3fea8d676e545ad2, 0xbfea8d676e545ad2, 0x3fe1dc1b64dc4872,
-	0x3feaa9547a2cb98e, 0x3fe1b250171373bf, 0xbfe1b250171373bf, 0x3feaa9547a2cb98e,
-	0x3fc95b49e9b62afa, 0x3fef5da6ed43685d, 0xbfef5da6ed43685d, 0x3fc95b49e9b62afa,
-	0x3fed9a00dd8b3d46, 0x3fd84f6aaaf3903f, 0xbfd84f6aaaf3903f, 0x3fed9a00dd8b3d46,
-	0x3fd8ac4b86d5ed44, 0x3fed86c48445a44f, 0xbfed86c48445a44f, 0x3fd8ac4b86d5ed44,
-	0x3fe6b25ced2fe29c, 0x3fe68ed1eaa19c71, 0xbfe68ed1eaa19c71, 0x3fe6b25ced2fe29c,
-	0x3f6921f8becca4ba, 0x3feffff621621d02, 0xbfeffff621621d02, 0x3f6921f8becca4ba,
-}
