+++ crypto/internal/falcon1024/tables.go
@@ -1,0 +1,1399 @@
+package falcon1024
+
+// gmb and igmb use the Falcon reference names for forward and inverse NTT roots.
+// Values are in the Falcon reference 2^16 Montgomery domain.
+var (
+	gmb = [...]fieldElement{
+		4091, 7888, 11060, 11208, 6960, 4342, 6275, 9759,
+		1591, 6399, 9477, 5266, 586, 5825, 7538, 9710,
+		1134, 6407, 1711, 965, 7099, 7674, 3743, 6442,
+		10414, 8100, 1885, 1688, 1364, 10329, 10164, 9180,
+		12210, 6240, 997, 117, 4783, 4407, 1549, 7072,
+		2829, 6458, 4431, 8877, 7144, 2564, 5664, 4042,
+		12189, 432, 10751, 1237, 7610, 1534, 3983, 7863,
+		2181, 6308, 8720, 6570, 4843, 1690, 14, 3872,
+		5569, 9368, 12163, 2019, 7543, 2315, 4673, 7340,
+		1553, 1156, 8401, 11389, 1020, 2967, 10772, 7045,
+		3316, 11236, 5285, 11578, 10637, 10086, 9493, 6180,
+		9277, 6130, 3323, 883, 10469, 489, 1502, 2851,
+		11061, 9729, 2742, 12241, 4970, 10481, 10078, 1195,
+		730, 1762, 3854, 2030, 5892, 10922, 9020, 5274,
+		9179, 3604, 3782, 10206, 3180, 3467, 4668, 2446,
+		7613, 9386, 834, 7703, 6836, 3403, 5351, 12276,
+		3580, 1739, 10820, 9787, 10209, 4070, 12250, 8525,
+		10401, 2749, 7338, 10574, 6040, 943, 9330, 1477,
+		6865, 9668, 3585, 6633, 12145, 4063, 3684, 7680,
+		8188, 6902, 3533, 9807, 6090, 727, 10099, 7003,
+		6945, 1949, 9731, 10559, 6057, 378, 7871, 8763,
+		8901, 9229, 8846, 4551, 9589, 11664, 7630, 8821,
+		5680, 4956, 6251, 8388, 10156, 8723, 2341, 3159,
+		1467, 5460, 8553, 7783, 2649, 2320, 9036, 6188,
+		737, 3698, 4699, 5753, 9046, 3687, 16, 914,
+		5186, 10531, 4552, 1964, 3509, 8436, 7516, 5381,
+		10733, 3281, 7037, 1060, 2895, 7156, 8887, 5357,
+		6409, 8197, 2962, 6375, 5064, 6634, 5625, 278,
+		932, 10229, 8927, 7642, 351, 9298, 237, 5858,
+		7692, 3146, 12126, 7586, 2053, 11285, 3802, 5204,
+		4602, 1748, 11300, 340, 3711, 4614, 300, 10993,
+		5070, 10049, 11616, 12247, 7421, 10707, 5746, 5654,
+		3835, 5553, 1224, 8476, 9237, 3845, 250, 11209,
+		4225, 6326, 9680, 12254, 4136, 2778, 692, 8808,
+		6410, 6718, 10105, 10418, 3759, 7356, 11361, 8433,
+		6437, 3652, 6342, 8978, 5391, 2272, 6476, 7416,
+		8418, 10824, 11986, 5733, 876, 7030, 2167, 2436,
+		3442, 9217, 8206, 4858, 5964, 2746, 7178, 1434,
+		7389, 8879, 10661, 11457, 4220, 1432, 10832, 4328,
+		8557, 1867, 9454, 2416, 3816, 9076, 686, 5393,
+		2523, 4339, 6115, 619, 937, 2834, 7775, 3279,
+		2363, 7488, 6112, 5056, 824, 10204, 11690, 1113,
+		2727, 9848, 896, 2028, 5075, 2654, 10464, 7884,
+		12169, 5434, 3070, 6400, 9132, 11672, 12153, 4520,
+		1273, 9739, 11468, 9937, 10039, 9720, 2262, 9399,
+		11192, 315, 4511, 1158, 6061, 6751, 11865, 357,
+		7367, 4550, 983, 8534, 8352, 10126, 7530, 9253,
+		4367, 5221, 3999, 8777, 3161, 6990, 4130, 11652,
+		3374, 11477, 1753, 292, 8681, 2806, 10378, 12188,
+		5800, 11811, 3181, 1988, 1024, 9340, 2477, 10928,
+		4582, 6750, 3619, 5503, 5233, 2463, 8470, 7650,
+		7964, 6395, 1071, 1272, 3474, 11045, 3291, 11344,
+		8502, 9478, 9837, 1253, 1857, 6233, 4720, 11561,
+		6034, 9817, 3339, 1797, 2879, 6242, 5200, 2114,
+		7962, 9353, 11363, 5475, 6084, 9601, 4108, 7323,
+		10438, 9471, 1271, 408, 6911, 3079, 360, 8276,
+		11535, 9156, 9049, 11539, 850, 8617, 784, 7919,
+		8334, 12170, 1846, 10213, 12184, 7827, 11903, 5600,
+		9779, 1012, 721, 2784, 6676, 6552, 5348, 4424,
+		6816, 8405, 9959, 5150, 2356, 5552, 5267, 1333,
+		8801, 9661, 7308, 5788, 4910, 909, 11613, 4395,
+		8238, 6686, 4302, 3044, 2285, 12249, 1963, 9216,
+		4296, 11918, 695, 4371, 9793, 4884, 2411, 10230,
+		2650, 841, 3890, 10231, 7248, 8505, 11196, 6688,
+		4059, 6060, 3686, 4722, 11853, 5816, 7058, 6868,
+		11137, 7926, 4894, 12284, 4102, 3908, 3610, 6525,
+		7938, 7982, 11977, 6755, 537, 4562, 1623, 8227,
+		11453, 7544, 906, 11816, 9548, 10858, 9703, 2815,
+		11736, 6813, 6979, 819, 8903, 6271, 10843, 348,
+		7514, 8339, 6439, 694, 852, 5659, 2781, 3716,
+		11589, 3024, 1523, 8659, 4114, 10738, 3303, 5885,
+		2978, 7289, 11884, 9123, 9323, 11830, 98, 2526,
+		2116, 4131, 11407, 1844, 3645, 3916, 8133, 2224,
+		10871, 8092, 9651, 5989, 7140, 8480, 1670, 159,
+		10923, 4918, 128, 7312, 725, 9157, 5006, 6393,
+		3494, 6043, 10972, 6181, 11838, 3423, 10514, 7668,
+		3693, 6658, 6905, 11953, 10212, 11922, 9101, 8365,
+		5110, 45, 2400, 1921, 4377, 2720, 1695, 51,
+		2808, 650, 1896, 9997, 9971, 11980, 8098, 4833,
+		4135, 4257, 5838, 4765, 10985, 11532, 590, 12198,
+		482, 12173, 2006, 7064, 10018, 3912, 12016, 10519,
+		11362, 6954, 2210, 284, 5413, 6601, 3865, 10339,
+		11188, 6231, 517, 9564, 11281, 3863, 1210, 4604,
+		8160, 11447, 153, 7204, 5763, 5089, 9248, 12154,
+		11748, 1354, 6672, 179, 5532, 2646, 5941, 12185,
+		862, 3158, 477, 7279, 5678, 7914, 4254, 302,
+		2893, 10114, 6890, 9560, 9647, 11905, 4098, 9824,
+		10269, 1353, 10715, 5325, 6254, 3951, 1807, 6449,
+		5159, 1308, 8315, 3404, 1877, 1231, 112, 6398,
+		11724, 12272, 7286, 1459, 12274, 9896, 3456, 800,
+		1397, 10678, 103, 7420, 7976, 936, 764, 632,
+		7996, 8223, 8445, 7758, 10870, 9571, 2508, 1946,
+		6524, 10158, 1044, 4338, 2457, 3641, 1659, 4139,
+		4688, 9733, 11148, 3946, 2082, 5261, 2036, 11850,
+		7636, 12236, 5366, 2380, 1399, 7720, 2100, 3217,
+		10912, 8898, 7578, 11995, 2791, 1215, 3355, 2711,
+		2267, 2004, 8568, 10176, 3214, 2337, 1750, 4729,
+		4997, 7415, 6315, 12044, 4374, 7157, 4844, 211,
+		8003, 10159, 9290, 11481, 1735, 2336, 5793, 9875,
+		8192, 986, 7527, 1401, 870, 3615, 8465, 2756,
+		9770, 2034, 10168, 3264, 6132, 54, 2880, 4763,
+		11805, 3074, 8286, 9428, 4881, 6933, 1090, 10038,
+		2567, 708, 893, 6465, 4962, 10024, 2090, 5718,
+		10743, 780, 4733, 4623, 2134, 2087, 4802, 884,
+		5372, 5795, 5938, 4333, 6559, 7549, 5269, 10664,
+		4252, 3260, 5917, 10814, 5768, 9983, 8096, 7791,
+		6800, 7491, 6272, 1907, 10947, 6289, 11803, 6032,
+		11449, 1171, 9201, 7933, 2479, 7970, 11337, 7062,
+		8911, 6728, 6542, 8114, 8828, 6595, 3545, 4348,
+		4610, 2205, 6999, 8106, 5560, 10390, 9321, 2499,
+		2413, 7272, 6881, 10582, 9308, 9437, 3554, 3326,
+		5991, 11969, 3415, 12283, 9838, 12063, 4332, 7830,
+		11329, 6605, 12271, 2044, 11611, 7353, 11201, 11582,
+		3733, 8943, 9978, 1627, 7168, 3935, 5050, 2762,
+		7496, 10383, 755, 1654, 12053, 4952, 10134, 4394,
+		6592, 7898, 7497, 8904, 12029, 3581, 10748, 5674,
+		10358, 4901, 7414, 8771, 710, 6764, 8462, 7193,
+		5371, 7274, 11084, 290, 7864, 6827, 11822, 2509,
+		6578, 4026, 5807, 1458, 5721, 5762, 4178, 2105,
+		11621, 4852, 8897, 2856, 11510, 9264, 2520, 8776,
+		7011, 2647, 1898, 7039, 5950, 11163, 5488, 6277,
+		9182, 11456, 633, 10046, 11554, 5633, 9587, 2333,
+		7008, 7084, 5047, 7199, 9865, 8997, 569, 6390,
+		10845, 9679, 8268, 11472, 4203, 1997, 2, 9331,
+		162, 6182, 2000, 3649, 9792, 6363, 7557, 6187,
+		8510, 9935, 5536, 9019, 3706, 12009, 1452, 3067,
+		5494, 9692, 4865, 6019, 7106, 9610, 4588, 10165,
+		6261, 5887, 2652, 10172, 1580, 10379, 4638, 9949,
+	}
+
+	igmb = [...]fieldElement{
+		4091, 4401, 1081, 1229, 2530, 6014, 7947, 5329,
+		2579, 4751, 6464, 11703, 7023, 2812, 5890, 10698,
+		3109, 2125, 1960, 10925, 10601, 10404, 4189, 1875,
+		5847, 8546, 4615, 5190, 11324, 10578, 5882, 11155,
+		8417, 12275, 10599, 7446, 5719, 3569, 5981, 10108,
+		4426, 8306, 10755, 4679, 11052, 1538, 11857, 100,
+		8247, 6625, 9725, 5145, 3412, 7858, 5831, 9460,
+		5217, 10740, 7882, 7506, 12172, 11292, 6049, 79,
+		13, 6938, 8886, 5453, 4586, 11455, 2903, 4676,
+		9843, 7621, 8822, 9109, 2083, 8507, 8685, 3110,
+		7015, 3269, 1367, 6397, 10259, 8435, 10527, 11559,
+		11094, 2211, 1808, 7319, 48, 9547, 2560, 1228,
+		9438, 10787, 11800, 1820, 11406, 8966, 6159, 3012,
+		6109, 2796, 2203, 1652, 711, 7004, 1053, 8973,
+		5244, 1517, 9322, 11269, 900, 3888, 11133, 10736,
+		4949, 7616, 9974, 4746, 10270, 126, 2921, 6720,
+		6635, 6543, 1582, 4868, 42, 673, 2240, 7219,
+		1296, 11989, 7675, 8578, 11949, 989, 10541, 7687,
+		7085, 8487, 1004, 10236, 4703, 163, 9143, 4597,
+		6431, 12052, 2991, 11938, 4647, 3362, 2060, 11357,
+		12011, 6664, 5655, 7225, 5914, 9327, 4092, 5880,
+		6932, 3402, 5133, 9394, 11229, 5252, 9008, 1556,
+		6908, 4773, 3853, 8780, 10325, 7737, 1758, 7103,
+		11375, 12273, 8602, 3243, 6536, 7590, 8591, 11552,
+		6101, 3253, 9969, 9640, 4506, 3736, 6829, 10822,
+		9130, 9948, 3566, 2133, 3901, 6038, 7333, 6609,
+		3468, 4659, 625, 2700, 7738, 3443, 3060, 3388,
+		3526, 4418, 11911, 6232, 1730, 2558, 10340, 5344,
+		5286, 2190, 11562, 6199, 2482, 8756, 5387, 4101,
+		4609, 8605, 8226, 144, 5656, 8704, 2621, 5424,
+		10812, 2959, 11346, 6249, 1715, 4951, 9540, 1888,
+		3764, 39, 8219, 2080, 2502, 1469, 10550, 8709,
+		5601, 1093, 3784, 5041, 2058, 8399, 11448, 9639,
+		2059, 9878, 7405, 2496, 7918, 11594, 371, 7993,
+		3073, 10326, 40, 10004, 9245, 7987, 5603, 4051,
+		7894, 676, 11380, 7379, 6501, 4981, 2628, 3488,
+		10956, 7022, 6737, 9933, 7139, 2330, 3884, 5473,
+		7865, 6941, 5737, 5613, 9505, 11568, 11277, 2510,
+		6689, 386, 4462, 105, 2076, 10443, 119, 3955,
+		4370, 11505, 3672, 11439, 750, 3240, 3133, 754,
+		4013, 11929, 9210, 5378, 11881, 11018, 2818, 1851,
+		4966, 8181, 2688, 6205, 6814, 926, 2936, 4327,
+		10175, 7089, 6047, 9410, 10492, 8950, 2472, 6255,
+		728, 7569, 6056, 10432, 11036, 2452, 2811, 3787,
+		945, 8998, 1244, 8815, 11017, 11218, 5894, 4325,
+		4639, 3819, 9826, 7056, 6786, 8670, 5539, 7707,
+		1361, 9812, 2949, 11265, 10301, 9108, 478, 6489,
+		101, 1911, 9483, 3608, 11997, 10536, 812, 8915,
+		637, 8159, 5299, 9128, 3512, 8290, 7068, 7922,
+		3036, 4759, 2163, 3937, 3755, 11306, 7739, 4922,
+		11932, 424, 5538, 6228, 11131, 7778, 11974, 1097,
+		2890, 10027, 2569, 2250, 2352, 821, 2550, 11016,
+		7769, 136, 617, 3157, 5889, 9219, 6855, 120,
+		4405, 1825, 9635, 7214, 10261, 11393, 2441, 9562,
+		11176, 599, 2085, 11465, 7233, 6177, 4801, 9926,
+		9010, 4514, 9455, 11352, 11670, 6174, 7950, 9766,
+		6896, 11603, 3213, 8473, 9873, 2835, 10422, 3732,
+		7961, 1457, 10857, 8069, 832, 1628, 3410, 4900,
+		10855, 5111, 9543, 6325, 7431, 4083, 3072, 8847,
+		9853, 10122, 5259, 11413, 6556, 303, 1465, 3871,
+		4873, 5813, 10017, 6898, 3311, 5947, 8637, 5852,
+		3856, 928, 4933, 8530, 1871, 2184, 5571, 5879,
+		3481, 11597, 9511, 8153, 35, 2609, 5963, 8064,
+		1080, 12039, 8444, 3052, 3813, 11065, 6736, 8454,
+		2340, 7651, 1910, 10709, 2117, 9637, 6402, 6028,
+		2124, 7701, 2679, 5183, 6270, 7424, 2597, 6795,
+		9222, 10837, 280, 8583, 3270, 6753, 2354, 3779,
+		6102, 4732, 5926, 2497, 8640, 10289, 6107, 12127,
+		2958, 12287, 10292, 8086, 817, 4021, 2610, 1444,
+		5899, 11720, 3292, 2424, 5090, 7242, 5205, 5281,
+		9956, 2702, 6656, 735, 2243, 11656, 833, 3107,
+		6012, 6801, 1126, 6339, 5250, 10391, 9642, 5278,
+		3513, 9769, 3025, 779, 9433, 3392, 7437, 668,
+		10184, 8111, 6527, 6568, 10831, 6482, 8263, 5711,
+		9780, 467, 5462, 4425, 11999, 1205, 5015, 6918,
+		5096, 3827, 5525, 11579, 3518, 4875, 7388, 1931,
+		6615, 1541, 8708, 260, 3385, 4792, 4391, 5697,
+		7895, 2155, 7337, 236, 10635, 11534, 1906, 4793,
+		9527, 7239, 8354, 5121, 10662, 2311, 3346, 8556,
+		707, 1088, 4936, 678, 10245, 18, 5684, 960,
+		4459, 7957, 226, 2451, 6, 8874, 320, 6298,
+		8963, 8735, 2852, 2981, 1707, 5408, 5017, 9876,
+		9790, 2968, 1899, 6729, 4183, 5290, 10084, 7679,
+		7941, 8744, 5694, 3461, 4175, 5747, 5561, 3378,
+		5227, 952, 4319, 9810, 4356, 3088, 11118, 840,
+		6257, 486, 6000, 1342, 10382, 6017, 4798, 5489,
+		4498, 4193, 2306, 6521, 1475, 6372, 9029, 8037,
+		1625, 7020, 4740, 5730, 7956, 6351, 6494, 6917,
+		11405, 7487, 10202, 10155, 7666, 7556, 11509, 1546,
+		6571, 10199, 2265, 7327, 5824, 11396, 11581, 9722,
+		2251, 11199, 5356, 7408, 2861, 4003, 9215, 484,
+		7526, 9409, 12235, 6157, 9025, 2121, 10255, 2519,
+		9533, 3824, 8674, 11419, 10888, 4762, 11303, 4097,
+		2414, 6496, 9953, 10554, 808, 2999, 2130, 4286,
+		12078, 7445, 5132, 7915, 245, 5974, 4874, 7292,
+		7560, 10539, 9952, 9075, 2113, 3721, 10285, 10022,
+		9578, 8934, 11074, 9498, 294, 4711, 3391, 1377,
+		9072, 10189, 4569, 10890, 9909, 6923, 53, 4653,
+		439, 10253, 7028, 10207, 8343, 1141, 2556, 7601,
+		8150, 10630, 8648, 9832, 7951, 11245, 2131, 5765,
+		10343, 9781, 2718, 1419, 4531, 3844, 4066, 4293,
+		11657, 11525, 11353, 4313, 4869, 12186, 1611, 10892,
+		11489, 8833, 2393, 15, 10830, 5003, 17, 565,
+		5891, 12177, 11058, 10412, 8885, 3974, 10981, 7130,
+		5840, 10482, 8338, 6035, 6964, 1574, 10936, 2020,
+		2465, 8191, 384, 2642, 2729, 5399, 2175, 9396,
+		11987, 8035, 4375, 6611, 5010, 11812, 9131, 11427,
+		104, 6348, 9643, 6757, 12110, 5617, 10935, 541,
+		135, 3041, 7200, 6526, 5085, 12136, 842, 4129,
+		7685, 11079, 8426, 1008, 2725, 11772, 6058, 1101,
+		1950, 8424, 5688, 6876, 12005, 10079, 5335, 927,
+		1770, 273, 8377, 2271, 5225, 10283, 116, 11807,
+		91, 11699, 757, 1304, 7524, 6451, 8032, 8154,
+		7456, 4191, 309, 2318, 2292, 10393, 11639, 9481,
+		12238, 10594, 9569, 7912, 10368, 9889, 12244, 7179,
+		3924, 3188, 367, 2077, 336, 5384, 5631, 8596,
+		4621, 1775, 8866, 451, 6108, 1317, 6246, 8795,
+		5896, 7283, 3132, 11564, 4977, 12161, 7371, 1366,
+		12130, 10619, 3809, 5149, 6300, 2638, 4197, 1418,
+		10065, 4156, 8373, 8644, 10445, 882, 8158, 10173,
+		9763, 12191, 459, 2966, 3166, 405, 5000, 9311,
+		6404, 8986, 1551, 8175, 3630, 10766, 9265, 700,
+		8573, 9508, 6630, 11437, 11595, 5850, 3950, 4775,
+		11941, 1446, 6018, 3386, 11470, 5310, 5476, 553,
+		9474, 2586, 1431, 2741, 473, 11383, 4745, 836,
+		4062, 10666, 7727, 11752, 5534, 312, 4307, 4351,
+		5764, 8679, 8381, 8187, 5, 7395, 4363, 1152,
+		5421, 5231, 6473, 436, 7567, 8603, 6229, 8230,
+	}
+)
+
+var (
+	maxBlSmall = [...]int{1, 1, 2, 2, 4, 7, 14, 27, 53, 106, 209}
+	maxBlLarge = [...]int{2, 2, 5, 7, 12, 21, 40, 78, 157, 308}
+	bitLength  = [...]struct{ avg, std int }{
+		{4, 0},
+		{11, 1},
+		{24, 1},
+		{50, 1},
+		{102, 1},
+		{202, 2},
+		{401, 4},
+		{794, 5},
+		{1577, 8},
+		{3138, 13},
+		{6308, 25},
+	}
+)
+
+type smallPrime struct {
+	p, g, s uint32
+}
+
+var primes = [...]smallPrime{
+	{p: 2147473409, g: 383167813, s: 10239},
+	{p: 2147389441, g: 211808905, s: 471403745},
+	{p: 2147387393, g: 37672282, s: 1329335065},
+	{p: 2147377153, g: 1977035326, s: 968223422},
+	{p: 2147358721, g: 1067163706, s: 132460015},
+	{p: 2147352577, g: 1606082042, s: 598693809},
+	{p: 2147346433, g: 2033915641, s: 1056257184},
+	{p: 2147338241, g: 1653770625, s: 421286710},
+	{p: 2147309569, g: 631200819, s: 1111201074},
+	{p: 2147297281, g: 2038364663, s: 1042003613},
+	{p: 2147295233, g: 1962540515, s: 19440033},
+	{p: 2147239937, g: 2100082663, s: 353296760},
+	{p: 2147235841, g: 1991153006, s: 1703918027},
+	{p: 2147217409, g: 516405114, s: 1258919613},
+	{p: 2147205121, g: 409347988, s: 1089726929},
+	{p: 2147196929, g: 927788991, s: 1946238668},
+	{p: 2147178497, g: 1136922411, s: 1347028164},
+	{p: 2147100673, g: 868626236, s: 701164723},
+	{p: 2147082241, g: 1897279176, s: 617820870},
+	{p: 2147074049, g: 1888819123, s: 158382189},
+	{p: 2147051521, g: 25006327, s: 522758543},
+	{p: 2147043329, g: 327546255, s: 37227845},
+	{p: 2147039233, g: 766324424, s: 1133356428},
+	{p: 2146988033, g: 1862817362, s: 73861329},
+	{p: 2146963457, g: 404622040, s: 653019435},
+	{p: 2146959361, g: 1936581214, s: 995143093},
+	{p: 2146938881, g: 1559770096, s: 634921513},
+	{p: 2146908161, g: 422623708, s: 1985060172},
+	{p: 2146885633, g: 1751189170, s: 298238186},
+	{p: 2146871297, g: 578919515, s: 291810829},
+	{p: 2146846721, g: 1114060353, s: 915902322},
+	{p: 2146834433, g: 2069565474, s: 47859524},
+	{p: 2146818049, g: 1552824584, s: 646281055},
+	{p: 2146775041, g: 1906267847, s: 1597832891},
+	{p: 2146756609, g: 1847414714, s: 1228090888},
+	{p: 2146744321, g: 1818792070, s: 1176377637},
+	{p: 2146738177, g: 1118066398, s: 1054971214},
+	{p: 2146736129, g: 52057278, s: 933422153},
+	{p: 2146713601, g: 592259376, s: 1406621510},
+	{p: 2146695169, g: 263161877, s: 1514178701},
+	{p: 2146656257, g: 685363115, s: 384505091},
+	{p: 2146650113, g: 927727032, s: 537575289},
+	{p: 2146646017, g: 52575506, s: 1799464037},
+	{p: 2146643969, g: 1276803876, s: 1348954416},
+	{p: 2146603009, g: 814028633, s: 1521547704},
+	{p: 2146572289, g: 1846678872, s: 1310832121},
+	{p: 2146547713, g: 919368090, s: 1019041349},
+	{p: 2146508801, g: 671847612, s: 38582496},
+	{p: 2146492417, g: 283911680, s: 532424562},
+	{p: 2146490369, g: 1780044827, s: 896447978},
+	{p: 2146459649, g: 327980850, s: 1327906900},
+	{p: 2146447361, g: 1310561493, s: 958645253},
+	{p: 2146441217, g: 412148926, s: 287271128},
+	{p: 2146437121, g: 293186449, s: 2009822534},
+	{p: 2146430977, g: 179034356, s: 1359155584},
+	{p: 2146418689, g: 1517345488, s: 1790248672},
+	{p: 2146406401, g: 1615820390, s: 1584833571},
+	{p: 2146404353, g: 826651445, s: 607120498},
+	{p: 2146379777, g: 3816988, s: 1897049071},
+	{p: 2146363393, g: 1221409784, s: 1986921567},
+	{p: 2146355201, g: 1388081168, s: 849968120},
+	{p: 2146336769, g: 1803473237, s: 1655544036},
+	{p: 2146312193, g: 1023484977, s: 273671831},
+	{p: 2146293761, g: 1074591448, s: 467406983},
+	{p: 2146283521, g: 831604668, s: 1523950494},
+	{p: 2146203649, g: 712865423, s: 1170834574},
+	{p: 2146154497, g: 1764991362, s: 1064856763},
+	{p: 2146142209, g: 627386213, s: 1406840151},
+	{p: 2146127873, g: 1638674429, s: 2088393537},
+	{p: 2146099201, g: 1516001018, s: 690673370},
+	{p: 2146093057, g: 1294931393, s: 315136610},
+	{p: 2146091009, g: 1942399533, s: 973539425},
+	{p: 2146078721, g: 1843461814, s: 2132275436},
+	{p: 2146060289, g: 1098740778, s: 360423481},
+	{p: 2146048001, g: 1617213232, s: 1951981294},
+	{p: 2146041857, g: 1805783169, s: 2075683489},
+	{p: 2146019329, g: 272027909, s: 1753219918},
+	{p: 2145986561, g: 1206530344, s: 2034028118},
+	{p: 2145976321, g: 1243769360, s: 1173377644},
+	{p: 2145964033, g: 887200839, s: 1281344586},
+	{p: 2145906689, g: 1651026455, s: 906178216},
+	{p: 2145875969, g: 1673238256, s: 1043521212},
+	{p: 2145871873, g: 1226591210, s: 1399796492},
+	{p: 2145841153, g: 1465353397, s: 1324527802},
+	{p: 2145832961, g: 1150638905, s: 554084759},
+	{p: 2145816577, g: 221601706, s: 427340863},
+	{p: 2145785857, g: 608896761, s: 316590738},
+	{p: 2145755137, g: 1712054942, s: 1684294304},
+	{p: 2145742849, g: 1302302867, s: 724873116},
+	{p: 2145728513, g: 516717693, s: 431671476},
+	{p: 2145699841, g: 524575579, s: 1619722537},
+	{p: 2145691649, g: 1925625239, s: 982974435},
+	{p: 2145687553, g: 463795662, s: 1293154300},
+	{p: 2145673217, g: 771716636, s: 881778029},
+	{p: 2145630209, g: 1509556977, s: 837364988},
+	{p: 2145595393, g: 229091856, s: 851648427},
+	{p: 2145587201, g: 1796903241, s: 635342424},
+	{p: 2145525761, g: 715310882, s: 1677228081},
+	{p: 2145495041, g: 1040930522, s: 200685896},
+	{p: 2145466369, g: 949804237, s: 1809146322},
+	{p: 2145445889, g: 1673903706, s: 95316881},
+	{p: 2145390593, g: 806941852, s: 1428671135},
+	{p: 2145372161, g: 1402525292, s: 159350694},
+	{p: 2145361921, g: 2124760298, s: 1589134749},
+	{p: 2145359873, g: 1217503067, s: 1561543010},
+	{p: 2145355777, g: 338341402, s: 83865711},
+	{p: 2145343489, g: 1381532164, s: 641430002},
+	{p: 2145325057, g: 1883895478, s: 1528469895},
+	{p: 2145318913, g: 1335370424, s: 65809740},
+	{p: 2145312769, g: 2000008042, s: 1919775760},
+	{p: 2145300481, g: 961450962, s: 1229540578},
+	{p: 2145282049, g: 910466767, s: 1964062701},
+	{p: 2145232897, g: 816527501, s: 450152063},
+	{p: 2145218561, g: 1435128058, s: 1794509700},
+	{p: 2145187841, g: 33505311, s: 1272467582},
+	{p: 2145181697, g: 269767433, s: 1380363849},
+	{p: 2145175553, g: 56386299, s: 1316870546},
+	{p: 2145079297, g: 2106880293, s: 1391797340},
+	{p: 2145021953, g: 1347906152, s: 720510798},
+	{p: 2145015809, g: 206769262, s: 1651459955},
+	{p: 2145003521, g: 1885513236, s: 1393381284},
+	{p: 2144960513, g: 1810381315, s: 31937275},
+	{p: 2144944129, g: 1306487838, s: 2019419520},
+	{p: 2144935937, g: 37304730, s: 1841489054},
+	{p: 2144894977, g: 1601434616, s: 157985831},
+	{p: 2144888833, g: 98749330, s: 2128592228},
+	{p: 2144880641, g: 1772327002, s: 2076128344},
+	{p: 2144864257, g: 1404514762, s: 2029969964},
+	{p: 2144827393, g: 801236594, s: 406627220},
+	{p: 2144806913, g: 349217443, s: 1501080290},
+	{p: 2144796673, g: 1542656776, s: 2084736519},
+	{p: 2144778241, g: 1210734884, s: 1746416203},
+	{p: 2144759809, g: 1146598851, s: 716464489},
+	{p: 2144757761, g: 286328400, s: 1823728177},
+	{p: 2144729089, g: 1347555695, s: 1836644881},
+	{p: 2144727041, g: 1795703790, s: 520296412},
+	{p: 2144696321, g: 1302475157, s: 852964281},
+	{p: 2144667649, g: 1075877614, s: 504992927},
+	{p: 2144573441, g: 198765808, s: 1617144982},
+	{p: 2144555009, g: 321528767, s: 155821259},
+	{p: 2144550913, g: 814139516, s: 1819937644},
+	{p: 2144536577, g: 571143206, s: 962942255},
+	{p: 2144524289, g: 1746733766, s: 2471321},
+	{p: 2144512001, g: 1821415077, s: 124190939},
+	{p: 2144468993, g: 917871546, s: 1260072806},
+	{p: 2144458753, g: 378417981, s: 1569240563},
+	{p: 2144421889, g: 175229668, s: 1825620763},
+	{p: 2144409601, g: 1699216963, s: 351648117},
+	{p: 2144370689, g: 1071885991, s: 958186029},
+	{p: 2144348161, g: 1763151227, s: 540353574},
+	{p: 2144335873, g: 1060214804, s: 919598847},
+	{p: 2144329729, g: 663515846, s: 1448552668},
+	{p: 2144327681, g: 1057776305, s: 590222840},
+	{p: 2144309249, g: 1705149168, s: 1459294624},
+	{p: 2144296961, g: 325823721, s: 1649016934},
+	{p: 2144290817, g: 738775789, s: 447427206},
+	{p: 2144243713, g: 962347618, s: 893050215},
+	{p: 2144237569, g: 1655257077, s: 900860862},
+	{p: 2144161793, g: 242206694, s: 1567868672},
+	{p: 2144155649, g: 769415308, s: 1247993134},
+	{p: 2144137217, g: 320492023, s: 515841070},
+	{p: 2144120833, g: 1639388522, s: 770877302},
+	{p: 2144071681, g: 1761785233, s: 964296120},
+	{p: 2144065537, g: 419817825, s: 204564472},
+	{p: 2144028673, g: 666050597, s: 2091019760},
+	{p: 2144010241, g: 1413657615, s: 1518702610},
+	{p: 2143952897, g: 1238327946, s: 475672271},
+	{p: 2143940609, g: 307063413, s: 1176750846},
+	{p: 2143918081, g: 2062905559, s: 786785803},
+	{p: 2143899649, g: 1338112849, s: 1562292083},
+	{p: 2143891457, g: 68149545, s: 87166451},
+	{p: 2143885313, g: 921750778, s: 394460854},
+	{p: 2143854593, g: 719766593, s: 133877196},
+	{p: 2143836161, g: 1149399850, s: 1861591875},
+	{p: 2143762433, g: 1848739366, s: 1335934145},
+	{p: 2143756289, g: 1326674710, s: 102999236},
+	{p: 2143713281, g: 808061791, s: 1156900308},
+	{p: 2143690753, g: 388399459, s: 1926468019},
+	{p: 2143670273, g: 1427891374, s: 1756689401},
+	{p: 2143666177, g: 1912173949, s: 986629565},
+	{p: 2143645697, g: 2041160111, s: 371842865},
+	{p: 2143641601, g: 1279906897, s: 2023974350},
+	{p: 2143635457, g: 720473174, s: 1389027526},
+	{p: 2143621121, g: 1298309455, s: 1732632006},
+	{p: 2143598593, g: 1548762216, s: 1825417506},
+	{p: 2143567873, g: 620475784, s: 1073787233},
+	{p: 2143561729, g: 1932954575, s: 949167309},
+	{p: 2143553537, g: 354315656, s: 1652037534},
+	{p: 2143541249, g: 577424288, s: 1097027618},
+	{p: 2143531009, g: 357862822, s: 478640055},
+	{p: 2143522817, g: 2017706025, s: 1550531668},
+	{p: 2143506433, g: 2078127419, s: 1824320165},
+	{p: 2143488001, g: 613475285, s: 1604011510},
+	{p: 2143469569, g: 1466594987, s: 502095196},
+	{p: 2143426561, g: 1115430331, s: 1044637111},
+	{p: 2143383553, g: 9778045, s: 1902463734},
+	{p: 2143377409, g: 1557401276, s: 2056861771},
+	{p: 2143363073, g: 652036455, s: 1965915971},
+	{p: 2143260673, g: 1464581171, s: 1523257541},
+	{p: 2143246337, g: 1876119649, s: 764541916},
+	{p: 2143209473, g: 1614992673, s: 1920672844},
+	{p: 2143203329, g: 981052047, s: 2049774209},
+	{p: 2143160321, g: 1847355533, s: 728535665},
+	{p: 2143129601, g: 965558457, s: 603052992},
+	{p: 2143123457, g: 2140817191, s: 8348679},
+	{p: 2143100929, g: 1547263683, s: 694209023},
+	{p: 2143092737, g: 643459066, s: 1979934533},
+	{p: 2143082497, g: 188603778, s: 2026175670},
+	{p: 2143062017, g: 1657329695, s: 377451099},
+	{p: 2143051777, g: 114967950, s: 979255473},
+	{p: 2143025153, g: 1698431342, s: 1449196896},
+	{p: 2143006721, g: 1862741675, s: 1739650365},
+	{p: 2142996481, g: 756660457, s: 996160050},
+	{p: 2142976001, g: 927864010, s: 1166847574},
+	{p: 2142965761, g: 905070557, s: 661974566},
+	{p: 2142916609, g: 40932754, s: 1787161127},
+	{p: 2142892033, g: 1987985648, s: 675335382},
+	{p: 2142885889, g: 797497211, s: 1323096997},
+	{p: 2142871553, g: 2068025830, s: 1411877159},
+	{p: 2142861313, g: 1217177090, s: 1438410687},
+	{p: 2142830593, g: 409906375, s: 1767860634},
+	{p: 2142803969, g: 1197788993, s: 359782919},
+	{p: 2142785537, g: 643817365, s: 513932862},
+	{p: 2142779393, g: 1717046338, s: 218943121},
+	{p: 2142724097, g: 89336830, s: 416687049},
+	{p: 2142707713, g: 5944581, s: 1356813523},
+	{p: 2142658561, g: 887942135, s: 2074011722},
+	{p: 2142638081, g: 151851972, s: 1647339939},
+	{p: 2142564353, g: 1691505537, s: 1483107336},
+	{p: 2142533633, g: 1989920200, s: 1135938817},
+	{p: 2142529537, g: 959263126, s: 1531961857},
+	{p: 2142527489, g: 453251129, s: 1725566162},
+	{p: 2142502913, g: 1536028102, s: 182053257},
+	{p: 2142498817, g: 570138730, s: 701443447},
+	{p: 2142416897, g: 326965800, s: 411931819},
+	{p: 2142363649, g: 1675665410, s: 1517191733},
+	{p: 2142351361, g: 968529566, s: 1575712703},
+	{p: 2142330881, g: 1384953238, s: 1769087884},
+	{p: 2142314497, g: 1977173242, s: 1833745524},
+	{p: 2142289921, g: 95082313, s: 1714775493},
+	{p: 2142283777, g: 109377615, s: 1070584533},
+	{p: 2142277633, g: 16960510, s: 702157145},
+	{p: 2142263297, g: 553850819, s: 431364395},
+	{p: 2142208001, g: 241466367, s: 2053967982},
+	{p: 2142164993, g: 1795661326, s: 1031836848},
+	{p: 2142097409, g: 1212530046, s: 712772031},
+	{p: 2142087169, g: 1763869720, s: 822276067},
+	{p: 2142078977, g: 644065713, s: 1765268066},
+	{p: 2142074881, g: 112671944, s: 643204925},
+	{p: 2142044161, g: 1387785471, s: 1297890174},
+	{p: 2142025729, g: 783885537, s: 1000425730},
+	{p: 2142011393, g: 905662232, s: 1679401033},
+	{p: 2141974529, g: 799788433, s: 468119557},
+	{p: 2141943809, g: 1932544124, s: 449305555},
+	{p: 2141933569, g: 1527403256, s: 841867925},
+	{p: 2141931521, g: 1247076451, s: 743823916},
+	{p: 2141902849, g: 1199660531, s: 401687910},
+	{p: 2141890561, g: 150132350, s: 1720336972},
+	{p: 2141857793, g: 1287438162, s: 663880489},
+	{p: 2141833217, g: 618017731, s: 1819208266},
+	{p: 2141820929, g: 999578638, s: 1403090096},
+	{p: 2141786113, g: 81834325, s: 1523542501},
+	{p: 2141771777, g: 120001928, s: 463556492},
+	{p: 2141759489, g: 122455485, s: 2124928282},
+	{p: 2141749249, g: 141986041, s: 940339153},
+	{p: 2141685761, g: 889088734, s: 477141499},
+	{p: 2141673473, g: 324212681, s: 1122558298},
+	{p: 2141669377, g: 1175806187, s: 1373818177},
+	{p: 2141655041, g: 1113654822, s: 296887082},
+	{p: 2141587457, g: 991103258, s: 1585913875},
+	{p: 2141583361, g: 1401451409, s: 1802457360},
+	{p: 2141575169, g: 1571977166, s: 712760980},
+	{p: 2141546497, g: 1107849376, s: 1250270109},
+	{p: 2141515777, g: 196544219, s: 356001130},
+	{p: 2141495297, g: 1733571506, s: 1060744866},
+	{p: 2141483009, g: 321552363, s: 1168297026},
+	{p: 2141458433, g: 505818251, s: 733225819},
+	{p: 2141360129, g: 1026840098, s: 948342276},
+	{p: 2141325313, g: 945133744, s: 2129965998},
+	{p: 2141317121, g: 1871100260, s: 1843844634},
+	{p: 2141286401, g: 1790639498, s: 1750465696},
+	{p: 2141267969, g: 1376858592, s: 186160720},
+	{p: 2141255681, g: 2129698296, s: 1876677959},
+	{p: 2141243393, g: 2138900688, s: 1340009628},
+	{p: 2141214721, g: 1933049835, s: 1087819477},
+	{p: 2141212673, g: 1898664939, s: 1786328049},
+	{p: 2141202433, g: 990234828, s: 940682169},
+	{p: 2141175809, g: 1406392421, s: 993089586},
+	{p: 2141165569, g: 1263518371, s: 289019479},
+	{p: 2141073409, g: 1485624211, s: 507864514},
+	{p: 2141052929, g: 1885134788, s: 311252465},
+	{p: 2141040641, g: 1285021247, s: 280941862},
+	{p: 2141028353, g: 1527610374, s: 375035110},
+	{p: 2141011969, g: 1400626168, s: 164696620},
+	{p: 2140999681, g: 632959608, s: 966175067},
+	{p: 2140997633, g: 2045628978, s: 1290889438},
+	{p: 2140993537, g: 1412755491, s: 375366253},
+	{p: 2140942337, g: 719477232, s: 785367828},
+	{p: 2140925953, g: 45224252, s: 836552317},
+	{p: 2140917761, g: 1157376588, s: 1001839569},
+	{p: 2140887041, g: 278480752, s: 2098732796},
+	{p: 2140837889, g: 1663139953, s: 924094810},
+	{p: 2140788737, g: 802501511, s: 2045368990},
+	{p: 2140766209, g: 1820083885, s: 1800295504},
+	{p: 2140764161, g: 1169561905, s: 2106792035},
+	{p: 2140696577, g: 127781498, s: 1885987531},
+	{p: 2140684289, g: 16014477, s: 1098116827},
+	{p: 2140653569, g: 665960598, s: 1796728247},
+	{p: 2140594177, g: 1043085491, s: 377310938},
+	{p: 2140579841, g: 1732838211, s: 1504505945},
+	{p: 2140569601, g: 302071939, s: 358291016},
+	{p: 2140567553, g: 192393733, s: 1909137143},
+	{p: 2140557313, g: 406595731, s: 1175330270},
+	{p: 2140549121, g: 1748850918, s: 525007007},
+	{p: 2140477441, g: 499436566, s: 1031159814},
+	{p: 2140469249, g: 1886004401, s: 1029951320},
+	{p: 2140426241, g: 1483168100, s: 1676273461},
+	{p: 2140420097, g: 1779917297, s: 846024476},
+	{p: 2140413953, g: 522948893, s: 1816354149},
+	{p: 2140383233, g: 1931364473, s: 1296921241},
+	{p: 2140366849, g: 1917356555, s: 147196204},
+	{p: 2140354561, g: 16466177, s: 1349052107},
+	{p: 2140348417, g: 1875366972, s: 1860485634},
+	{p: 2140323841, g: 456498717, s: 1790256483},
+	{p: 2140321793, g: 1629493973, s: 150031888},
+	{p: 2140315649, g: 1904063898, s: 395510935},
+	{p: 2140280833, g: 1784104328, s: 831417909},
+	{p: 2140250113, g: 256087139, s: 697349101},
+	{p: 2140229633, g: 388553070, s: 243875754},
+	{p: 2140223489, g: 747459608, s: 1396270850},
+	{p: 2140200961, g: 507423743, s: 1895572209},
+	{p: 2140162049, g: 580106016, s: 2045297469},
+	{p: 2140149761, g: 712426444, s: 785217995},
+	{p: 2140137473, g: 1441607584, s: 536866543},
+	{p: 2140119041, g: 346538902, s: 1740434653},
+	{p: 2140090369, g: 282642885, s: 21051094},
+	{p: 2140076033, g: 1407456228, s: 319910029},
+	{p: 2140047361, g: 1619330500, s: 1488632070},
+	{p: 2140041217, g: 2089408064, s: 2012026134},
+	{p: 2140008449, g: 1705524800, s: 1613440760},
+	{p: 2139924481, g: 1846208233, s: 1280649481},
+	{p: 2139906049, g: 989438755, s: 1185646076},
+	{p: 2139867137, g: 1522314850, s: 372783595},
+	{p: 2139842561, g: 1681587377, s: 216848235},
+	{p: 2139826177, g: 2066284988, s: 1784999464},
+	{p: 2139824129, g: 480888214, s: 1513323027},
+	{p: 2139789313, g: 847937200, s: 858192859},
+	{p: 2139783169, g: 1642000434, s: 1583261448},
+	{p: 2139770881, g: 940699589, s: 179702100},
+	{p: 2139768833, g: 315623242, s: 964612676},
+	{p: 2139666433, g: 331649203, s: 764666914},
+	{p: 2139641857, g: 2118730799, s: 1313764644},
+	{p: 2139635713, g: 519149027, s: 519212449},
+	{p: 2139598849, g: 1526413634, s: 1769667104},
+	{p: 2139574273, g: 551148610, s: 820739925},
+	{p: 2139568129, g: 1386800242, s: 472447405},
+	{p: 2139549697, g: 813760130, s: 1412328531},
+	{p: 2139537409, g: 1615286260, s: 1609362979},
+	{p: 2139475969, g: 1352559299, s: 1696720421},
+	{p: 2139455489, g: 1048691649, s: 1584935400},
+	{p: 2139432961, g: 836025845, s: 950121150},
+	{p: 2139424769, g: 1558281165, s: 1635486858},
+	{p: 2139406337, g: 1728402143, s: 1674423301},
+	{p: 2139396097, g: 1727715782, s: 1483470544},
+	{p: 2139383809, g: 1092853491, s: 1741699084},
+	{p: 2139369473, g: 690776899, s: 1242798709},
+	{p: 2139351041, g: 1768782380, s: 2120712049},
+	{p: 2139334657, g: 1739968247, s: 1427249225},
+	{p: 2139332609, g: 1547189119, s: 623011170},
+	{p: 2139310081, g: 1346827917, s: 1605466350},
+	{p: 2139303937, g: 369317948, s: 828392831},
+	{p: 2139301889, g: 1560417239, s: 1788073219},
+	{p: 2139283457, g: 1303121623, s: 595079358},
+	{p: 2139248641, g: 1354555286, s: 573424177},
+	{p: 2139240449, g: 60974056, s: 885781403},
+	{p: 2139222017, g: 355573421, s: 1221054839},
+	{p: 2139215873, g: 566477826, s: 1724006500},
+	{p: 2139150337, g: 871437673, s: 1609133294},
+	{p: 2139144193, g: 1478130914, s: 1137491905},
+	{p: 2139117569, g: 1854880922, s: 964728507},
+	{p: 2139076609, g: 202405335, s: 756508944},
+	{p: 2139062273, g: 1399715741, s: 884826059},
+	{p: 2139045889, g: 1051045798, s: 1202295476},
+	{p: 2139033601, g: 1707715206, s: 632234634},
+	{p: 2139006977, g: 2035853139, s: 231626690},
+	{p: 2138951681, g: 183867876, s: 838350879},
+	{p: 2138945537, g: 1403254661, s: 404460202},
+	{p: 2138920961, g: 310865011, s: 1282911681},
+	{p: 2138910721, g: 1328496553, s: 103472415},
+	{p: 2138904577, g: 78831681, s: 993513549},
+	{p: 2138902529, g: 1319697451, s: 1055904361},
+	{p: 2138816513, g: 384338872, s: 1706202469},
+	{p: 2138810369, g: 1084868275, s: 405677177},
+	{p: 2138787841, g: 401181788, s: 1964773901},
+	{p: 2138775553, g: 1850532988, s: 1247087473},
+	{p: 2138767361, g: 874261901, s: 1576073565},
+	{p: 2138757121, g: 1187474742, s: 993541415},
+	{p: 2138748929, g: 1782458888, s: 1043206483},
+	{p: 2138744833, g: 1221500487, s: 800141243},
+	{p: 2138738689, g: 413465368, s: 1450660558},
+	{p: 2138695681, g: 739045140, s: 342611472},
+	{p: 2138658817, g: 1355845756, s: 672674190},
+	{p: 2138644481, g: 608379162, s: 1538874380},
+	{p: 2138632193, g: 1444914034, s: 686911254},
+	{p: 2138607617, g: 484707818, s: 1435142134},
+	{p: 2138591233, g: 539460669, s: 1290458549},
+	{p: 2138572801, g: 2093538990, s: 2011138646},
+	{p: 2138552321, g: 1149786988, s: 1076414907},
+	{p: 2138546177, g: 840688206, s: 2108985273},
+	{p: 2138533889, g: 209669619, s: 198172413},
+	{p: 2138523649, g: 1975879426, s: 1277003968},
+	{p: 2138490881, g: 1351891144, s: 1976858109},
+	{p: 2138460161, g: 1817321013, s: 1979278293},
+	{p: 2138429441, g: 1950077177, s: 203441928},
+	{p: 2138400769, g: 908970113, s: 628395069},
+	{p: 2138398721, g: 219890864, s: 758486760},
+	{p: 2138376193, g: 1306654379, s: 977554090},
+	{p: 2138351617, g: 298822498, s: 2004708503},
+	{p: 2138337281, g: 441457816, s: 1049002108},
+	{p: 2138320897, g: 1517731724, s: 1442269609},
+	{p: 2138290177, g: 1355911197, s: 1647139103},
+	{p: 2138234881, g: 531313247, s: 1746591962},
+	{p: 2138214401, g: 1899410930, s: 781416444},
+	{p: 2138202113, g: 1813477173, s: 1622508515},
+	{p: 2138191873, g: 1086458299, s: 1025408615},
+	{p: 2138183681, g: 1998800427, s: 827063290},
+	{p: 2138173441, g: 1921308898, s: 749670117},
+	{p: 2138103809, g: 1620902804, s: 2126787647},
+	{p: 2138099713, g: 828647069, s: 1892961817},
+	{p: 2138085377, g: 179405355, s: 1525506535},
+	{p: 2138060801, g: 615683235, s: 1259580138},
+	{p: 2138044417, g: 2030277840, s: 1731266562},
+	{p: 2138042369, g: 2087222316, s: 1627902259},
+	{p: 2138032129, g: 126388712, s: 1108640984},
+	{p: 2138011649, g: 715026550, s: 1017980050},
+	{p: 2137993217, g: 1693714349, s: 1351778704},
+	{p: 2137888769, g: 1289762259, s: 1053090405},
+	{p: 2137853953, g: 199991890, s: 1254192789},
+	{p: 2137833473, g: 941421685, s: 896995556},
+	{p: 2137817089, g: 750416446, s: 1251031181},
+	{p: 2137792513, g: 798075119, s: 368077456},
+	{p: 2137786369, g: 878543495, s: 1035375025},
+	{p: 2137767937, g: 9351178, s: 1156563902},
+	{p: 2137755649, g: 1382297614, s: 1686559583},
+	{p: 2137724929, g: 1345472850, s: 1681096331},
+	{p: 2137704449, g: 834666929, s: 630551727},
+	{p: 2137673729, g: 1646165729, s: 1892091571},
+	{p: 2137620481, g: 778943821, s: 48456461},
+	{p: 2137618433, g: 1730837875, s: 1713336725},
+	{p: 2137581569, g: 805610339, s: 1378891359},
+	{p: 2137538561, g: 204342388, s: 1950165220},
+	{p: 2137526273, g: 1947629754, s: 1500789441},
+	{p: 2137516033, g: 719902645, s: 1499525372},
+	{p: 2137491457, g: 230451261, s: 556382829},
+	{p: 2137440257, g: 979573541, s: 412760291},
+	{p: 2137374721, g: 927841248, s: 1954137185},
+	{p: 2137362433, g: 1243778559, s: 861024672},
+	{p: 2137313281, g: 1341338501, s: 980638386},
+	{p: 2137311233, g: 937415182, s: 1793212117},
+	{p: 2137255937, g: 795331324, s: 1410253405},
+	{p: 2137243649, g: 150756339, s: 1966999887},
+	{p: 2137182209, g: 163346914, s: 1939301431},
+	{p: 2137171969, g: 1952552395, s: 758913141},
+	{p: 2137159681, g: 570788721, s: 218668666},
+	{p: 2137147393, g: 1896656810, s: 2045670345},
+	{p: 2137141249, g: 358493842, s: 518199643},
+	{p: 2137139201, g: 1505023029, s: 674695848},
+	{p: 2137133057, g: 27911103, s: 830956306},
+	{p: 2137122817, g: 439771337, s: 1555268614},
+	{p: 2137116673, g: 790988579, s: 1871449599},
+	{p: 2137110529, g: 432109234, s: 811805080},
+	{p: 2137102337, g: 1357900653, s: 1184997641},
+	{p: 2137098241, g: 515119035, s: 1715693095},
+	{p: 2137090049, g: 408575203, s: 2085660657},
+	{p: 2137085953, g: 2097793407, s: 1349626963},
+	{p: 2137055233, g: 1556739954, s: 1449960883},
+	{p: 2137030657, g: 1545758650, s: 1369303716},
+	{p: 2136987649, g: 332602570, s: 103875114},
+	{p: 2136969217, g: 1499989506, s: 1662964115},
+	{p: 2136924161, g: 857040753, s: 4738842},
+	{p: 2136895489, g: 1948872712, s: 570436091},
+	{p: 2136893441, g: 58969960, s: 1568349634},
+	{p: 2136887297, g: 2127193379, s: 273612548},
+	{p: 2136850433, g: 111208983, s: 1181257116},
+	{p: 2136809473, g: 1627275942, s: 1680317971},
+	{p: 2136764417, g: 1574888217, s: 14011331},
+	{p: 2136741889, g: 14011055, s: 1129154251},
+	{p: 2136727553, g: 35862563, s: 1838555253},
+	{p: 2136721409, g: 310235666, s: 1363928244},
+	{p: 2136698881, g: 1612429202, s: 1560383828},
+	{p: 2136649729, g: 1138540131, s: 800014364},
+	{p: 2136606721, g: 602323503, s: 1433096652},
+	{p: 2136563713, g: 182209265, s: 1919611038},
+	{p: 2136555521, g: 324156477, s: 165591039},
+	{p: 2136549377, g: 195513113, s: 217165345},
+	{p: 2136526849, g: 1050768046, s: 939647887},
+	{p: 2136508417, g: 1886286237, s: 1619926572},
+	{p: 2136477697, g: 609647664, s: 35065157},
+	{p: 2136471553, g: 679352216, s: 1452259468},
+	{p: 2136457217, g: 128630031, s: 824816521},
+	{p: 2136422401, g: 19787464, s: 1526049830},
+	{p: 2136420353, g: 698316836, s: 1530623527},
+	{p: 2136371201, g: 1651862373, s: 1804812805},
+	{p: 2136334337, g: 326596005, s: 336977082},
+	{p: 2136322049, g: 63253370, s: 1904972151},
+	{p: 2136297473, g: 312176076, s: 172182411},
+	{p: 2136248321, g: 381261841, s: 369032670},
+	{p: 2136242177, g: 358688773, s: 1640007994},
+	{p: 2136229889, g: 512677188, s: 75585225},
+	{p: 2136219649, g: 2095003250, s: 1970086149},
+	{p: 2136207361, g: 1909650722, s: 537760675},
+	{p: 2136176641, g: 1334616195, s: 1533487619},
+	{p: 2136158209, g: 2096285632, s: 1793285210},
+	{p: 2136143873, g: 1897347517, s: 293843959},
+	{p: 2136133633, g: 923586222, s: 1022655978},
+	{p: 2136096769, g: 1464868191, s: 1515074410},
+	{p: 2136094721, g: 2020679520, s: 2061636104},
+	{p: 2136076289, g: 290798503, s: 1814726809},
+	{p: 2136041473, g: 156415894, s: 1250757633},
+	{p: 2135996417, g: 297459940, s: 1132158924},
+	{p: 2135955457, g: 538755304, s: 1688831340},
+	{p: 0, g: 0, s: 0},
+}
+
+// rev10 is the Falcon reference REV10 table for 10-bit bit reversal.
+// Source: https://falcon-sign.info/impl/keygen.c.html
+var rev10 = [...]uint16{
+	0, 512, 256, 768, 128, 640, 384, 896, 64, 576, 320, 832, 192, 704, 448, 960,
+	32, 544, 288, 800, 160, 672, 416, 928, 96, 608, 352, 864, 224, 736, 480, 992,
+	16, 528, 272, 784, 144, 656, 400, 912, 80, 592, 336, 848, 208, 720, 464, 976,
+	48, 560, 304, 816, 176, 688, 432, 944, 112, 624, 368, 880, 240, 752, 496, 1008,
+	8, 520, 264, 776, 136, 648, 392, 904, 72, 584, 328, 840, 200, 712, 456, 968,
+	40, 552, 296, 808, 168, 680, 424, 936, 104, 616, 360, 872, 232, 744, 488, 1000,
+	24, 536, 280, 792, 152, 664, 408, 920, 88, 600, 344, 856, 216, 728, 472, 984,
+	56, 568, 312, 824, 184, 696, 440, 952, 120, 632, 376, 888, 248, 760, 504, 1016,
+	4, 516, 260, 772, 132, 644, 388, 900, 68, 580, 324, 836, 196, 708, 452, 964,
+	36, 548, 292, 804, 164, 676, 420, 932, 100, 612, 356, 868, 228, 740, 484, 996,
+	20, 532, 276, 788, 148, 660, 404, 916, 84, 596, 340, 852, 212, 724, 468, 980,
+	52, 564, 308, 820, 180, 692, 436, 948, 116, 628, 372, 884, 244, 756, 500, 1012,
+	12, 524, 268, 780, 140, 652, 396, 908, 76, 588, 332, 844, 204, 716, 460, 972,
+	44, 556, 300, 812, 172, 684, 428, 940, 108, 620, 364, 876, 236, 748, 492, 1004,
+	28, 540, 284, 796, 156, 668, 412, 924, 92, 604, 348, 860, 220, 732, 476, 988,
+	60, 572, 316, 828, 188, 700, 444, 956, 124, 636, 380, 892, 252, 764, 508, 1020,
+	2, 514, 258, 770, 130, 642, 386, 898, 66, 578, 322, 834, 194, 706, 450, 962,
+	34, 546, 290, 802, 162, 674, 418, 930, 98, 610, 354, 866, 226, 738, 482, 994,
+	18, 530, 274, 786, 146, 658, 402, 914, 82, 594, 338, 850, 210, 722, 466, 978,
+	50, 562, 306, 818, 178, 690, 434, 946, 114, 626, 370, 882, 242, 754, 498, 1010,
+	10, 522, 266, 778, 138, 650, 394, 906, 74, 586, 330, 842, 202, 714, 458, 970,
+	42, 554, 298, 810, 170, 682, 426, 938, 106, 618, 362, 874, 234, 746, 490, 1002,
+	26, 538, 282, 794, 154, 666, 410, 922, 90, 602, 346, 858, 218, 730, 474, 986,
+	58, 570, 314, 826, 186, 698, 442, 954, 122, 634, 378, 890, 250, 762, 506, 1018,
+	6, 518, 262, 774, 134, 646, 390, 902, 70, 582, 326, 838, 198, 710, 454, 966,
+	38, 550, 294, 806, 166, 678, 422, 934, 102, 614, 358, 870, 230, 742, 486, 998,
+	22, 534, 278, 790, 150, 662, 406, 918, 86, 598, 342, 854, 214, 726, 470, 982,
+	54, 566, 310, 822, 182, 694, 438, 950, 118, 630, 374, 886, 246, 758, 502, 1014,
+	14, 526, 270, 782, 142, 654, 398, 910, 78, 590, 334, 846, 206, 718, 462, 974,
+	46, 558, 302, 814, 174, 686, 430, 942, 110, 622, 366, 878, 238, 750, 494, 1006,
+	30, 542, 286, 798, 158, 670, 414, 926, 94, 606, 350, 862, 222, 734, 478, 990,
+	62, 574, 318, 830, 190, 702, 446, 958, 126, 638, 382, 894, 254, 766, 510, 1022,
+	1, 513, 257, 769, 129, 641, 385, 897, 65, 577, 321, 833, 193, 705, 449, 961,
+	33, 545, 289, 801, 161, 673, 417, 929, 97, 609, 353, 865, 225, 737, 481, 993,
+	17, 529, 273, 785, 145, 657, 401, 913, 81, 593, 337, 849, 209, 721, 465, 977,
+	49, 561, 305, 817, 177, 689, 433, 945, 113, 625, 369, 881, 241, 753, 497, 1009,
+	9, 521, 265, 777, 137, 649, 393, 905, 73, 585, 329, 841, 201, 713, 457, 969,
+	41, 553, 297, 809, 169, 681, 425, 937, 105, 617, 361, 873, 233, 745, 489, 1001,
+	25, 537, 281, 793, 153, 665, 409, 921, 89, 601, 345, 857, 217, 729, 473, 985,
+	57, 569, 313, 825, 185, 697, 441, 953, 121, 633, 377, 889, 249, 761, 505, 1017,
+	5, 517, 261, 773, 133, 645, 389, 901, 69, 581, 325, 837, 197, 709, 453, 965,
+	37, 549, 293, 805, 165, 677, 421, 933, 101, 613, 357, 869, 229, 741, 485, 997,
+	21, 533, 277, 789, 149, 661, 405, 917, 85, 597, 341, 853, 213, 725, 469, 981,
+	53, 565, 309, 821, 181, 693, 437, 949, 117, 629, 373, 885, 245, 757, 501, 1013,
+	13, 525, 269, 781, 141, 653, 397, 909, 77, 589, 333, 845, 205, 717, 461, 973,
+	45, 557, 301, 813, 173, 685, 429, 941, 109, 621, 365, 877, 237, 749, 493, 1005,
+	29, 541, 285, 797, 157, 669, 413, 925, 93, 605, 349, 861, 221, 733, 477, 989,
+	61, 573, 317, 829, 189, 701, 445, 957, 125, 637, 381, 893, 253, 765, 509, 1021,
+	3, 515, 259, 771, 131, 643, 387, 899, 67, 579, 323, 835, 195, 707, 451, 963,
+	35, 547, 291, 803, 163, 675, 419, 931, 99, 611, 355, 867, 227, 739, 483, 995,
+	19, 531, 275, 787, 147, 659, 403, 915, 83, 595, 339, 851, 211, 723, 467, 979,
+	51, 563, 307, 819, 179, 691, 435, 947, 115, 627, 371, 883, 243, 755, 499, 1011,
+	11, 523, 267, 779, 139, 651, 395, 907, 75, 587, 331, 843, 203, 715, 459, 971,
+	43, 555, 299, 811, 171, 683, 427, 939, 107, 619, 363, 875, 235, 747, 491, 1003,
+	27, 539, 283, 795, 155, 667, 411, 923, 91, 603, 347, 859, 219, 731, 475, 987,
+	59, 571, 315, 827, 187, 699, 443, 955, 123, 635, 379, 891, 251, 763, 507, 1019,
+	7, 519, 263, 775, 135, 647, 391, 903, 71, 583, 327, 839, 199, 711, 455, 967,
+	39, 551, 295, 807, 167, 679, 423, 935, 103, 615, 359, 871, 231, 743, 487, 999,
+	23, 535, 279, 791, 151, 663, 407, 919, 87, 599, 343, 855, 215, 727, 471, 983,
+	55, 567, 311, 823, 183, 695, 439, 951, 119, 631, 375, 887, 247, 759, 503, 1015,
+	15, 527, 271, 783, 143, 655, 399, 911, 79, 591, 335, 847, 207, 719, 463, 975,
+	47, 559, 303, 815, 175, 687, 431, 943, 111, 623, 367, 879, 239, 751, 495, 1007,
+	31, 543, 287, 799, 159, 671, 415, 927, 95, 607, 351, 863, 223, 735, 479, 991,
+	63, 575, 319, 831, 191, 703, 447, 959, 127, 639, 383, 895, 255, 767, 511, 1023,
+}
+
+// falconGMTabBits contains Falcon reference fpr_gm_tab entries as float64 bit patterns.
+var falconGMTabBits = [2 * n]uint64{
+	0x0000000000000000, 0x0000000000000000, 0x8000000000000000, 0x3ff0000000000000,
+	0x3fe6a09e667f3bcd, 0x3fe6a09e667f3bcd, 0xbfe6a09e667f3bcd, 0x3fe6a09e667f3bcd,
+	0x3fed906bcf328d46, 0x3fd87de2a6aea963, 0xbfd87de2a6aea963, 0x3fed906bcf328d46,
+	0x3fd87de2a6aea963, 0x3fed906bcf328d46, 0xbfed906bcf328d46, 0x3fd87de2a6aea963,
+	0x3fef6297cff75cb0, 0x3fc8f8b83c69a60b, 0xbfc8f8b83c69a60b, 0x3fef6297cff75cb0,
+	0x3fe1c73b39ae68c8, 0x3fea9b66290ea1a3, 0xbfea9b66290ea1a3, 0x3fe1c73b39ae68c8,
+	0x3fea9b66290ea1a3, 0x3fe1c73b39ae68c8, 0xbfe1c73b39ae68c8, 0x3fea9b66290ea1a3,
+	0x3fc8f8b83c69a60b, 0x3fef6297cff75cb0, 0xbfef6297cff75cb0, 0x3fc8f8b83c69a60b,
+	0x3fefd88da3d12526, 0x3fb917a6bc29b42c, 0xbfb917a6bc29b42c, 0x3fefd88da3d12526,
+	0x3fe44cf325091dd6, 0x3fe8bc806b151741, 0xbfe8bc806b151741, 0x3fe44cf325091dd6,
+	0x3fec38b2f180bdb1, 0x3fde2b5d3806f63b, 0xbfde2b5d3806f63b, 0x3fec38b2f180bdb1,
+	0x3fd294062ed59f06, 0x3fee9f4156c62dda, 0xbfee9f4156c62dda, 0x3fd294062ed59f06,
+	0x3fee9f4156c62dda, 0x3fd294062ed59f06, 0xbfd294062ed59f06, 0x3fee9f4156c62dda,
+	0x3fde2b5d3806f63b, 0x3fec38b2f180bdb1, 0xbfec38b2f180bdb1, 0x3fde2b5d3806f63b,
+	0x3fe8bc806b151741, 0x3fe44cf325091dd6, 0xbfe44cf325091dd6, 0x3fe8bc806b151741,
+	0x3fb917a6bc29b42c, 0x3fefd88da3d12526, 0xbfefd88da3d12526, 0x3fb917a6bc29b42c,
+	0x3feff621e3796d7e, 0x3fa91f65f10dd814, 0xbfa91f65f10dd814, 0x3feff621e3796d7e,
+	0x3fe57d69348ceca0, 0x3fe7b5df226aafaf, 0xbfe7b5df226aafaf, 0x3fe57d69348ceca0,
+	0x3feced7af43cc773, 0x3fdb5d1009e15cc0, 0xbfdb5d1009e15cc0, 0x3feced7af43cc773,
+	0x3fd58f9a75ab1fdd, 0x3fee212104f686e5, 0xbfee212104f686e5, 0x3fd58f9a75ab1fdd,
+	0x3fef0a7efb9230d7, 0x3fcf19f97b215f1b, 0xbfcf19f97b215f1b, 0x3fef0a7efb9230d7,
+	0x3fe073879922ffee, 0x3feb728345196e3e, 0xbfeb728345196e3e, 0x3fe073879922ffee,
+	0x3fe9b3e047f38741, 0x3fe30ff7fce17035, 0xbfe30ff7fce17035, 0x3fe9b3e047f38741,
+	0x3fc2c8106e8e613a, 0x3fefa7557f08a517, 0xbfefa7557f08a517, 0x3fc2c8106e8e613a,
+	0x3fefa7557f08a517, 0x3fc2c8106e8e613a, 0xbfc2c8106e8e613a, 0x3fefa7557f08a517,
+	0x3fe30ff7fce17035, 0x3fe9b3e047f38741, 0xbfe9b3e047f38741, 0x3fe30ff7fce17035,
+	0x3feb728345196e3e, 0x3fe073879922ffee, 0xbfe073879922ffee, 0x3feb728345196e3e,
+	0x3fcf19f97b215f1b, 0x3fef0a7efb9230d7, 0xbfef0a7efb9230d7, 0x3fcf19f97b215f1b,
+	0x3fee212104f686e5, 0x3fd58f9a75ab1fdd, 0xbfd58f9a75ab1fdd, 0x3fee212104f686e5,
+	0x3fdb5d1009e15cc0, 0x3feced7af43cc773, 0xbfeced7af43cc773, 0x3fdb5d1009e15cc0,
+	0x3fe7b5df226aafaf, 0x3fe57d69348ceca0, 0xbfe57d69348ceca0, 0x3fe7b5df226aafaf,
+	0x3fa91f65f10dd814, 0x3feff621e3796d7e, 0xbfeff621e3796d7e, 0x3fa91f65f10dd814,
+	0x3feffd886084cd0d, 0x3f992155f7a3667e, 0xbf992155f7a3667e, 0x3feffd886084cd0d,
+	0x3fe610b7551d2cdf, 0x3fe72d0837efff96, 0xbfe72d0837efff96, 0x3fe610b7551d2cdf,
+	0x3fed4134d14dc93a, 0x3fd9ef7943a8ed8a, 0xbfd9ef7943a8ed8a, 0x3fed4134d14dc93a,
+	0x3fd7088530fa459f, 0x3feddb13b6ccc23c, 0xbfeddb13b6ccc23c, 0x3fd7088530fa459f,
+	0x3fef38f3ac64e589, 0x3fcc0b826a7e4f63, 0xbfcc0b826a7e4f63, 0x3fef38f3ac64e589,
+	0x3fe11eb3541b4b23, 0x3feb090a58150200, 0xbfeb090a58150200, 0x3fe11eb3541b4b23,
+	0x3fea29a7a0462782, 0x3fe26d054cdd12df, 0xbfe26d054cdd12df, 0x3fea29a7a0462782,
+	0x3fc5e214448b3fc6, 0x3fef8764fa714ba9, 0xbfef8764fa714ba9, 0x3fc5e214448b3fc6,
+	0x3fefc26470e19fd3, 0x3fbf564e56a9730e, 0xbfbf564e56a9730e, 0x3fefc26470e19fd3,
+	0x3fe3affa292050b9, 0x3fe93a22499263fb, 0xbfe93a22499263fb, 0x3fe3affa292050b9,
+	0x3febd7c0ac6f952a, 0x3fdf8ba4dbf89aba, 0xbfdf8ba4dbf89aba, 0x3febd7c0ac6f952a,
+	0x3fd111d262b1f677, 0x3feed740e7684963, 0xbfeed740e7684963, 0x3fd111d262b1f677,
+	0x3fee6288ec48e112, 0x3fd4135c94176601, 0xbfd4135c94176601, 0x3fee6288ec48e112,
+	0x3fdcc66e9931c45e, 0x3fec954b213411f5, 0xbfec954b213411f5, 0x3fdcc66e9931c45e,
+	0x3fe83b0e0bff976e, 0x3fe4e6cabbe3e5e9, 0xbfe4e6cabbe3e5e9, 0x3fe83b0e0bff976e,
+	0x3fb2d52092ce19f6, 0x3fefe9cdad01883a, 0xbfefe9cdad01883a, 0x3fb2d52092ce19f6,
+	0x3fefe9cdad01883a, 0x3fb2d52092ce19f6, 0xbfb2d52092ce19f6, 0x3fefe9cdad01883a,
+	0x3fe4e6cabbe3e5e9, 0x3fe83b0e0bff976e, 0xbfe83b0e0bff976e, 0x3fe4e6cabbe3e5e9,
+	0x3fec954b213411f5, 0x3fdcc66e9931c45e, 0xbfdcc66e9931c45e, 0x3fec954b213411f5,
+	0x3fd4135c94176601, 0x3fee6288ec48e112, 0xbfee6288ec48e112, 0x3fd4135c94176601,
+	0x3feed740e7684963, 0x3fd111d262b1f677, 0xbfd111d262b1f677, 0x3feed740e7684963,
+	0x3fdf8ba4dbf89aba, 0x3febd7c0ac6f952a, 0xbfebd7c0ac6f952a, 0x3fdf8ba4dbf89aba,
+	0x3fe93a22499263fb, 0x3fe3affa292050b9, 0xbfe3affa292050b9, 0x3fe93a22499263fb,
+	0x3fbf564e56a9730e, 0x3fefc26470e19fd3, 0xbfefc26470e19fd3, 0x3fbf564e56a9730e,
+	0x3fef8764fa714ba9, 0x3fc5e214448b3fc6, 0xbfc5e214448b3fc6, 0x3fef8764fa714ba9,
+	0x3fe26d054cdd12df, 0x3fea29a7a0462782, 0xbfea29a7a0462782, 0x3fe26d054cdd12df,
+	0x3feb090a58150200, 0x3fe11eb3541b4b23, 0xbfe11eb3541b4b23, 0x3feb090a58150200,
+	0x3fcc0b826a7e4f63, 0x3fef38f3ac64e589, 0xbfef38f3ac64e589, 0x3fcc0b826a7e4f63,
+	0x3feddb13b6ccc23c, 0x3fd7088530fa459f, 0xbfd7088530fa459f, 0x3feddb13b6ccc23c,
+	0x3fd9ef7943a8ed8a, 0x3fed4134d14dc93a, 0xbfed4134d14dc93a, 0x3fd9ef7943a8ed8a,
+	0x3fe72d0837efff96, 0x3fe610b7551d2cdf, 0xbfe610b7551d2cdf, 0x3fe72d0837efff96,
+	0x3f992155f7a3667e, 0x3feffd886084cd0d, 0xbfeffd886084cd0d, 0x3f992155f7a3667e,
+	0x3fefff62169b92db, 0x3f8921d1fcdec784, 0xbf8921d1fcdec784, 0x3fefff62169b92db,
+	0x3fe6591925f0783d, 0x3fe6e74454eaa8af, 0xbfe6e74454eaa8af, 0x3fe6591925f0783d,
+	0x3fed696173c9e68b, 0x3fd9372a63bc93d7, 0xbfd9372a63bc93d7, 0x3fed696173c9e68b,
+	0x3fd7c3a9311dcce7, 0x3fedb6526238a09b, 0xbfedb6526238a09b, 0x3fd7c3a9311dcce7,
+	0x3fef4e603b0b2f2d, 0x3fca82a025b00451, 0xbfca82a025b00451, 0x3fef4e603b0b2f2d,
+	0x3fe1734d63dedb49, 0x3fead2bc9e21d511, 0xbfead2bc9e21d511, 0x3fe1734d63dedb49,
+	0x3fea63091b02fae2, 0x3fe21a799933eb59, 0xbfe21a799933eb59, 0x3fea63091b02fae2,
+	0x3fc76dd9de50bf31, 0x3fef7599a3a12077, 0xbfef7599a3a12077, 0x3fc76dd9de50bf31,
+	0x3fefce15fd6da67b, 0x3fbc3785c79ec2d5, 0xbfbc3785c79ec2d5, 0x3fefce15fd6da67b,
+	0x3fe3fed9534556d4, 0x3fe8fbcca3ef940d, 0xbfe8fbcca3ef940d, 0x3fe3fed9534556d4,
+	0x3fec08c426725549, 0x3fdedc1952ef78d6, 0xbfdedc1952ef78d6, 0x3fec08c426725549,
+	0x3fd1d3443f4cdb3e, 0x3feebbd8c8df0b74, 0xbfeebbd8c8df0b74, 0x3fd1d3443f4cdb3e,
+	0x3fee817bab4cd10d, 0x3fd35410c2e18152, 0xbfd35410c2e18152, 0x3fee817bab4cd10d,
+	0x3fdd79775b86e389, 0x3fec678b3488739b, 0xbfec678b3488739b, 0x3fdd79775b86e389,
+	0x3fe87c400fba2ebf, 0x3fe49a449b9b0939, 0xbfe49a449b9b0939, 0x3fe87c400fba2ebf,
+	0x3fb5f6d00a9aa419, 0x3fefe1cafcbd5b09, 0xbfefe1cafcbd5b09, 0x3fb5f6d00a9aa419,
+	0x3feff095658e71ad, 0x3faf656e79f820e0, 0xbfaf656e79f820e0, 0x3feff095658e71ad,
+	0x3fe5328292a35596, 0x3fe7f8ece3571771, 0xbfe7f8ece3571771, 0x3fe5328292a35596,
+	0x3fecc1f0f3fcfc5c, 0x3fdc1249d8011ee7, 0xbfdc1249d8011ee7, 0x3fecc1f0f3fcfc5c,
+	0x3fd4d1e24278e76a, 0x3fee426a4b2bc17e, 0xbfee426a4b2bc17e, 0x3fd4d1e24278e76a,
+	0x3feef178a3e473c2, 0x3fd04fb80e37fdae, 0xbfd04fb80e37fdae, 0x3feef178a3e473c2,
+	0x3fe01cfc874c3eb7, 0x3feba5aa673590d2, 0xbfeba5aa673590d2, 0x3fe01cfc874c3eb7,
+	0x3fe9777ef4c7d742, 0x3fe36058b10659f3, 0xbfe36058b10659f3, 0x3fe9777ef4c7d742,
+	0x3fc139f0cedaf577, 0x3fefb5797195d741, 0xbfefb5797195d741, 0x3fc139f0cedaf577,
+	0x3fef97f924c9099b, 0x3fc45576b1293e5a, 0xbfc45576b1293e5a, 0x3fef97f924c9099b,
+	0x3fe2bedb25faf3ea, 0x3fe9ef43ef29af94, 0xbfe9ef43ef29af94, 0x3fe2bedb25faf3ea,
+	0x3feb3e4d3ef55712, 0x3fe0c9704d5d898f, 0xbfe0c9704d5d898f, 0x3feb3e4d3ef55712,
+	0x3fcd934fe5454311, 0x3fef2252f7763ada, 0xbfef2252f7763ada, 0x3fcd934fe5454311,
+	0x3fedfeae622dbe2b, 0x3fd64c7ddd3f27c6, 0xbfd64c7ddd3f27c6, 0x3fedfeae622dbe2b,
+	0x3fdaa6c82b6d3fca, 0x3fed17e7743e35dc, 0xbfed17e7743e35dc, 0x3fdaa6c82b6d3fca,
+	0x3fe771e75f037261, 0x3fe5c77bbe65018c, 0xbfe5c77bbe65018c, 0x3fe771e75f037261,
+	0x3fa2d865759455cd, 0x3feffa72effef75d, 0xbfeffa72effef75d, 0x3fa2d865759455cd,
+	0x3feffa72effef75d, 0x3fa2d865759455cd, 0xbfa2d865759455cd, 0x3feffa72effef75d,
+	0x3fe5c77bbe65018c, 0x3fe771e75f037261, 0xbfe771e75f037261, 0x3fe5c77bbe65018c,
+	0x3fed17e7743e35dc, 0x3fdaa6c82b6d3fca, 0xbfdaa6c82b6d3fca, 0x3fed17e7743e35dc,
+	0x3fd64c7ddd3f27c6, 0x3fedfeae622dbe2b, 0xbfedfeae622dbe2b, 0x3fd64c7ddd3f27c6,
+	0x3fef2252f7763ada, 0x3fcd934fe5454311, 0xbfcd934fe5454311, 0x3fef2252f7763ada,
+	0x3fe0c9704d5d898f, 0x3feb3e4d3ef55712, 0xbfeb3e4d3ef55712, 0x3fe0c9704d5d898f,
+	0x3fe9ef43ef29af94, 0x3fe2bedb25faf3ea, 0xbfe2bedb25faf3ea, 0x3fe9ef43ef29af94,
+	0x3fc45576b1293e5a, 0x3fef97f924c9099b, 0xbfef97f924c9099b, 0x3fc45576b1293e5a,
+	0x3fefb5797195d741, 0x3fc139f0cedaf577, 0xbfc139f0cedaf577, 0x3fefb5797195d741,
+	0x3fe36058b10659f3, 0x3fe9777ef4c7d742, 0xbfe9777ef4c7d742, 0x3fe36058b10659f3,
+	0x3feba5aa673590d2, 0x3fe01cfc874c3eb7, 0xbfe01cfc874c3eb7, 0x3feba5aa673590d2,
+	0x3fd04fb80e37fdae, 0x3feef178a3e473c2, 0xbfeef178a3e473c2, 0x3fd04fb80e37fdae,
+	0x3fee426a4b2bc17e, 0x3fd4d1e24278e76a, 0xbfd4d1e24278e76a, 0x3fee426a4b2bc17e,
+	0x3fdc1249d8011ee7, 0x3fecc1f0f3fcfc5c, 0xbfecc1f0f3fcfc5c, 0x3fdc1249d8011ee7,
+	0x3fe7f8ece3571771, 0x3fe5328292a35596, 0xbfe5328292a35596, 0x3fe7f8ece3571771,
+	0x3faf656e79f820e0, 0x3feff095658e71ad, 0xbfeff095658e71ad, 0x3faf656e79f820e0,
+	0x3fefe1cafcbd5b09, 0x3fb5f6d00a9aa419, 0xbfb5f6d00a9aa419, 0x3fefe1cafcbd5b09,
+	0x3fe49a449b9b0939, 0x3fe87c400fba2ebf, 0xbfe87c400fba2ebf, 0x3fe49a449b9b0939,
+	0x3fec678b3488739b, 0x3fdd79775b86e389, 0xbfdd79775b86e389, 0x3fec678b3488739b,
+	0x3fd35410c2e18152, 0x3fee817bab4cd10d, 0xbfee817bab4cd10d, 0x3fd35410c2e18152,
+	0x3feebbd8c8df0b74, 0x3fd1d3443f4cdb3e, 0xbfd1d3443f4cdb3e, 0x3feebbd8c8df0b74,
+	0x3fdedc1952ef78d6, 0x3fec08c426725549, 0xbfec08c426725549, 0x3fdedc1952ef78d6,
+	0x3fe8fbcca3ef940d, 0x3fe3fed9534556d4, 0xbfe3fed9534556d4, 0x3fe8fbcca3ef940d,
+	0x3fbc3785c79ec2d5, 0x3fefce15fd6da67b, 0xbfefce15fd6da67b, 0x3fbc3785c79ec2d5,
+	0x3fef7599a3a12077, 0x3fc76dd9de50bf31, 0xbfc76dd9de50bf31, 0x3fef7599a3a12077,
+	0x3fe21a799933eb59, 0x3fea63091b02fae2, 0xbfea63091b02fae2, 0x3fe21a799933eb59,
+	0x3fead2bc9e21d511, 0x3fe1734d63dedb49, 0xbfe1734d63dedb49, 0x3fead2bc9e21d511,
+	0x3fca82a025b00451, 0x3fef4e603b0b2f2d, 0xbfef4e603b0b2f2d, 0x3fca82a025b00451,
+	0x3fedb6526238a09b, 0x3fd7c3a9311dcce7, 0xbfd7c3a9311dcce7, 0x3fedb6526238a09b,
+	0x3fd9372a63bc93d7, 0x3fed696173c9e68b, 0xbfed696173c9e68b, 0x3fd9372a63bc93d7,
+	0x3fe6e74454eaa8af, 0x3fe6591925f0783d, 0xbfe6591925f0783d, 0x3fe6e74454eaa8af,
+	0x3f8921d1fcdec784, 0x3fefff62169b92db, 0xbfefff62169b92db, 0x3f8921d1fcdec784,
+	0x3fefffd8858e8a92, 0x3f7921f0fe670071, 0xbf7921f0fe670071, 0x3fefffd8858e8a92,
+	0x3fe67cf78491af10, 0x3fe6c40d73c18275, 0xbfe6c40d73c18275, 0x3fe67cf78491af10,
+	0x3fed7d0b02b8ecf9, 0x3fd8daa52ec8a4b0, 0xbfd8daa52ec8a4b0, 0x3fed7d0b02b8ecf9,
+	0x3fd820e3b04eaac4, 0x3feda383a9668988, 0xbfeda383a9668988, 0x3fd820e3b04eaac4,
+	0x3fef58a2b1789e84, 0x3fc9bdcbf2dc4366, 0xbfc9bdcbf2dc4366, 0x3fef58a2b1789e84,
+	0x3fe19d5a09f2b9b8, 0x3feab7325916c0d4, 0xbfeab7325916c0d4, 0x3fe19d5a09f2b9b8,
+	0x3fea7f58529fe69d, 0x3fe1f0f08bbc861b, 0xbfe1f0f08bbc861b, 0x3fea7f58529fe69d,
+	0x3fc83366e89c64c6, 0x3fef6c3f7df5bbb7, 0xbfef6c3f7df5bbb7, 0x3fc83366e89c64c6,
+	0x3fefd37914220b84, 0x3fbaa7b724495c03, 0xbfbaa7b724495c03, 0x3fefd37914220b84,
+	0x3fe425ff178e6bb1, 0x3fe8dc45331698cc, 0xbfe8dc45331698cc, 0x3fe425ff178e6bb1,
+	0x3fec20de3fa971b0, 0x3fde83e0eaf85114, 0xbfde83e0eaf85114, 0x3fec20de3fa971b0,
+	0x3fd233bbabc3bb71, 0x3feeadb2e8e7a88e, 0xbfeeadb2e8e7a88e, 0x3fd233bbabc3bb71,
+	0x3fee9084361df7f2, 0x3fd2f422daec0387, 0xbfd2f422daec0387, 0x3fee9084361df7f2,
+	0x3fddd28f1481cc58, 0x3fec5042012b6907, 0xbfec5042012b6907, 0x3fddd28f1481cc58,
+	0x3fe89c7e9a4dd4aa, 0x3fe473b51b987347, 0xbfe473b51b987347, 0x3fe89c7e9a4dd4aa,
+	0x3fb787586a5d5b21, 0x3fefdd539ff1f456, 0xbfefdd539ff1f456, 0x3fb787586a5d5b21,
+	0x3feff3830f8d575c, 0x3fac428d12c0d7e3, 0xbfac428d12c0d7e3, 0x3feff3830f8d575c,
+	0x3fe5581038975137, 0x3fe7d7836cc33db2, 0xbfe7d7836cc33db2, 0x3fe5581038975137,
+	0x3fecd7d9898b32f6, 0x3fdbb7cf2304bd01, 0xbfdbb7cf2304bd01, 0x3fecd7d9898b32f6,
+	0x3fd530d880af3c24, 0x3fee31eae870ce25, 0xbfee31eae870ce25, 0x3fd530d880af3c24,
+	0x3feefe220c0b95ec, 0x3fcfdcdc1adfedf9, 0xbfcfdcdc1adfedf9, 0x3feefe220c0b95ec,
+	0x3fe0485626ae221a, 0x3feb8c38d27504e9, 0xbfeb8c38d27504e9, 0x3fe0485626ae221a,
+	0x3fe995cf2ed80d22, 0x3fe338400d0c8e57, 0xbfe338400d0c8e57, 0x3fe995cf2ed80d22,
+	0x3fc20116d4ec7bcf, 0x3fefae8e8e46cfbb, 0xbfefae8e8e46cfbb, 0x3fc20116d4ec7bcf,
+	0x3fef9fce55adb2c8, 0x3fc38edbb0cd8d14, 0xbfc38edbb0cd8d14, 0x3fef9fce55adb2c8,
+	0x3fe2e780e3e8ea17, 0x3fe9d1b1f5ea80d5, 0xbfe9d1b1f5ea80d5, 0x3fe2e780e3e8ea17,
+	0x3feb5889fe921405, 0x3fe09e907417c5e1, 0xbfe09e907417c5e1, 0x3feb5889fe921405,
+	0x3fce56ca1e101a1b, 0x3fef168f53f7205d, 0xbfef168f53f7205d, 0x3fce56ca1e101a1b,
+	0x3fee100cca2980ac, 0x3fd5ee27379ea693, 0xbfd5ee27379ea693, 0x3fee100cca2980ac,
+	0x3fdb020d6c7f4009, 0x3fed02d4feb2bd92, 0xbfed02d4feb2bd92, 0x3fdb020d6c7f4009,
+	0x3fe79400574f55e5, 0x3fe5a28d2a5d7250, 0xbfe5a28d2a5d7250, 0x3fe79400574f55e5,
+	0x3fa5fc00d290cd43, 0x3feff871dadb81df, 0xbfeff871dadb81df, 0x3fa5fc00d290cd43,
+	0x3feffc251df1d3f8, 0x3f9f693731d1cf01, 0xbf9f693731d1cf01, 0x3feffc251df1d3f8,
+	0x3fe5ec3495837074, 0x3fe74f948da8d28d, 0xbfe74f948da8d28d, 0x3fe5ec3495837074,
+	0x3fed2cb220e0ef9f, 0x3fda4b4127dea1e5, 0xbfda4b4127dea1e5, 0x3fed2cb220e0ef9f,
+	0x3fd6aa9d7dc77e17, 0x3feded05f7de47da, 0xbfeded05f7de47da, 0x3fd6aa9d7dc77e17,
+	0x3fef2dc9c9089a9d, 0x3fcccf8cb312b286, 0xbfcccf8cb312b286, 0x3fef2dc9c9089a9d,
+	0x3fe0f426bb2a8e7e, 0x3feb23cd470013b4, 0xbfeb23cd470013b4, 0x3fe0f426bb2a8e7e,
+	0x3fea0c95eabaf937, 0x3fe2960727629ca8, 0xbfe2960727629ca8, 0x3fea0c95eabaf937,
+	0x3fc51bdf8597c5f2, 0x3fef8fd5ffae41db, 0xbfef8fd5ffae41db, 0x3fc51bdf8597c5f2,
+	0x3fefbc1617e44186, 0x3fc072a047ba831d, 0xbfc072a047ba831d, 0x3fefbc1617e44186,
+	0x3fe3884185dfeb22, 0x3fe958efe48e6dd7, 0xbfe958efe48e6dd7, 0x3fe3884185dfeb22,
+	0x3febbed7c49380ea, 0x3fdfe2f64be71210, 0xbfdfe2f64be71210, 0x3febbed7c49380ea,
+	0x3fd0b0d9cfdbdb90, 0x3feee482e25a9dbc, 0xbfeee482e25a9dbc, 0x3fd0b0d9cfdbdb90,
+	0x3fee529f04729ffc, 0x3fd472b8a5571054, 0xbfd472b8a5571054, 0x3fee529f04729ffc,
+	0x3fdc6c7f4997000b, 0x3fecabc169a0b900, 0xbfecabc169a0b900, 0x3fdc6c7f4997000b,
+	0x3fe81a1b33b57acc, 0x3fe50cc09f59a09b, 0xbfe50cc09f59a09b, 0x3fe81a1b33b57acc,
+	0x3fb1440134d709b3, 0x3fefed58ecb673c4, 0xbfefed58ecb673c4, 0x3fb1440134d709b3,
+	0x3fefe5f3af2e3940, 0x3fb4661179272096, 0xbfb4661179272096, 0x3fefe5f3af2e3940,
+	0x3fe4c0a145ec0004, 0x3fe85bc51ae958cc, 0xbfe85bc51ae958cc, 0x3fe4c0a145ec0004,
+	0x3fec7e8e52233cf3, 0x3fdd2016e8e9db5b, 0xbfdd2016e8e9db5b, 0x3fec7e8e52233cf3,
+	0x3fd3b3cefa0414b7, 0x3fee7227db6a9744, 0xbfee7227db6a9744, 0x3fd3b3cefa0414b7,
+	0x3feec9b2d3c3bf84, 0x3fd172a0d7765177, 0xbfd172a0d7765177, 0x3feec9b2d3c3bf84,
+	0x3fdf3405963fd067, 0x3febf064e15377dd, 0xbfebf064e15377dd, 0x3fdf3405963fd067,
+	0x3fe91b166fd49da2, 0x3fe3d78238c58344, 0xbfe3d78238c58344, 0x3fe91b166fd49da2,
+	0x3fbdc70ecbae9fc9, 0x3fefc8646cfeb721, 0xbfefc8646cfeb721, 0x3fbdc70ecbae9fc9,
+	0x3fef7ea629e63d6e, 0x3fc6a81304f64ab2, 0xbfc6a81304f64ab2, 0x3fef7ea629e63d6e,
+	0x3fe243d5fb98ac1f, 0x3fea4678c8119ac8, 0xbfea4678c8119ac8, 0x3fe243d5fb98ac1f,
+	0x3feaee04b43c1474, 0x3fe14915af336ceb, 0xbfe14915af336ceb, 0x3feaee04b43c1474,
+	0x3fcb4732ef3d6722, 0x3fef43d085ff92dd, 0xbfef43d085ff92dd, 0x3fcb4732ef3d6722,
+	0x3fedc8d7cb410260, 0x3fd766340f2418f6, 0xbfd766340f2418f6, 0x3fedc8d7cb410260,
+	0x3fd993716141bdff, 0x3fed556f52e93eb1, 0xbfed556f52e93eb1, 0x3fd993716141bdff,
+	0x3fe70a42b3176d7a, 0x3fe63503a31c1be9, 0xbfe63503a31c1be9, 0x3fe70a42b3176d7a,
+	0x3f92d936bbe30efd, 0x3feffe9cb44b51a1, 0xbfeffe9cb44b51a1, 0x3f92d936bbe30efd,
+	0x3feffe9cb44b51a1, 0x3f92d936bbe30efd, 0xbf92d936bbe30efd, 0x3feffe9cb44b51a1,
+	0x3fe63503a31c1be9, 0x3fe70a42b3176d7a, 0xbfe70a42b3176d7a, 0x3fe63503a31c1be9,
+	0x3fed556f52e93eb1, 0x3fd993716141bdff, 0xbfd993716141bdff, 0x3fed556f52e93eb1,
+	0x3fd766340f2418f6, 0x3fedc8d7cb410260, 0xbfedc8d7cb410260, 0x3fd766340f2418f6,
+	0x3fef43d085ff92dd, 0x3fcb4732ef3d6722, 0xbfcb4732ef3d6722, 0x3fef43d085ff92dd,
+	0x3fe14915af336ceb, 0x3feaee04b43c1474, 0xbfeaee04b43c1474, 0x3fe14915af336ceb,
+	0x3fea4678c8119ac8, 0x3fe243d5fb98ac1f, 0xbfe243d5fb98ac1f, 0x3fea4678c8119ac8,
+	0x3fc6a81304f64ab2, 0x3fef7ea629e63d6e, 0xbfef7ea629e63d6e, 0x3fc6a81304f64ab2,
+	0x3fefc8646cfeb721, 0x3fbdc70ecbae9fc9, 0xbfbdc70ecbae9fc9, 0x3fefc8646cfeb721,
+	0x3fe3d78238c58344, 0x3fe91b166fd49da2, 0xbfe91b166fd49da2, 0x3fe3d78238c58344,
+	0x3febf064e15377dd, 0x3fdf3405963fd067, 0xbfdf3405963fd067, 0x3febf064e15377dd,
+	0x3fd172a0d7765177, 0x3feec9b2d3c3bf84, 0xbfeec9b2d3c3bf84, 0x3fd172a0d7765177,
+	0x3fee7227db6a9744, 0x3fd3b3cefa0414b7, 0xbfd3b3cefa0414b7, 0x3fee7227db6a9744,
+	0x3fdd2016e8e9db5b, 0x3fec7e8e52233cf3, 0xbfec7e8e52233cf3, 0x3fdd2016e8e9db5b,
+	0x3fe85bc51ae958cc, 0x3fe4c0a145ec0004, 0xbfe4c0a145ec0004, 0x3fe85bc51ae958cc,
+	0x3fb4661179272096, 0x3fefe5f3af2e3940, 0xbfefe5f3af2e3940, 0x3fb4661179272096,
+	0x3fefed58ecb673c4, 0x3fb1440134d709b3, 0xbfb1440134d709b3, 0x3fefed58ecb673c4,
+	0x3fe50cc09f59a09b, 0x3fe81a1b33b57acc, 0xbfe81a1b33b57acc, 0x3fe50cc09f59a09b,
+	0x3fecabc169a0b900, 0x3fdc6c7f4997000b, 0xbfdc6c7f4997000b, 0x3fecabc169a0b900,
+	0x3fd472b8a5571054, 0x3fee529f04729ffc, 0xbfee529f04729ffc, 0x3fd472b8a5571054,
+	0x3feee482e25a9dbc, 0x3fd0b0d9cfdbdb90, 0xbfd0b0d9cfdbdb90, 0x3feee482e25a9dbc,
+	0x3fdfe2f64be71210, 0x3febbed7c49380ea, 0xbfebbed7c49380ea, 0x3fdfe2f64be71210,
+	0x3fe958efe48e6dd7, 0x3fe3884185dfeb22, 0xbfe3884185dfeb22, 0x3fe958efe48e6dd7,
+	0x3fc072a047ba831d, 0x3fefbc1617e44186, 0xbfefbc1617e44186, 0x3fc072a047ba831d,
+	0x3fef8fd5ffae41db, 0x3fc51bdf8597c5f2, 0xbfc51bdf8597c5f2, 0x3fef8fd5ffae41db,
+	0x3fe2960727629ca8, 0x3fea0c95eabaf937, 0xbfea0c95eabaf937, 0x3fe2960727629ca8,
+	0x3feb23cd470013b4, 0x3fe0f426bb2a8e7e, 0xbfe0f426bb2a8e7e, 0x3feb23cd470013b4,
+	0x3fcccf8cb312b286, 0x3fef2dc9c9089a9d, 0xbfef2dc9c9089a9d, 0x3fcccf8cb312b286,
+	0x3feded05f7de47da, 0x3fd6aa9d7dc77e17, 0xbfd6aa9d7dc77e17, 0x3feded05f7de47da,
+	0x3fda4b4127dea1e5, 0x3fed2cb220e0ef9f, 0xbfed2cb220e0ef9f, 0x3fda4b4127dea1e5,
+	0x3fe74f948da8d28d, 0x3fe5ec3495837074, 0xbfe5ec3495837074, 0x3fe74f948da8d28d,
+	0x3f9f693731d1cf01, 0x3feffc251df1d3f8, 0xbfeffc251df1d3f8, 0x3f9f693731d1cf01,
+	0x3feff871dadb81df, 0x3fa5fc00d290cd43, 0xbfa5fc00d290cd43, 0x3feff871dadb81df,
+	0x3fe5a28d2a5d7250, 0x3fe79400574f55e5, 0xbfe79400574f55e5, 0x3fe5a28d2a5d7250,
+	0x3fed02d4feb2bd92, 0x3fdb020d6c7f4009, 0xbfdb020d6c7f4009, 0x3fed02d4feb2bd92,
+	0x3fd5ee27379ea693, 0x3fee100cca2980ac, 0xbfee100cca2980ac, 0x3fd5ee27379ea693,
+	0x3fef168f53f7205d, 0x3fce56ca1e101a1b, 0xbfce56ca1e101a1b, 0x3fef168f53f7205d,
+	0x3fe09e907417c5e1, 0x3feb5889fe921405, 0xbfeb5889fe921405, 0x3fe09e907417c5e1,
+	0x3fe9d1b1f5ea80d5, 0x3fe2e780e3e8ea17, 0xbfe2e780e3e8ea17, 0x3fe9d1b1f5ea80d5,
+	0x3fc38edbb0cd8d14, 0x3fef9fce55adb2c8, 0xbfef9fce55adb2c8, 0x3fc38edbb0cd8d14,
+	0x3fefae8e8e46cfbb, 0x3fc20116d4ec7bcf, 0xbfc20116d4ec7bcf, 0x3fefae8e8e46cfbb,
+	0x3fe338400d0c8e57, 0x3fe995cf2ed80d22, 0xbfe995cf2ed80d22, 0x3fe338400d0c8e57,
+	0x3feb8c38d27504e9, 0x3fe0485626ae221a, 0xbfe0485626ae221a, 0x3feb8c38d27504e9,
+	0x3fcfdcdc1adfedf9, 0x3feefe220c0b95ec, 0xbfeefe220c0b95ec, 0x3fcfdcdc1adfedf9,
+	0x3fee31eae870ce25, 0x3fd530d880af3c24, 0xbfd530d880af3c24, 0x3fee31eae870ce25,
+	0x3fdbb7cf2304bd01, 0x3fecd7d9898b32f6, 0xbfecd7d9898b32f6, 0x3fdbb7cf2304bd01,
+	0x3fe7d7836cc33db2, 0x3fe5581038975137, 0xbfe5581038975137, 0x3fe7d7836cc33db2,
+	0x3fac428d12c0d7e3, 0x3feff3830f8d575c, 0xbfeff3830f8d575c, 0x3fac428d12c0d7e3,
+	0x3fefdd539ff1f456, 0x3fb787586a5d5b21, 0xbfb787586a5d5b21, 0x3fefdd539ff1f456,
+	0x3fe473b51b987347, 0x3fe89c7e9a4dd4aa, 0xbfe89c7e9a4dd4aa, 0x3fe473b51b987347,
+	0x3fec5042012b6907, 0x3fddd28f1481cc58, 0xbfddd28f1481cc58, 0x3fec5042012b6907,
+	0x3fd2f422daec0387, 0x3fee9084361df7f2, 0xbfee9084361df7f2, 0x3fd2f422daec0387,
+	0x3feeadb2e8e7a88e, 0x3fd233bbabc3bb71, 0xbfd233bbabc3bb71, 0x3feeadb2e8e7a88e,
+	0x3fde83e0eaf85114, 0x3fec20de3fa971b0, 0xbfec20de3fa971b0, 0x3fde83e0eaf85114,
+	0x3fe8dc45331698cc, 0x3fe425ff178e6bb1, 0xbfe425ff178e6bb1, 0x3fe8dc45331698cc,
+	0x3fbaa7b724495c03, 0x3fefd37914220b84, 0xbfefd37914220b84, 0x3fbaa7b724495c03,
+	0x3fef6c3f7df5bbb7, 0x3fc83366e89c64c6, 0xbfc83366e89c64c6, 0x3fef6c3f7df5bbb7,
+	0x3fe1f0f08bbc861b, 0x3fea7f58529fe69d, 0xbfea7f58529fe69d, 0x3fe1f0f08bbc861b,
+	0x3feab7325916c0d4, 0x3fe19d5a09f2b9b8, 0xbfe19d5a09f2b9b8, 0x3feab7325916c0d4,
+	0x3fc9bdcbf2dc4366, 0x3fef58a2b1789e84, 0xbfef58a2b1789e84, 0x3fc9bdcbf2dc4366,
+	0x3feda383a9668988, 0x3fd820e3b04eaac4, 0xbfd820e3b04eaac4, 0x3feda383a9668988,
+	0x3fd8daa52ec8a4b0, 0x3fed7d0b02b8ecf9, 0xbfed7d0b02b8ecf9, 0x3fd8daa52ec8a4b0,
+	0x3fe6c40d73c18275, 0x3fe67cf78491af10, 0xbfe67cf78491af10, 0x3fe6c40d73c18275,
+	0x3f7921f0fe670071, 0x3fefffd8858e8a92, 0xbfefffd8858e8a92, 0x3f7921f0fe670071,
+	0x3feffff621621d02, 0x3f6921f8becca4ba, 0xbf6921f8becca4ba, 0x3feffff621621d02,
+	0x3fe68ed1eaa19c71, 0x3fe6b25ced2fe29c, 0xbfe6b25ced2fe29c, 0x3fe68ed1eaa19c71,
+	0x3fed86c48445a44f, 0x3fd8ac4b86d5ed44, 0xbfd8ac4b86d5ed44, 0x3fed86c48445a44f,
+	0x3fd84f6aaaf3903f, 0x3fed9a00dd8b3d46, 0xbfed9a00dd8b3d46, 0x3fd84f6aaaf3903f,
+	0x3fef5da6ed43685d, 0x3fc95b49e9b62afa, 0xbfc95b49e9b62afa, 0x3fef5da6ed43685d,
+	0x3fe1b250171373bf, 0x3feaa9547a2cb98e, 0xbfeaa9547a2cb98e, 0x3fe1b250171373bf,
+	0x3fea8d676e545ad2, 0x3fe1dc1b64dc4872, 0xbfe1dc1b64dc4872, 0x3fea8d676e545ad2,
+	0x3fc8961727c41804, 0x3fef677556883cee, 0xbfef677556883cee, 0x3fc8961727c41804,
+	0x3fefd60d2da75c9e, 0x3fb9dfb6eb24a85c, 0xbfb9dfb6eb24a85c, 0x3fefd60d2da75c9e,
+	0x3fe4397f5b2a4380, 0x3fe8cc6a75184655, 0xbfe8cc6a75184655, 0x3fe4397f5b2a4380,
+	0x3fec2cd14931e3f1, 0x3fde57a86d3cd825, 0xbfde57a86d3cd825, 0x3fec2cd14931e3f1,
+	0x3fd263e6995554ba, 0x3feea68393e65800, 0xbfeea68393e65800, 0x3fd263e6995554ba,
+	0x3fee97ec36016b30, 0x3fd2c41a4e954520, 0xbfd2c41a4e954520, 0x3fee97ec36016b30,
+	0x3fddfeff66a941de, 0x3fec44833141c004, 0xbfec44833141c004, 0x3fddfeff66a941de,
+	0x3fe8ac871ede1d88, 0x3fe4605a692b32a2, 0xbfe4605a692b32a2, 0x3fe8ac871ede1d88,
+	0x3fb84f8712c130a1, 0x3fefdafa7514538c, 0xbfefdafa7514538c, 0x3fb84f8712c130a1,
+	0x3feff4dc54b1bed3, 0x3faab101bd5f8317, 0xbfaab101bd5f8317, 0x3feff4dc54b1bed3,
+	0x3fe56ac35197649f, 0x3fe7c6b89ce2d333, 0xbfe7c6b89ce2d333, 0x3fe56ac35197649f,
+	0x3fece2b32799a060, 0x3fdb8a7814fd5693, 0xbfdb8a7814fd5693, 0x3fece2b32799a060,
+	0x3fd5604012f467b4, 0x3fee298f4439197a, 0xbfee298f4439197a, 0x3fd5604012f467b4,
+	0x3fef045a14cf738c, 0x3fcf7b7480bd3802, 0xbfcf7b7480bd3802, 0x3fef045a14cf738c,
+	0x3fe05df3ec31b8b7, 0x3feb7f6686e792e9, 0xbfeb7f6686e792e9, 0x3fe05df3ec31b8b7,
+	0x3fe9a4dfa42b06b2, 0x3fe32421ec49a61f, 0xbfe32421ec49a61f, 0x3fe9a4dfa42b06b2,
+	0x3fc264994dfd3409, 0x3fefaafbcb0cfddc, 0xbfefaafbcb0cfddc, 0x3fc264994dfd3409,
+	0x3fefa39bac7a1791, 0x3fc32b7bf94516a7, 0xbfc32b7bf94516a7, 0x3fefa39bac7a1791,
+	0x3fe2fbc24b441015, 0x3fe9c2d110f075c2, 0xbfe9c2d110f075c2, 0x3fe2fbc24b441015,
+	0x3feb658f14fdbc47, 0x3fe089112032b08c, 0xbfe089112032b08c, 0x3feb658f14fdbc47,
+	0x3fceb86b462de348, 0x3fef1090bc898f5f, 0xbfef1090bc898f5f, 0x3fceb86b462de348,
+	0x3fee18a02fdc66d9, 0x3fd5bee78b9db3b6, 0xbfd5bee78b9db3b6, 0x3fee18a02fdc66d9,
+	0x3fdb2f971db31972, 0x3fecf830e8ce467b, 0xbfecf830e8ce467b, 0x3fdb2f971db31972,
+	0x3fe7a4f707bf97d2, 0x3fe59001d5f723df, 0xbfe59001d5f723df, 0x3fe7a4f707bf97d2,
+	0x3fa78dbaa5874686, 0x3feff753bb1b9164, 0xbfeff753bb1b9164, 0x3fa78dbaa5874686,
+	0x3feffce09ce2a679, 0x3f9c454f4ce53b1d, 0xbf9c454f4ce53b1d, 0x3feffce09ce2a679,
+	0x3fe5fe7cbde56a10, 0x3fe73e558e079942, 0xbfe73e558e079942, 0x3fe5fe7cbde56a10,
+	0x3fed36fc7bcbfbdc, 0x3fda1d6543b50ac0, 0xbfda1d6543b50ac0, 0x3fed36fc7bcbfbdc,
+	0x3fd6d998638a0cb6, 0x3fede4160f6d8d81, 0xbfede4160f6d8d81, 0x3fd6d998638a0cb6,
+	0x3fef33685a3aaef0, 0x3fcc6d90535d74dd, 0xbfcc6d90535d74dd, 0x3fef33685a3aaef0,
+	0x3fe1097248d0a957, 0x3feb16742a4ca2f5, 0xbfeb16742a4ca2f5, 0x3fe1097248d0a957,
+	0x3fea1b26d2c0a75e, 0x3fe2818bef4d3cba, 0xbfe2818bef4d3cba, 0x3fea1b26d2c0a75e,
+	0x3fc57f008654cbde, 0x3fef8ba737cb4b78, 0xbfef8ba737cb4b78, 0x3fc57f008654cbde,
+	0x3fefbf470f0a8d88, 0x3fc00ee8ad6fb85b, 0xbfc00ee8ad6fb85b, 0x3fefbf470f0a8d88,
+	0x3fe39c23e3d63029, 0x3fe94990e3ac4a6c, 0xbfe94990e3ac4a6c, 0x3fe39c23e3d63029,
+	0x3febcb54cb0d2327, 0x3fdfb7575c24d2de, 0xbfdfb7575c24d2de, 0x3febcb54cb0d2327,
+	0x3fd0e15b4e1749ce, 0x3feeddeb6a078651, 0xbfeeddeb6a078651, 0x3fd0e15b4e1749ce,
+	0x3fee5a9d550467d3, 0x3fd44310dc8936f0, 0xbfd44310dc8936f0, 0x3fee5a9d550467d3,
+	0x3fdc997fc3865389, 0x3feca08f19b9c449, 0xbfeca08f19b9c449, 0x3fdc997fc3865389,
+	0x3fe82a9c13f545ff, 0x3fe4f9cc25cca486, 0xbfe4f9cc25cca486, 0x3fe82a9c13f545ff,
+	0x3fb20c9674ed444d, 0x3fefeb9d2530410f, 0xbfefeb9d2530410f, 0x3fb20c9674ed444d,
+	0x3fefe7ea85482d60, 0x3fb39d9f12c5a299, 0xbfb39d9f12c5a299, 0x3fefe7ea85482d60,
+	0x3fe4d3bc6d589f7f, 0x3fe84b7111af83fa, 0xbfe84b7111af83fa, 0x3fe4d3bc6d589f7f,
+	0x3fec89f587029c13, 0x3fdcf34baee1cd21, 0xbfdcf34baee1cd21, 0x3fec89f587029c13,
+	0x3fd3e39be96ec271, 0x3fee6a61c55d53a7, 0xbfee6a61c55d53a7, 0x3fd3e39be96ec271,
+	0x3feed0835e999009, 0x3fd1423eefc69378, 0xbfd1423eefc69378, 0x3feed0835e999009,
+	0x3fdf5fdee656cda3, 0x3febe41b611154c1, 0xbfebe41b611154c1, 0x3fdf5fdee656cda3,
+	0x3fe92aa41fc5a815, 0x3fe3c3c44981c518, 0xbfe3c3c44981c518, 0x3fe92aa41fc5a815,
+	0x3fbe8eb7fde4aa3f, 0x3fefc56e3b7d9af6, 0xbfefc56e3b7d9af6, 0x3fbe8eb7fde4aa3f,
+	0x3fef830f4a40c60c, 0x3fc6451a831d830d, 0xbfc6451a831d830d, 0x3fef830f4a40c60c,
+	0x3fe258734cbb7110, 0x3fea38184a593bc6, 0xbfea38184a593bc6, 0x3fe258734cbb7110,
+	0x3feafb8fd89f57b6, 0x3fe133e9cfee254f, 0xbfe133e9cfee254f, 0x3feafb8fd89f57b6,
+	0x3fcba96334f15dad, 0x3fef3e6bbc1bbc65, 0xbfef3e6bbc1bbc65, 0x3fcba96334f15dad,
+	0x3fedd1fef38a915a, 0x3fd73763c9261092, 0xbfd73763c9261092, 0x3fedd1fef38a915a,
+	0x3fd9c17d440df9f2, 0x3fed4b5b1b187524, 0xbfed4b5b1b187524, 0x3fd9c17d440df9f2,
+	0x3fe71bac960e41bf, 0x3fe622e44fec22ff, 0xbfe622e44fec22ff, 0x3fe71bac960e41bf,
+	0x3f95fd4d21fab226, 0x3feffe1c6870cb77, 0xbfeffe1c6870cb77, 0x3f95fd4d21fab226,
+	0x3fefff0943c53bd1, 0x3f8f6a296ab997cb, 0xbf8f6a296ab997cb, 0x3fefff0943c53bd1,
+	0x3fe64715437f535b, 0x3fe6f8ca99c95b75, 0xbfe6f8ca99c95b75, 0x3fe64715437f535b,
+	0x3fed5f7172888a7f, 0x3fd96555b7ab948f, 0xbfd96555b7ab948f, 0x3fed5f7172888a7f,
+	0x3fd794f5e613dfae, 0x3fedbf9e4395759a, 0xbfedbf9e4395759a, 0x3fd794f5e613dfae,
+	0x3fef492206bcabb4, 0x3fcae4f1d5f3b9ab, 0xbfcae4f1d5f3b9ab, 0x3fef492206bcabb4,
+	0x3fe15e36e4dbe2bc, 0x3feae068f345ecef, 0xbfeae068f345ecef, 0x3fe15e36e4dbe2bc,
+	0x3fea54c91090f523, 0x3fe22f2d662c13e2, 0xbfe22f2d662c13e2, 0x3fea54c91090f523,
+	0x3fc70afd8d08c4ff, 0x3fef7a299c1a322a, 0xbfef7a299c1a322a, 0x3fc70afd8d08c4ff,
+	0x3fefcb4703914354, 0x3fbcff533b307dc1, 0xbfbcff533b307dc1, 0x3fefcb4703914354,
+	0x3fe3eb33eabe0680, 0x3fe90b7943575efe, 0xbfe90b7943575efe, 0x3fe3eb33eabe0680,
+	0x3febfc9d25a1b147, 0x3fdf081906bff7fe, 0xbfdf081906bff7fe, 0x3febfc9d25a1b147,
+	0x3fd1a2f7fbe8f243, 0x3feec2cf4b1af6b2, 0xbfeec2cf4b1af6b2, 0x3fd1a2f7fbe8f243,
+	0x3fee79db29a5165a, 0x3fd383f5e353b6ab, 0xbfd383f5e353b6ab, 0x3fee79db29a5165a,
+	0x3fdd4cd02ba8609d, 0x3fec7315899eaad7, 0xbfec7315899eaad7, 0x3fdd4cd02ba8609d,
+	0x3fe86c0a1d9aa195, 0x3fe4ad79516722f1, 0xbfe4ad79516722f1, 0x3fe86c0a1d9aa195,
+	0x3fb52e774a4d4d0a, 0x3fefe3e92be9d886, 0xbfefe3e92be9d886, 0x3fb52e774a4d4d0a,
+	0x3fefef0102826191, 0x3fb07b614e463064, 0xbfb07b614e463064, 0x3fefef0102826191,
+	0x3fe51fa81cd99aa6, 0x3fe8098b756e52fa, 0xbfe8098b756e52fa, 0x3fe51fa81cd99aa6,
+	0x3fecb6e20a00da99, 0x3fdc3f6d47263129, 0xbfdc3f6d47263129, 0x3fecb6e20a00da99,
+	0x3fd4a253d11b82f3, 0x3fee4a8dff81ce5e, 0xbfee4a8dff81ce5e, 0x3fd4a253d11b82f3,
+	0x3feeeb074c50a544, 0x3fd0804e05eb661e, 0xbfd0804e05eb661e, 0x3feeeb074c50a544,
+	0x3fe00740c82b82e1, 0x3febb249a0b6c40d, 0xbfebb249a0b6c40d, 0x3fe00740c82b82e1,
+	0x3fe9683f42bd7fe1, 0x3fe374531b817f8d, 0xbfe374531b817f8d, 0x3fe9683f42bd7fe1,
+	0x3fc0d64dbcb26786, 0x3fefb8d18d66adb7, 0xbfefb8d18d66adb7, 0x3fc0d64dbcb26786,
+	0x3fef93f14f85ac08, 0x3fc4b8b17f79fa88, 0xbfc4b8b17f79fa88, 0x3fef93f14f85ac08,
+	0x3fe2aa76e87aeb58, 0x3fe9fdf4f13149de, 0xbfe9fdf4f13149de, 0x3fe2aa76e87aeb58,
+	0x3feb3115a5f37bf3, 0x3fe0ded0b84bc4b6, 0xbfe0ded0b84bc4b6, 0x3feb3115a5f37bf3,
+	0x3fcd31774d2cbdee, 0x3fef2817fc4609ce, 0xbfef2817fc4609ce, 0x3fcd31774d2cbdee,
+	0x3fedf5e36a9ba59c, 0x3fd67b949cad63cb, 0xbfd67b949cad63cb, 0x3fedf5e36a9ba59c,
+	0x3fda790cd3dbf31b, 0x3fed2255c6e5a4e1, 0xbfed2255c6e5a4e1, 0x3fda790cd3dbf31b,
+	0x3fe760c52c304764, 0x3fe5d9dee73e345c, 0xbfe5d9dee73e345c, 0x3fe760c52c304764,
+	0x3fa14685db42c17f, 0x3feffb55e425fdae, 0xbfeffb55e425fdae, 0x3fa14685db42c17f,
+	0x3feff97c4208c014, 0x3fa46a396ff86179, 0xbfa46a396ff86179, 0x3feff97c4208c014,
+	0x3fe5b50b264f7448, 0x3fe782fb1b90b35b, 0xbfe782fb1b90b35b, 0x3fe5b50b264f7448,
+	0x3fed0d672f59d2b9, 0x3fdad473125cdc09, 0xbfdad473125cdc09, 0x3fed0d672f59d2b9,
+	0x3fd61d595c88c202, 0x3fee0766d9280f54, 0xbfee0766d9280f54, 0x3fd61d595c88c202,
+	0x3fef1c7abe284708, 0x3fcdf5163f01099a, 0xbfcdf5163f01099a, 0x3fef1c7abe284708,
+	0x3fe0b405878f85ec, 0x3feb4b7409de7925, 0xbfeb4b7409de7925, 0x3fe0b405878f85ec,
+	0x3fe9e082edb42472, 0x3fe2d333d34e9bb8, 0xbfe2d333d34e9bb8, 0x3fe9e082edb42472,
+	0x3fc3f22f57db4893, 0x3fef9bed7cfbde29, 0xbfef9bed7cfbde29, 0x3fc3f22f57db4893,
+	0x3fefb20dc681d54d, 0x3fc19d8940be24e7, 0xbfc19d8940be24e7, 0x3fefb20dc681d54d,
+	0x3fe34c5252c14de1, 0x3fe986aef1457594, 0xbfe986aef1457594, 0x3fe34c5252c14de1,
+	0x3feb98fa1fd9155e, 0x3fe032ae55edbd96, 0xbfe032ae55edbd96, 0x3feb98fa1fd9155e,
+	0x3fd01f1806b9fdd2, 0x3feef7d6e51ca3c0, 0xbfeef7d6e51ca3c0, 0x3fd01f1806b9fdd2,
+	0x3fee3a33ec75ce85, 0x3fd50163dc197048, 0xbfd50163dc197048, 0x3fee3a33ec75ce85,
+	0x3fdbe51517ffc0d9, 0x3fecccee20c2dea0, 0xbfecccee20c2dea0, 0x3fdbe51517ffc0d9,
+	0x3fe7e83f87b03686, 0x3fe5454ff5159dfc, 0xbfe5454ff5159dfc, 0x3fe7e83f87b03686,
+	0x3fadd406f9808ec9, 0x3feff21614e131ed, 0xbfeff21614e131ed, 0x3fadd406f9808ec9,
+	0x3fefdf9922f73307, 0x3fb6bf1b3e79b129, 0xbfb6bf1b3e79b129, 0x3fefdf9922f73307,
+	0x3fe48703306091ff, 0x3fe88c66e7481ba1, 0xbfe88c66e7481ba1, 0x3fe48703306091ff,
+	0x3fec5bef59fef85a, 0x3fdda60c5cfa10d9, 0xbfdda60c5cfa10d9, 0x3fec5bef59fef85a,
+	0x3fd3241fb638baaf, 0x3fee89095bad6025, 0xbfee89095bad6025, 0x3fd3241fb638baaf,
+	0x3feeb4cf515b8811, 0x3fd2038583d727be, 0xbfd2038583d727be, 0x3feeb4cf515b8811,
+	0x3fdeb00695f25620, 0x3fec14d9dc465e57, 0xbfec14d9dc465e57, 0x3fdeb00695f25620,
+	0x3fe8ec109b486c49, 0x3fe41272663d108c, 0xbfe41272663d108c, 0x3fe8ec109b486c49,
+	0x3fbb6fa6ec38f64c, 0x3fefd0d158d86087, 0xbfefd0d158d86087, 0x3fbb6fa6ec38f64c,
+	0x3fef70f6434b7eb7, 0x3fc7d0a7bbd2cb1c, 0xbfc7d0a7bbd2cb1c, 0x3fef70f6434b7eb7,
+	0x3fe205baa17560d6, 0x3fea7138de9d60f5, 0xbfea7138de9d60f5, 0x3fe205baa17560d6,
+	0x3feac4ffbd3efac8, 0x3fe188591f3a46e5, 0xbfe188591f3a46e5, 0x3feac4ffbd3efac8,
+	0x3fca203e1b1831da, 0x3fef538b1faf2d07, 0xbfef538b1faf2d07, 0x3fca203e1b1831da,
+	0x3fedacf42ce68ab9, 0x3fd7f24dd37341e4, 0xbfd7f24dd37341e4, 0x3fedacf42ce68ab9,
+	0x3fd908ef81ef7bd1, 0x3fed733f508c0dff, 0xbfed733f508c0dff, 0x3fd908ef81ef7bd1,
+	0x3fe6d5afef4aafcd, 0x3fe66b0f3f52b386, 0xbfe66b0f3f52b386, 0x3fe6d5afef4aafcd,
+	0x3f82d96b0e509703, 0x3fefffa72c978c4f, 0xbfefffa72c978c4f, 0x3f82d96b0e509703,
+	0x3fefffa72c978c4f, 0x3f82d96b0e509703, 0xbf82d96b0e509703, 0x3fefffa72c978c4f,
+	0x3fe66b0f3f52b386, 0x3fe6d5afef4aafcd, 0xbfe6d5afef4aafcd, 0x3fe66b0f3f52b386,
+	0x3fed733f508c0dff, 0x3fd908ef81ef7bd1, 0xbfd908ef81ef7bd1, 0x3fed733f508c0dff,
+	0x3fd7f24dd37341e4, 0x3fedacf42ce68ab9, 0xbfedacf42ce68ab9, 0x3fd7f24dd37341e4,
+	0x3fef538b1faf2d07, 0x3fca203e1b1831da, 0xbfca203e1b1831da, 0x3fef538b1faf2d07,
+	0x3fe188591f3a46e5, 0x3feac4ffbd3efac8, 0xbfeac4ffbd3efac8, 0x3fe188591f3a46e5,
+	0x3fea7138de9d60f5, 0x3fe205baa17560d6, 0xbfe205baa17560d6, 0x3fea7138de9d60f5,
+	0x3fc7d0a7bbd2cb1c, 0x3fef70f6434b7eb7, 0xbfef70f6434b7eb7, 0x3fc7d0a7bbd2cb1c,
+	0x3fefd0d158d86087, 0x3fbb6fa6ec38f64c, 0xbfbb6fa6ec38f64c, 0x3fefd0d158d86087,
+	0x3fe41272663d108c, 0x3fe8ec109b486c49, 0xbfe8ec109b486c49, 0x3fe41272663d108c,
+	0x3fec14d9dc465e57, 0x3fdeb00695f25620, 0xbfdeb00695f25620, 0x3fec14d9dc465e57,
+	0x3fd2038583d727be, 0x3feeb4cf515b8811, 0xbfeeb4cf515b8811, 0x3fd2038583d727be,
+	0x3fee89095bad6025, 0x3fd3241fb638baaf, 0xbfd3241fb638baaf, 0x3fee89095bad6025,
+	0x3fdda60c5cfa10d9, 0x3fec5bef59fef85a, 0xbfec5bef59fef85a, 0x3fdda60c5cfa10d9,
+	0x3fe88c66e7481ba1, 0x3fe48703306091ff, 0xbfe48703306091ff, 0x3fe88c66e7481ba1,
+	0x3fb6bf1b3e79b129, 0x3fefdf9922f73307, 0xbfefdf9922f73307, 0x3fb6bf1b3e79b129,
+	0x3feff21614e131ed, 0x3fadd406f9808ec9, 0xbfadd406f9808ec9, 0x3feff21614e131ed,
+	0x3fe5454ff5159dfc, 0x3fe7e83f87b03686, 0xbfe7e83f87b03686, 0x3fe5454ff5159dfc,
+	0x3fecccee20c2dea0, 0x3fdbe51517ffc0d9, 0xbfdbe51517ffc0d9, 0x3fecccee20c2dea0,
+	0x3fd50163dc197048, 0x3fee3a33ec75ce85, 0xbfee3a33ec75ce85, 0x3fd50163dc197048,
+	0x3feef7d6e51ca3c0, 0x3fd01f1806b9fdd2, 0xbfd01f1806b9fdd2, 0x3feef7d6e51ca3c0,
+	0x3fe032ae55edbd96, 0x3feb98fa1fd9155e, 0xbfeb98fa1fd9155e, 0x3fe032ae55edbd96,
+	0x3fe986aef1457594, 0x3fe34c5252c14de1, 0xbfe34c5252c14de1, 0x3fe986aef1457594,
+	0x3fc19d8940be24e7, 0x3fefb20dc681d54d, 0xbfefb20dc681d54d, 0x3fc19d8940be24e7,
+	0x3fef9bed7cfbde29, 0x3fc3f22f57db4893, 0xbfc3f22f57db4893, 0x3fef9bed7cfbde29,
+	0x3fe2d333d34e9bb8, 0x3fe9e082edb42472, 0xbfe9e082edb42472, 0x3fe2d333d34e9bb8,
+	0x3feb4b7409de7925, 0x3fe0b405878f85ec, 0xbfe0b405878f85ec, 0x3feb4b7409de7925,
+	0x3fcdf5163f01099a, 0x3fef1c7abe284708, 0xbfef1c7abe284708, 0x3fcdf5163f01099a,
+	0x3fee0766d9280f54, 0x3fd61d595c88c202, 0xbfd61d595c88c202, 0x3fee0766d9280f54,
+	0x3fdad473125cdc09, 0x3fed0d672f59d2b9, 0xbfed0d672f59d2b9, 0x3fdad473125cdc09,
+	0x3fe782fb1b90b35b, 0x3fe5b50b264f7448, 0xbfe5b50b264f7448, 0x3fe782fb1b90b35b,
+	0x3fa46a396ff86179, 0x3feff97c4208c014, 0xbfeff97c4208c014, 0x3fa46a396ff86179,
+	0x3feffb55e425fdae, 0x3fa14685db42c17f, 0xbfa14685db42c17f, 0x3feffb55e425fdae,
+	0x3fe5d9dee73e345c, 0x3fe760c52c304764, 0xbfe760c52c304764, 0x3fe5d9dee73e345c,
+	0x3fed2255c6e5a4e1, 0x3fda790cd3dbf31b, 0xbfda790cd3dbf31b, 0x3fed2255c6e5a4e1,
+	0x3fd67b949cad63cb, 0x3fedf5e36a9ba59c, 0xbfedf5e36a9ba59c, 0x3fd67b949cad63cb,
+	0x3fef2817fc4609ce, 0x3fcd31774d2cbdee, 0xbfcd31774d2cbdee, 0x3fef2817fc4609ce,
+	0x3fe0ded0b84bc4b6, 0x3feb3115a5f37bf3, 0xbfeb3115a5f37bf3, 0x3fe0ded0b84bc4b6,
+	0x3fe9fdf4f13149de, 0x3fe2aa76e87aeb58, 0xbfe2aa76e87aeb58, 0x3fe9fdf4f13149de,
+	0x3fc4b8b17f79fa88, 0x3fef93f14f85ac08, 0xbfef93f14f85ac08, 0x3fc4b8b17f79fa88,
+	0x3fefb8d18d66adb7, 0x3fc0d64dbcb26786, 0xbfc0d64dbcb26786, 0x3fefb8d18d66adb7,
+	0x3fe374531b817f8d, 0x3fe9683f42bd7fe1, 0xbfe9683f42bd7fe1, 0x3fe374531b817f8d,
+	0x3febb249a0b6c40d, 0x3fe00740c82b82e1, 0xbfe00740c82b82e1, 0x3febb249a0b6c40d,
+	0x3fd0804e05eb661e, 0x3feeeb074c50a544, 0xbfeeeb074c50a544, 0x3fd0804e05eb661e,
+	0x3fee4a8dff81ce5e, 0x3fd4a253d11b82f3, 0xbfd4a253d11b82f3, 0x3fee4a8dff81ce5e,
+	0x3fdc3f6d47263129, 0x3fecb6e20a00da99, 0xbfecb6e20a00da99, 0x3fdc3f6d47263129,
+	0x3fe8098b756e52fa, 0x3fe51fa81cd99aa6, 0xbfe51fa81cd99aa6, 0x3fe8098b756e52fa,
+	0x3fb07b614e463064, 0x3fefef0102826191, 0xbfefef0102826191, 0x3fb07b614e463064,
+	0x3fefe3e92be9d886, 0x3fb52e774a4d4d0a, 0xbfb52e774a4d4d0a, 0x3fefe3e92be9d886,
+	0x3fe4ad79516722f1, 0x3fe86c0a1d9aa195, 0xbfe86c0a1d9aa195, 0x3fe4ad79516722f1,
+	0x3fec7315899eaad7, 0x3fdd4cd02ba8609d, 0xbfdd4cd02ba8609d, 0x3fec7315899eaad7,
+	0x3fd383f5e353b6ab, 0x3fee79db29a5165a, 0xbfee79db29a5165a, 0x3fd383f5e353b6ab,
+	0x3feec2cf4b1af6b2, 0x3fd1a2f7fbe8f243, 0xbfd1a2f7fbe8f243, 0x3feec2cf4b1af6b2,
+	0x3fdf081906bff7fe, 0x3febfc9d25a1b147, 0xbfebfc9d25a1b147, 0x3fdf081906bff7fe,
+	0x3fe90b7943575efe, 0x3fe3eb33eabe0680, 0xbfe3eb33eabe0680, 0x3fe90b7943575efe,
+	0x3fbcff533b307dc1, 0x3fefcb4703914354, 0xbfefcb4703914354, 0x3fbcff533b307dc1,
+	0x3fef7a299c1a322a, 0x3fc70afd8d08c4ff, 0xbfc70afd8d08c4ff, 0x3fef7a299c1a322a,
+	0x3fe22f2d662c13e2, 0x3fea54c91090f523, 0xbfea54c91090f523, 0x3fe22f2d662c13e2,
+	0x3feae068f345ecef, 0x3fe15e36e4dbe2bc, 0xbfe15e36e4dbe2bc, 0x3feae068f345ecef,
+	0x3fcae4f1d5f3b9ab, 0x3fef492206bcabb4, 0xbfef492206bcabb4, 0x3fcae4f1d5f3b9ab,
+	0x3fedbf9e4395759a, 0x3fd794f5e613dfae, 0xbfd794f5e613dfae, 0x3fedbf9e4395759a,
+	0x3fd96555b7ab948f, 0x3fed5f7172888a7f, 0xbfed5f7172888a7f, 0x3fd96555b7ab948f,
+	0x3fe6f8ca99c95b75, 0x3fe64715437f535b, 0xbfe64715437f535b, 0x3fe6f8ca99c95b75,
+	0x3f8f6a296ab997cb, 0x3fefff0943c53bd1, 0xbfefff0943c53bd1, 0x3f8f6a296ab997cb,
+	0x3feffe1c6870cb77, 0x3f95fd4d21fab226, 0xbf95fd4d21fab226, 0x3feffe1c6870cb77,
+	0x3fe622e44fec22ff, 0x3fe71bac960e41bf, 0xbfe71bac960e41bf, 0x3fe622e44fec22ff,
+	0x3fed4b5b1b187524, 0x3fd9c17d440df9f2, 0xbfd9c17d440df9f2, 0x3fed4b5b1b187524,
+	0x3fd73763c9261092, 0x3fedd1fef38a915a, 0xbfedd1fef38a915a, 0x3fd73763c9261092,
+	0x3fef3e6bbc1bbc65, 0x3fcba96334f15dad, 0xbfcba96334f15dad, 0x3fef3e6bbc1bbc65,
+	0x3fe133e9cfee254f, 0x3feafb8fd89f57b6, 0xbfeafb8fd89f57b6, 0x3fe133e9cfee254f,
+	0x3fea38184a593bc6, 0x3fe258734cbb7110, 0xbfe258734cbb7110, 0x3fea38184a593bc6,
+	0x3fc6451a831d830d, 0x3fef830f4a40c60c, 0xbfef830f4a40c60c, 0x3fc6451a831d830d,
+	0x3fefc56e3b7d9af6, 0x3fbe8eb7fde4aa3f, 0xbfbe8eb7fde4aa3f, 0x3fefc56e3b7d9af6,
+	0x3fe3c3c44981c518, 0x3fe92aa41fc5a815, 0xbfe92aa41fc5a815, 0x3fe3c3c44981c518,
+	0x3febe41b611154c1, 0x3fdf5fdee656cda3, 0xbfdf5fdee656cda3, 0x3febe41b611154c1,
+	0x3fd1423eefc69378, 0x3feed0835e999009, 0xbfeed0835e999009, 0x3fd1423eefc69378,
+	0x3fee6a61c55d53a7, 0x3fd3e39be96ec271, 0xbfd3e39be96ec271, 0x3fee6a61c55d53a7,
+	0x3fdcf34baee1cd21, 0x3fec89f587029c13, 0xbfec89f587029c13, 0x3fdcf34baee1cd21,
+	0x3fe84b7111af83fa, 0x3fe4d3bc6d589f7f, 0xbfe4d3bc6d589f7f, 0x3fe84b7111af83fa,
+	0x3fb39d9f12c5a299, 0x3fefe7ea85482d60, 0xbfefe7ea85482d60, 0x3fb39d9f12c5a299,
+	0x3fefeb9d2530410f, 0x3fb20c9674ed444d, 0xbfb20c9674ed444d, 0x3fefeb9d2530410f,
+	0x3fe4f9cc25cca486, 0x3fe82a9c13f545ff, 0xbfe82a9c13f545ff, 0x3fe4f9cc25cca486,
+	0x3feca08f19b9c449, 0x3fdc997fc3865389, 0xbfdc997fc3865389, 0x3feca08f19b9c449,
+	0x3fd44310dc8936f0, 0x3fee5a9d550467d3, 0xbfee5a9d550467d3, 0x3fd44310dc8936f0,
+	0x3feeddeb6a078651, 0x3fd0e15b4e1749ce, 0xbfd0e15b4e1749ce, 0x3feeddeb6a078651,
+	0x3fdfb7575c24d2de, 0x3febcb54cb0d2327, 0xbfebcb54cb0d2327, 0x3fdfb7575c24d2de,
+	0x3fe94990e3ac4a6c, 0x3fe39c23e3d63029, 0xbfe39c23e3d63029, 0x3fe94990e3ac4a6c,
+	0x3fc00ee8ad6fb85b, 0x3fefbf470f0a8d88, 0xbfefbf470f0a8d88, 0x3fc00ee8ad6fb85b,
+	0x3fef8ba737cb4b78, 0x3fc57f008654cbde, 0xbfc57f008654cbde, 0x3fef8ba737cb4b78,
+	0x3fe2818bef4d3cba, 0x3fea1b26d2c0a75e, 0xbfea1b26d2c0a75e, 0x3fe2818bef4d3cba,
+	0x3feb16742a4ca2f5, 0x3fe1097248d0a957, 0xbfe1097248d0a957, 0x3feb16742a4ca2f5,
+	0x3fcc6d90535d74dd, 0x3fef33685a3aaef0, 0xbfef33685a3aaef0, 0x3fcc6d90535d74dd,
+	0x3fede4160f6d8d81, 0x3fd6d998638a0cb6, 0xbfd6d998638a0cb6, 0x3fede4160f6d8d81,
+	0x3fda1d6543b50ac0, 0x3fed36fc7bcbfbdc, 0xbfed36fc7bcbfbdc, 0x3fda1d6543b50ac0,
+	0x3fe73e558e079942, 0x3fe5fe7cbde56a10, 0xbfe5fe7cbde56a10, 0x3fe73e558e079942,
+	0x3f9c454f4ce53b1d, 0x3feffce09ce2a679, 0xbfeffce09ce2a679, 0x3f9c454f4ce53b1d,
+	0x3feff753bb1b9164, 0x3fa78dbaa5874686, 0xbfa78dbaa5874686, 0x3feff753bb1b9164,
+	0x3fe59001d5f723df, 0x3fe7a4f707bf97d2, 0xbfe7a4f707bf97d2, 0x3fe59001d5f723df,
+	0x3fecf830e8ce467b, 0x3fdb2f971db31972, 0xbfdb2f971db31972, 0x3fecf830e8ce467b,
+	0x3fd5bee78b9db3b6, 0x3fee18a02fdc66d9, 0xbfee18a02fdc66d9, 0x3fd5bee78b9db3b6,
+	0x3fef1090bc898f5f, 0x3fceb86b462de348, 0xbfceb86b462de348, 0x3fef1090bc898f5f,
+	0x3fe089112032b08c, 0x3feb658f14fdbc47, 0xbfeb658f14fdbc47, 0x3fe089112032b08c,
+	0x3fe9c2d110f075c2, 0x3fe2fbc24b441015, 0xbfe2fbc24b441015, 0x3fe9c2d110f075c2,
+	0x3fc32b7bf94516a7, 0x3fefa39bac7a1791, 0xbfefa39bac7a1791, 0x3fc32b7bf94516a7,
+	0x3fefaafbcb0cfddc, 0x3fc264994dfd3409, 0xbfc264994dfd3409, 0x3fefaafbcb0cfddc,
+	0x3fe32421ec49a61f, 0x3fe9a4dfa42b06b2, 0xbfe9a4dfa42b06b2, 0x3fe32421ec49a61f,
+	0x3feb7f6686e792e9, 0x3fe05df3ec31b8b7, 0xbfe05df3ec31b8b7, 0x3feb7f6686e792e9,
+	0x3fcf7b7480bd3802, 0x3fef045a14cf738c, 0xbfef045a14cf738c, 0x3fcf7b7480bd3802,
+	0x3fee298f4439197a, 0x3fd5604012f467b4, 0xbfd5604012f467b4, 0x3fee298f4439197a,
+	0x3fdb8a7814fd5693, 0x3fece2b32799a060, 0xbfece2b32799a060, 0x3fdb8a7814fd5693,
+	0x3fe7c6b89ce2d333, 0x3fe56ac35197649f, 0xbfe56ac35197649f, 0x3fe7c6b89ce2d333,
+	0x3faab101bd5f8317, 0x3feff4dc54b1bed3, 0xbfeff4dc54b1bed3, 0x3faab101bd5f8317,
+	0x3fefdafa7514538c, 0x3fb84f8712c130a1, 0xbfb84f8712c130a1, 0x3fefdafa7514538c,
+	0x3fe4605a692b32a2, 0x3fe8ac871ede1d88, 0xbfe8ac871ede1d88, 0x3fe4605a692b32a2,
+	0x3fec44833141c004, 0x3fddfeff66a941de, 0xbfddfeff66a941de, 0x3fec44833141c004,
+	0x3fd2c41a4e954520, 0x3fee97ec36016b30, 0xbfee97ec36016b30, 0x3fd2c41a4e954520,
+	0x3feea68393e65800, 0x3fd263e6995554ba, 0xbfd263e6995554ba, 0x3feea68393e65800,
+	0x3fde57a86d3cd825, 0x3fec2cd14931e3f1, 0xbfec2cd14931e3f1, 0x3fde57a86d3cd825,
+	0x3fe8cc6a75184655, 0x3fe4397f5b2a4380, 0xbfe4397f5b2a4380, 0x3fe8cc6a75184655,
+	0x3fb9dfb6eb24a85c, 0x3fefd60d2da75c9e, 0xbfefd60d2da75c9e, 0x3fb9dfb6eb24a85c,
+	0x3fef677556883cee, 0x3fc8961727c41804, 0xbfc8961727c41804, 0x3fef677556883cee,
+	0x3fe1dc1b64dc4872, 0x3fea8d676e545ad2, 0xbfea8d676e545ad2, 0x3fe1dc1b64dc4872,
+	0x3feaa9547a2cb98e, 0x3fe1b250171373bf, 0xbfe1b250171373bf, 0x3feaa9547a2cb98e,
+	0x3fc95b49e9b62afa, 0x3fef5da6ed43685d, 0xbfef5da6ed43685d, 0x3fc95b49e9b62afa,
+	0x3fed9a00dd8b3d46, 0x3fd84f6aaaf3903f, 0xbfd84f6aaaf3903f, 0x3fed9a00dd8b3d46,
+	0x3fd8ac4b86d5ed44, 0x3fed86c48445a44f, 0xbfed86c48445a44f, 0x3fd8ac4b86d5ed44,
+	0x3fe6b25ced2fe29c, 0x3fe68ed1eaa19c71, 0xbfe68ed1eaa19c71, 0x3fe6b25ced2fe29c,
+	0x3f6921f8becca4ba, 0x3feffff621621d02, 0xbfeffff621621d02, 0x3f6921f8becca4ba,
+}
